Grisette.Experimental

Description

Experimental features

The experimental features are likely to be changed in the future, and they do not comply with the semantics versioning policy.

Use the APIs with caution.

Symbolic Generation with Errors Class

class Mergeable a => GenSymConstrained spec a where

Class of types in which symbolic values can be generated with some specification.

See GenSym for more details. The difference of this class is that it allows constraints to be generated along with the generation of symbolic values.

Minimal complete definition

Nothing

Methods

freshConstrained :: (MonadFresh m, MonadError e m, MonadUnion m) => e -> spec -> m (Union a)

Generates a symbolic value with the given specification.

Constraint violations will throw an error in the monadic environment.

>>> runFreshT (freshConstrained () (SymOrdUpperBound (1 :: SymInteger) ())) "a" :: ExceptT () Union (Union SymInteger)
ExceptT <If (<= 1 a@0) (Left ()) (Right {a@0})>

default freshConstrained :: (GenSymSimpleConstrained spec a, MonadFresh m, MonadError e m, MonadUnion m) => e -> spec -> m (Union a)

class Mergeable a => GenSymSimpleConstrained spec a where

Class of types in which symbolic values can be generated with some specification.

See GenSymSimple for more details. The difference of this class is that it allows constraints to be generated along with the generation of symbolic values.

Methods

simpleFreshConstrained :: (MonadFresh m, MonadError e m, MonadUnion m) => e -> spec -> m a

Generates a symbolic value with the given specification.

Constraint violations will throw an error in the monadic environment.

>>> runFreshT (simpleFreshConstrained () (SymOrdUpperBound (1 :: SymInteger) ())) "a" :: ExceptT () Union SymInteger
ExceptT <If (<= 1 a@0) (Left ()) (Right a@0)>

genSymConstrained :: (GenSymConstrained spec a, Mergeable e) => e -> spec -> Identifier -> ExceptT e Union (Union a)

Generates a symbolic value with the given specification, also place the necessary constraints.

genSymSimpleConstrained :: (GenSymSimpleConstrained spec a, Mergeable e) => e -> spec -> Identifier -> ExceptT e Union a

Generates a symbolic value with the given specification, also place the necessary constraints.

derivedSimpleFreshConstrainedNoSpec :: (Generic a, GenSymSimpleConstrainedNoSpec (Rep a), MonadFresh m, MonadError e m, MonadUnion m, Mergeable a) => e -> () -> m a

We cannot provide DerivingVia style derivation for GenSymSimpleConstrained, while you can use this simpleFreshConstrained implementation to implement GenSymSimpleConstrained fo your own types.

This simpleFreshConstrained implementation is for the types that does not need any specification. It will generate product types by generating each fields with () as specification. It will not work on sum types.

Note: Never use on recursive types.

derivedSimpleFreshConstrainedSameShape :: (Generic a, GenSymConstrainedSameShape (Rep a), Mergeable a, MonadFresh m, MonadError e m, MonadUnion m) => e -> a -> m a

We cannot provide DerivingVia style derivation for GenSymSimpleConstrained, while you can use this simpleFreshConstrained implementation to implement GenSymSimpleConstrained fo your own types.

This simpleFreshConstrained implementation is for the types that can be generated with a reference value of the same type.

For sum types, it will generate the result with the same data constructor. For product types, it will generate the result by generating each field with the corresponding reference value.

Note: Can be used on recursive types.

derivedFreshConstrainedNoSpec :: (Generic a, GenSymConstrainedNoSpec (Rep a), Mergeable a, MonadFresh m, MonadError e m, MonadUnion m) => e -> () -> m (Union a)

We cannot provide DerivingVia style derivation for GenSymConstrained, while you can use this freshConstrained implementation to implement GenSymConstrained for your own types.

This freshConstrained implementation is for the types that does not need any specification. It will generate product types by generating each fields with () as specification, and generate all possible values for a sum type.

Note: Never use on recursive types.

Some common GenSymConstrained specifications

data SymOrdUpperBound a spec

Exclusive bound, generates the values with the specification, then filters out the ones that are greater than or equal to the bound

Constructors

SymOrdUpperBound a spec

data SymOrdLowerBound a spec

Inclusive bound, generates the values with the specification, then filters out the ones that are less than the bound

Constructors

SymOrdLowerBound a spec

data SymOrdBound a spec

Left-inclusive, right-exclusive bound, generates the values with the specification, then filters out the ones that are out-of-bound

Constructors

SymOrdBound a a spec