Grisette.Internal.Backend.SymBiMap

Description

Documentation

data SymBiMap

A bidirectional map between symbolic Grisette terms and sbv terms.

Constructors

SymBiMap

Fields

biMapToSBV :: HashMap SomeTerm (QuantifiedStack -> Dynamic)

biMapSize :: Int

biMapFromSBV :: HashMap String SomeTypedAnySymbol

quantifiedSymbolNum :: Int

emptySymBiMap :: SymBiMap

An empty bidirectional map.

sizeBiMap :: SymBiMap -> Int

The size of the bidirectional map.

addBiMap :: forall (knd :: SymbolKind). HasCallStack => SomeTerm -> Dynamic -> String -> SomeTypedSymbol knd -> SymBiMap -> SymBiMap

Add a new entry to the bidirectional map.

addBiMapIntermediate :: HasCallStack => SomeTerm -> (QuantifiedStack -> Dynamic) -> SymBiMap -> SymBiMap

Add a new entry to the bidirectional map for intermediate values.

findStringToSymbol :: forall (knd :: SymbolKind). IsSymbolKind knd => String -> SymBiMap -> Maybe (SomeTypedSymbol knd)

Find a symbolic Grisette term from a string.

lookupTerm :: HasCallStack => SomeTerm -> SymBiMap -> Maybe (QuantifiedStack -> Dynamic)

Look up an sbv value with a symbolic Grisette term in the bidirectional map.

attachNextQuantifiedSymbolInfo :: SymBiMap -> TypedConstantSymbol a -> (SymBiMap, TypedConstantSymbol a)

Attach the next quantified symbol info to a symbol.