Grisette.Internal.Core.Control.Monad.CBMCExcept

Description

CBMC-like error handling

newtype CBMCEither a b

A wrapper type for Either. Uses different merging strategies.

Constructors

CBMCEither

Fields

runCBMCEither :: Either a b

newtype CBMCExceptT e (m :: Type -> Type) a

Similar to ExceptT, but with different error handling mechanism.

Constructors

CBMCExceptT

Fields

runCBMCExceptT :: m (CBMCEither e a)

cbmcExcept :: forall (m :: Type -> Type) e a. Monad m => Either e a -> CBMCExceptT e m a

Wrap an Either value in CBMCExceptT

mapCBMCExceptT :: (m (Either e a) -> n (Either e' b)) -> CBMCExceptT e m a -> CBMCExceptT e' n b

Map the error and values in a CBMCExceptT

withCBMCExceptT :: forall (m :: Type -> Type) e e' a. Functor m => (e -> e') -> CBMCExceptT e m a -> CBMCExceptT e' m a

Map the error in a CBMCExceptT

class Monad m => MonadError e (m :: Type -> Type) | m -> e where

The strategy of combining computations that can throw exceptions by bypassing bound functions from the point an exception is thrown to the point that it is handled.

Is parameterized over the type of error information and the monad type constructor. It is common to use Either String as the monad type constructor for an error monad in which error descriptions take the form of strings. In that case and many other common cases the resulting monad is already defined as an instance of the MonadError class. You can also define your own error type and/or use a monad type constructor other than Either String or Either IOError. In these cases you will have to explicitly define instances of the MonadError class. (If you are using the deprecated Control.Monad.Error or Control.Monad.Trans.Error, you may also have to define an Error instance.)

Methods

throwError :: e -> m a

Is used within a monadic computation to begin exception processing.

catchError :: m a -> (e -> m a) -> m a

A handler function to handle previous errors and return to normal execution. A common idiom is:

do { action1; action2; action3 } `catchError` handler

where the action functions can call throwError. Note that handler and the do-block must have the same return type.