Grisette.Internal.Core.Data.Class.AsKey

Documentation

class KeyEq a where

Type class for identity equality for terms.

Methods

keyEq :: a -> a -> Bool infix 4

class KeyEq a => KeyOrd a where

Type class for comparing terms based on their identity.

Methods

keyCompare :: a -> a -> Ordering infix 4

class KeyEq a => KeyHashable a where

Type class for hashing terms based on their identity.

Methods

keyHashWithSalt :: Int -> a -> Int infixl 0

class KeyEq1 (f :: Type -> Type) where

Methods

liftKeyEq :: (a -> b -> Bool) -> f a -> f b -> Bool

class KeyEq1 f => KeyOrd1 (f :: Type -> Type) where

Methods

liftKeyCompare :: (a -> b -> Ordering) -> f a -> f b -> Ordering

class KeyEq1 f => KeyHashable1 (f :: Type -> Type) where

Methods

liftKeyHashWithSalt :: (Int -> a -> Int) -> Int -> f a -> Int

newtype AsKey a

Use a term as a key with identity equality.

For example, AsKey SymBool uses the term identity for SymBool.

Constructors

AsKey

Fields

getAsKey :: a

newtype AsKey1 (f :: Type -> Type) a

Use a union as a key with identity equality.

For example, AsKey1 Union uses the term identity for Union.

Constructors

AsKey1

Fields

getAsKey1 :: f a

shouldUseAsKeyError :: HasCallStack => String -> String -> a

shouldUseAsKeyHasSymbolicVersionError :: HasCallStack => String -> String -> String -> a

shouldUseSymbolicVersionError :: HasCallStack => String -> String -> String -> a