Grisette.Internal.Core.Data.Class.BitCast

Description

Documentation

class BitCast from to where

Type class for bit-casting between types.

Special Considerations for Floating-Point Types:

Typically, bit-casting a value from type a to type b and then back to type a should result in the original value. However, this is not always true for floating-point values. In SMT-LIB2, there is only one NaN value with multiple bit representations.

Given this, we do not provide BitCast for the FP type, instead, we use the bitCastOrCanonical function to use a canonical representation for the NaN values.

If your application requires distinguishing between different NaN values, it is recommended to define your own floating-point type using bit-vectors. This allows you to check for NaN values and perform operations by bitcasting back to the provided floating-point types when they are not NaN values.

Methods

bitCast :: from -> to

class BitCastCanonical from to where

The canonical value when the bitcast cannot be precisely performed.

For example, with SMT-LIB2, there is only one NaN for floating point numbers, with multiple bit representations. Our underlying FP type also follows this convention. This means that we cannot precisely bitcast a FP to other types. So instead, we bitcast the NaN value to a canonical representation, defined with this type class.

Methods

bitCastCanonicalValue :: proxy from -> to

class BitCastOr from to where

Bitcasting a value. If the value cannot be precisely bitcast, use the default value.

Methods

bitCastOr :: to -> from -> to

bitCastOrCanonical :: BitCastOrCanonical from to => from -> to

Bitcasting a value and when the value cannot be precisely bitcast, use the canonical value.

type BitCastOrCanonical a b = (BitCastCanonical a b, BitCastOr a b)

Constraint for bitcasting a value and when the value cannot be precisely bitcast, use the canonical value.