Grisette.Internal.Core.Data.Class.CEGISSolver

Description

Note for the examples

The examples assumes that the z3 solver is available in PATH.

Generic CEGIS interface

data VerifierResult cex exception

The response from a verifier.

Constructors

CEGISVerifierFoundCex cex

CEGISVerifierNoCex Bool
True indicates that the verifier is sure that there is no counter-example, while False indicates that the verifier is not sure, but it cannot find a counter-example.

CEGISVerifierException exception

type SynthesisConstraintFun cex = cex -> IO SymBool

Build the synthesizer constraint from the verfication result. The first argument will be guaranteed to be distinct during each invocation of the function in the CEGIS algorithm, so it can be used to instantiate the identifiers for fresh variables.

type VerifierFun cex exception = Model -> IO (VerifierResult cex exception)

The verifier.

data CEGISResult exception

The result of the CEGIS procedure.

Constructors

CEGISSuccess Model

CEGISVerifierFailure exception

CEGISSolverFailure SolvingFailure

solverGenericCEGIS

Arguments

:: Solver handle

=> handle
Configuration of the solver.

-> Bool
Whether we should rerun the passed verifiers if any other verifier found a counter-example.

-> SymBool
The initial synthesis constraint.

-> SynthesisConstraintFun input
Synthesis constraint from counter-examples

-> [VerifierFun input exception]
The verifier functions.

-> IO ([input], CEGISResult exception)

Generic CEGIS procedure. See genericCEGIS for more details.

The difference from genericCEGIS is that this function accepts a solver handle for the synthesizer, instead of a solver configuration.

solverGenericCEGISWithRefinement

Arguments

:: Solver handle

=> handle
Configuration of the solver.

-> Bool
Whether we should rerun the passed verifiers if any other verifier found a counter-example.

-> SymBool
The initial synthesis constraint.

-> SynthesisConstraintFun input
Synthesis constraint from counter-examples

-> Maybe RefinementConditionFun
Refinement condition generator.

-> [VerifierFun input exception]
The verifier functions.

-> IO ([input], CEGISResult exception)

Generic CEGIS procedure with refinement. See genericCEGISWithRefinement for more details.

The difference from genericCEGISWithRefinement is that this function accepts a solver handle for the synthesizer, instead of a solver configuration.

genericCEGIS

Arguments

:: ConfigurableSolver config handle

=> config
Configuration of the solver.

-> Bool
Whether we should rerun the passed verifiers if any other verifier found a counter-example.

-> SymBool
The initial synthesis constraint.

-> SynthesisConstraintFun input
Synthesis constraint from counter-examples

-> [VerifierFun input exception]
The verifier functions.

-> IO ([input], CEGISResult exception)

Generic CEGIS procedure.

The CEGIS procedure will try to find a model that satisfies the initial synthesis constraint, and satisfies all the inputs generated by the verifier.

genericCEGISWithRefinement

Arguments

:: ConfigurableSolver config handle

=> config
Configuration of the solver.

-> Bool
Whether we should rerun the passed verifiers if any other verifier found a counter-example.

-> SymBool
The initial synthesis constraint.

-> SynthesisConstraintFun input
Synthesis constraint from counter-examples

-> Maybe RefinementConditionFun
Refinement condition generator.

-> [VerifierFun input exception]
The verifier functions.

-> IO ([input], CEGISResult exception)

Generic CEGIS procedure.

The CEGIS procedure will try to find a model that satisfies the initial synthesis constraint, and satisfies all the inputs generated by the verifier.

CEGIS interfaces with pre/post conditions

data CEGISCondition

The condition for CEGIS to solve.

The first argument is the pre-condition, and the second argument is the post-condition.

The CEGIS procedures would try to find a model for the formula

\[ \forall P. (\exists I. \mathrm{pre}(P, I)) \wedge (\forall I. \mathrm{pre}(P, I)\implies \mathrm{post}(P, I)) \]

In program synthesis tasks, \(P\) is the symbolic constants in the symbolic program, and \(I\) is the input. The pre-condition is used to restrict the search space of the program. The procedure would only return programs that meets the pre-conditions on every possible inputs, and there are at least one possible input. The post-condition is used to specify the desired program behaviors.

Constructors

CEGISCondition SymBool SymBool

cegisPostCond :: SymBool -> CEGISCondition

Construct a CEGIS condition with only a post-condition. The pre-condition would be set to true, meaning that all programs in the program space are allowed.

cegisPrePost :: SymBool -> SymBool -> CEGISCondition

Construct a CEGIS condition with both pre- and post-conditions.

solverCegisMultiInputs

Arguments

:: (EvalSym input, ExtractSym input, Solver handle)

=> handle

-> handle

-> [input]
Initial symbolic inputs. The solver will try to find a program that works on all the inputs representable by these inputs (see CEGISCondition).

-> (input -> CEGISCondition)
The condition for the solver to solve. All the symbolic constants that are not in the inputs will be considered as part of the symbolic program.

-> IO ([input], CEGISResult SolvingFailure)
The counter-examples generated during the CEGIS loop, and the model found by the solver.

CEGIS with multiple (possibly symbolic) inputs. See cegisMultiInputs for more details.

The difference from cegisMultiInputs is that this function accepts two solver handles, one for the synthesizer and one for the verifier.

The synthesizer solver will **not** be reset, while the verifier solver will be reset after each iteration.

solverCegis

Arguments

:: (Solver handle, EvalSym inputs, ExtractSym inputs, SymEq inputs)

=> handle
The synthesizer solver handle

-> handle
The verifier solver handle

-> inputs
Initial symbolic inputs. The solver will try to find a program that works on all the inputs representable by it (see CEGISCondition).

-> (inputs -> CEGISCondition)
The condition for the solver to solve. All the symbolic constants that are not in the inputs will be considered as part of the symbolic program.

-> IO ([inputs], CEGISResult SolvingFailure)
The counter-examples generated during the CEGIS loop, and the model found by the solver.

CEGIS with a single symbolic input to represent a set of inputs. See cegis for more details.

The difference from cegis is that this function accepts two solver handles, one for the synthesizer and one for the verifier.

The synthesizer solver will **not** be reset, while the verifier solver will be reset after each iteration.

solverCegisExcept :: forall t (u :: Type -> Type) e v inputs handle. (UnionWithExcept t u e v, PlainUnion u, Functor u, EvalSym inputs, ExtractSym inputs, Solver handle, SymEq inputs) => handle -> handle -> inputs -> (Either e v -> CEGISCondition) -> (inputs -> t) -> IO ([inputs], CEGISResult SolvingFailure)

CEGIS for symbolic programs with error handling, using a single symbolic input to represent a set of inputs. See cegisExcept for more details.

The difference from cegisExcept is that this function accepts two solver handles, one for the synthesizer and one for the verifier.

The synthesizer solver will **not** be reset, while the verifier solver will be reset after each iteration.

solverCegisExceptStdVC :: forall t (u :: Type -> Type) inputs handle. (UnionWithExcept t u VerificationConditions (), PlainUnion u, Monad u, EvalSym inputs, ExtractSym inputs, Solver handle, SymEq inputs) => handle -> handle -> inputs -> (inputs -> t) -> IO ([inputs], CEGISResult SolvingFailure)

CEGIS for symbolic programs with error handling, using a single symbolic input to represent a set of inputs. See cegisExceptStdVC for more details.

The difference from cegisExceptStdVC is that this function accepts two solver handles, one for the synthesizer and one for the verifier.

The synthesizer solver will **not** be reset, while the verifier solver will be reset after each iteration.

solverCegisExceptVC :: (UnionWithExcept t u e v, PlainUnion u, Monad u, EvalSym inputs, ExtractSym inputs, Solver handle, SymEq inputs) => handle -> handle -> inputs -> (Either e v -> u (Either VerificationConditions ())) -> (inputs -> t) -> IO ([inputs], CEGISResult SolvingFailure)

CEGIS for symbolic programs with error handling, using a single symbolic input to represent a set of inputs.

The errors should be translated to assertion or assumption violations.

The difference from cegisExceptVC is that this function accepts two solver handles, one for the synthesizer and one for the verifier.

The synthesizer solver will **not** be reset, while the verifier solver will be reset after each iteration.

solverCegisExceptMultiInputs :: forall handle inputs t (u :: Type -> Type) e v. (Solver handle, EvalSym inputs, ExtractSym inputs, UnionWithExcept t u e v, PlainUnion u, Monad u) => handle -> handle -> [inputs] -> (Either e v -> CEGISCondition) -> (inputs -> t) -> IO ([inputs], CEGISResult SolvingFailure)

CEGIS for symbolic programs with error handling, using multiple (possibly symbolic) inputs to represent a set of inputs.

The difference from cegisExceptMultiInputs is that this function accepts two solver handles, one for the synthesizer and one for the verifier.

The synthesizer solver will **not** be reset, while the verifier solver will be reset after each iteration.

solverCegisExceptStdVCMultiInputs :: forall handle inputs t (u :: Type -> Type). (Solver handle, EvalSym inputs, ExtractSym inputs, UnionWithExcept t u VerificationConditions (), PlainUnion u, Monad u) => handle -> handle -> [inputs] -> (inputs -> t) -> IO ([inputs], CEGISResult SolvingFailure)

CEGIS for symbolic programs with error handling, using multiple (possibly symbolic) inputs to represent a set of inputs. See cegisExceptStdVCMultiInputs for more details.

The difference from cegisExceptStdVCMultiInputs is that this function accepts two solver handles, one for the synthesizer and one for the verifier.

The synthesizer solver will **not** be reset, while the verifier solver will be reset after each iteration.

solverCegisExceptVCMultiInputs :: (Solver handle, EvalSym inputs, ExtractSym inputs, UnionWithExcept t u e v, PlainUnion u, Monad u) => handle -> handle -> [inputs] -> (Either e v -> u (Either VerificationConditions ())) -> (inputs -> t) -> IO ([inputs], CEGISResult SolvingFailure)

CEGIS for symbolic programs with error handling, using multiple (possibly symbolic) inputs to represent a set of inputs.

The errors should be translated to assertion or assumption violations.

The difference from cegisExceptVCMultiInputs is that this function accepts two solver handles, one for the synthesizer and one for the verifier.

The synthesizer solver will **not** be reset, while the verifier solver will be reset after each iteration.

solverCegisForAll

Arguments

:: (ExtractSym forallInput, Solver handle)

=> handle

-> handle

-> forallInput
A symbolic value. All the symbolic constants in the value are treated as for-all variables.

-> CEGISCondition

-> IO ([Model], CEGISResult SolvingFailure)
First output are the counter-examples for all the for-all variables, and the second output is the model for all other variables if CEGIS succeeds.

CEGIS with a single symbolic input to represent a set of inputs. See cegisForAll for more details.

The difference from cegisForAll is that this function accepts two solver handles, one for the synthesizer and one for the verifier.

The synthesizer solver will **not** be reset, while the verifier solver will be reset after each iteration.

solverCegisForAllExcept :: forall t (u :: Type -> Type) e v inputs handle. (UnionWithExcept t u e v, PlainUnion u, Functor u, EvalSym inputs, ExtractSym inputs, Solver handle, SymEq inputs) => handle -> handle -> inputs -> (Either e v -> CEGISCondition) -> t -> IO ([Model], CEGISResult SolvingFailure)

CEGIS for symbolic programs with error handling, with a forall variable.

See cegisForAllExcept, cegisForAll and cegisExcept.

The difference from cegisForAllExcept is that this function accepts two solver handles, one for the synthesizer and one for the verifier.

The synthesizer solver will **not** be reset, while the verifier solver will be reset after each iteration.

solverCegisForAllExceptStdVC :: forall t (u :: Type -> Type) inputs handle. (UnionWithExcept t u VerificationConditions (), PlainUnion u, Monad u, EvalSym inputs, ExtractSym inputs, Solver handle, SymEq inputs) => handle -> handle -> inputs -> t -> IO ([Model], CEGISResult SolvingFailure)

CEGIS for symbolic programs with error handling, with a forall variable.

See cegisForAllExceptStdVC cegisForAll and cegisExceptStdVC.

The difference from cegisForAllExceptStdVC is that this function accepts two solver handles, one for the synthesizer and one for the verifier.

The synthesizer solver will **not** be reset, while the verifier solver will be reset after each iteration.

solverCegisForAllExceptVC :: (UnionWithExcept t u e v, PlainUnion u, Monad u, EvalSym inputs, ExtractSym inputs, Solver handle, SymEq inputs) => handle -> handle -> inputs -> (Either e v -> u (Either VerificationConditions ())) -> t -> IO ([Model], CEGISResult SolvingFailure)

CEGIS for symbolic programs with error handling, with a forall variable.

See cegisForAllExceptVC cegisForAll and cegisExceptVC.

The difference from cegisForAllExceptVC is that this function accepts two solver handles, one for the synthesizer and one for the verifier.

The synthesizer solver will **not** be reset, while the verifier solver will be reset after each iteration.

cegisMultiInputs

Arguments

:: (EvalSym input, ExtractSym input, ConfigurableSolver config handle)

=> config
The configuration of the solver

-> [input]
Initial symbolic inputs. The solver will try to find a program that works on all the inputs representable by these inputs (see CEGISCondition).

-> (input -> CEGISCondition)
The condition for the solver to solve. All the symbolic constants that are not in the inputs will be considered as part of the symbolic program.

-> IO ([input], CEGISResult SolvingFailure)
The counter-examples generated during the CEGIS loop, and the model found by the solver.

CEGIS with multiple (possibly symbolic) inputs. Solves the following formula (see CEGISCondition for details).

\[ \forall P. (\exists I\in\mathrm{inputs}. \mathrm{pre}(P, I)) \wedge (\forall I\in\mathrm{inputs}. \mathrm{pre}(P, I)\implies \mathrm{post}(P, I)) \]

For simpler queries, where the inputs are representable by a single symbolic value, you may want to use cegis or cegisExcept instead. We have an example for the cegis call.

cegis

Arguments

:: (ConfigurableSolver config handle, EvalSym inputs, ExtractSym inputs, SymEq inputs)

=> config
The configuration of the solver

-> inputs
Initial symbolic inputs. The solver will try to find a program that works on all the inputs representable by it (see CEGISCondition).

-> (inputs -> CEGISCondition)
The condition for the solver to solve. All the symbolic constants that are not in the inputs will be considered as part of the symbolic program.

-> IO ([inputs], CEGISResult SolvingFailure)
The counter-examples generated during the CEGIS loop, and the model found by the solver.

CEGIS with a single symbolic input to represent a set of inputs.

The following example tries to find the value of c such that for all positive x, x * c && c -2. The c .> -2 clause is used to make the solution unique.

>>> let [x,c] = ["x","c"] :: [SymInteger]
>>> cegis z3 x (\x -> cegisPrePost (x .> 0) (x * c .< 0 .&& c .> -2))
(...,CEGISSuccess (Model {c -> -1 :: Integer}))

cegisExcept :: forall t (u :: Type -> Type) e v inputs config handle. (UnionWithExcept t u e v, PlainUnion u, Functor u, EvalSym inputs, ExtractSym inputs, ConfigurableSolver config handle, SymEq inputs) => config -> inputs -> (Either e v -> CEGISCondition) -> (inputs -> t) -> IO ([inputs], CEGISResult SolvingFailure)

CEGIS for symbolic programs with error handling, using a single symbolic input to represent a set of inputs.

cegisExcept is particularly useful when custom error types are used. With cegisExcept, you define how the errors are interpreted to the CEGIS conditions after the symbolic evaluation. This could increase the readability and modularity of the code.

The following example tries to find the value of c such that for all positive x, x * c && c -2. The c .> -2 assertion is used to make the solution unique.

>>> let [x,c] = ["x","c"] :: [SymInteger]
>>> import Control.Monad.Except
>>> :{
  res :: SymInteger -> ExceptT VerificationConditions Union ()
  res x = do
    symAssume $ x .> 0
    symAssert $ x * c .< 0
    symAssert $ c .> -2
:}

>>> :{
  translation (Left AssumptionViolation) = cegisPrePost (con False) (con True)
  translation (Left AssertionViolation) = cegisPostCond (con False)
  translation _ = cegisPostCond (con True)
:}

>>> cegisExcept z3 x translation res
([...],CEGISSuccess (Model {c -> -1 :: Integer}))

cegisExceptStdVC :: forall t (u :: Type -> Type) inputs config handle. (UnionWithExcept t u VerificationConditions (), PlainUnion u, Monad u, EvalSym inputs, ExtractSym inputs, ConfigurableSolver config handle, SymEq inputs) => config -> inputs -> (inputs -> t) -> IO ([inputs], CEGISResult SolvingFailure)

CEGIS for symbolic programs with error handling, using a single symbolic input to represent a set of inputs. This function saves the efforts to implement the translation function for the standard error type VerificationConditions, and the standard result type ().

This function translates assumption violations to failed pre-conditions, and translates assertion violations to failed post-conditions. The () result will not fail any conditions.

The following example tries to find the value of c such that for all positive x, x * c && c -2. The c .> -2 assertion is used to make the solution unique.

>>> let [x,c] = ["x","c"] :: [SymInteger]
>>> import Control.Monad.Except
>>> :{
  res :: SymInteger -> ExceptT VerificationConditions Union ()
  res x = do
    symAssume $ x .> 0
    symAssert $ x * c .< 0
    symAssert $ c .> -2
:}

>>> cegisExceptStdVC z3 x res
([...],CEGISSuccess (Model {c -> -1 :: Integer}))

cegisExceptVC :: (UnionWithExcept t u e v, PlainUnion u, Monad u, EvalSym inputs, ExtractSym inputs, ConfigurableSolver config handle, SymEq inputs) => config -> inputs -> (Either e v -> u (Either VerificationConditions ())) -> (inputs -> t) -> IO ([inputs], CEGISResult SolvingFailure)

CEGIS for symbolic programs with error handling, using a single symbolic input to represent a set of inputs.

The errors should be translated to assertion or assumption violations.

cegisExceptMultiInputs :: forall config handle inputs t (u :: Type -> Type) e v. (ConfigurableSolver config handle, EvalSym inputs, ExtractSym inputs, UnionWithExcept t u e v, PlainUnion u, Monad u) => config -> [inputs] -> (Either e v -> CEGISCondition) -> (inputs -> t) -> IO ([inputs], CEGISResult SolvingFailure)

CEGIS for symbolic programs with error handling, using multiple (possibly symbolic) inputs to represent a set of inputs.

cegisExceptStdVCMultiInputs :: forall config handle inputs t (u :: Type -> Type). (ConfigurableSolver config handle, EvalSym inputs, ExtractSym inputs, UnionWithExcept t u VerificationConditions (), PlainUnion u, Monad u) => config -> [inputs] -> (inputs -> t) -> IO ([inputs], CEGISResult SolvingFailure)

CEGIS for symbolic programs with error handling, using multiple (possibly symbolic) inputs to represent a set of inputs. This function saves the efforts to implement the translation function for the standard error type VerificationConditions, and the standard result type ().

This function translates assumption violations to failed pre-conditions, and translates assertion violations to failed post-conditions. The () result will not fail any conditions.

cegisExceptVCMultiInputs :: (ConfigurableSolver config handle, EvalSym inputs, ExtractSym inputs, UnionWithExcept t u e v, PlainUnion u, Monad u) => config -> [inputs] -> (Either e v -> u (Either VerificationConditions ())) -> (inputs -> t) -> IO ([inputs], CEGISResult SolvingFailure)

CEGIS for symbolic programs with error handling, using multiple (possibly symbolic) inputs to represent a set of inputs.

The errors should be translated to assertion or assumption violations.

cegisForAll

Arguments

:: (ExtractSym forallInput, ConfigurableSolver config handle)

=> config

-> forallInput
A symbolic value. All the symbolic constants in the value are treated as for-all variables.

-> CEGISCondition

-> IO ([Model], CEGISResult SolvingFailure)
First output are the counter-examples for all the for-all variables, and the second output is the model for all other variables if CEGIS succeeds.

CEGIS with a single symbolic input to represent a set of inputs.

The following example tries to find the value of c such that for all positive x, x * c && c -2. The c .> -2 clause is used to make the solution unique.

>>> let [x,c] = ["x","c"] :: [SymInteger]
>>> cegisForAll z3 x $ cegisPrePost (x .> 0) (x * c .< 0 .&& c .> -2)
(...,CEGISSuccess (Model {c -> -1 :: Integer}))

cegisForAllExcept :: forall t (u :: Type -> Type) e v inputs config handle. (UnionWithExcept t u e v, PlainUnion u, Functor u, EvalSym inputs, ExtractSym inputs, ConfigurableSolver config handle, SymEq inputs) => config -> inputs -> (Either e v -> CEGISCondition) -> t -> IO ([Model], CEGISResult SolvingFailure)

CEGIS for symbolic programs with error handling, with a forall variable.

See cegisForAll and cegisExcept.

cegisForAllExceptStdVC :: forall t (u :: Type -> Type) inputs config handle. (UnionWithExcept t u VerificationConditions (), PlainUnion u, Monad u, EvalSym inputs, ExtractSym inputs, ConfigurableSolver config handle, SymEq inputs) => config -> inputs -> t -> IO ([Model], CEGISResult SolvingFailure)

CEGIS for symbolic programs with error handling, with a forall variable.

See cegisForAll and cegisExceptStdVC.

cegisForAllExceptVC :: (UnionWithExcept t u e v, PlainUnion u, Monad u, EvalSym inputs, ExtractSym inputs, ConfigurableSolver config handle, SymEq inputs) => config -> inputs -> (Either e v -> u (Either VerificationConditions ())) -> t -> IO ([Model], CEGISResult SolvingFailure)

CEGIS for symbolic programs with error handling, with a forall variable.

See cegisForAll and cegisExceptVC.