Grisette.Internal.Core.Data.Class.Concrete

Documentation

class Concrete a

A tag for concrete types.

Can be used in code to constrain types to concrete ones. For example, when trying to support hash maps in Grisette, we may want to restrict the key type to be concrete. The Concrete type class is handy.