Grisette.Internal.Core.Data.Class.EvalSym

Description

Evaluating symbolic values with model

class EvalSym a where

Evaluating symbolic values with some model. This would substitute the symbols (symbolic constants) with the values in the model.

>>> let model = insertValue "a" (1 :: Integer) emptyModel :: Model
>>> evalSym False model ([ssym "a", ssym "b"] :: [SymInteger])
[1,b]

If we set the first argument true, the missing symbols will be filled in with some default values:

>>> evalSym True model ([ssym "a", ssym "b"] :: [SymInteger])
[1,0]

Note 1: This type class can be derived for algebraic data types. You may need the DerivingVia and DerivingStrategies extensions.

data X = ... deriving Generic deriving EvalSym via (Default X)

Methods

evalSym :: Bool -> Model -> a -> a

Evaluate a symbolic value with some model, possibly fill in values for the missing symbols.

evalSymToCon :: (ToCon a b, EvalSym a) => Model -> a -> b

Evaluate a symbolic value with some model, fill in values for the missing symbols, and convert the result to a concrete value.

>>> let model = insertValue "a" (1 :: Integer) emptyModel :: Model
>>> evalSymToCon model ([ssym "a", ssym "b"] :: [SymInteger]) :: [Integer]
[1,0]

class (forall a. EvalSym a => EvalSym (f a)) => EvalSym1 (f :: Type -> Type) where

Lifting of EvalSym to unary type constructors.

Methods

liftEvalSym :: (Bool -> Model -> a -> a) -> Bool -> Model -> f a -> f a

Lift the evalSym function to unary type constructors.

evalSym1 :: (EvalSym1 f, EvalSym a) => Bool -> Model -> f a -> f a

Lifting the standard evalSym to unary type constructors.

evalSymToCon1 :: (EvalSym1 f, EvalSym a, ToCon1 f g, ToCon a b) => Model -> f a -> g b

Evaluate and convert to concrete values with lifted standard evalSym for unary type constructors. See evalSymToCon.

class (forall a. EvalSym a => EvalSym1 (f a)) => EvalSym2 (f :: Type -> Type -> Type) where

Lifting of EvalSym1 to binary type constructors.

Methods

liftEvalSym2 :: (Bool -> Model -> a -> a) -> (Bool -> Model -> b -> b) -> Bool -> Model -> f a b -> f a b

Lift the evalSym function to binary type constructors.

evalSym2 :: (EvalSym2 f, EvalSym a, EvalSym b) => Bool -> Model -> f a b -> f a b

Lifting the standard evalSym to binary type constructors.

evalSymToCon2 :: (EvalSym2 f, EvalSym a, EvalSym c, ToCon2 f g, ToCon a b, ToCon c d) => Model -> f a c -> g b d

Evaluate and convert to concrete values with lifted standard evalSym for binary type constructors. See evalSymToCon.

Generic EvalSym

data family EvalSymArgs arity a

The arguments to the generic evalSym function.

class GEvalSym arity (f :: Type -> Type) where

The class of types that can be generically evaluated.

Methods

gevalSym :: EvalSymArgs arity a -> Bool -> Model -> f a -> f a

genericEvalSym :: (Generic a, GEvalSym Arity0 (Rep a)) => Bool -> Model -> a -> a

Generic evalSym function.

genericLiftEvalSym :: (Generic1 f, GEvalSym Arity1 (Rep1 f)) => (Bool -> Model -> a -> a) -> Bool -> Model -> f a -> f a

Generic liftEvalSym function.