Grisette.Internal.Core.Data.Class.Function

Description

Function operations

class Function f arg ret | f -> arg ret where

Abstraction for function-like types.

Methods

(#) :: f -> arg -> ret infixl 9

Function application operator.

The operator is not right associated (like ($)). It is left associated, and you can provide many arguments with this operator once at a time.

>>> (+1) # 2
3

>>> (+) # 2 # 3
5

class Apply uf where

Applying an uninterpreted function.

>>> let f = "f" :: SymInteger =~> SymInteger =~> SymInteger
>>> apply f "a" "b"
(apply (apply f a) b)

Note that for implementation reasons, you can also use apply function on a non-function symbolic value. In this case, the function is treated as an id function.

Associated Types

type FunType uf

Methods

apply :: uf -> FunType uf