Grisette.Internal.Core.Data.Class.IEEEFP

Description

Documentation

fpIsNaN :: RealFloat a => a -> Bool

Check if a floating-point number is not-a-number.

fpIsPositiveZero :: RealFloat a => a -> Bool

Check if a floating-point number is positive zero.

fpIsNegativeZero :: RealFloat a => a -> Bool

Check if a floating-point number is negative zero.

fpIsPositiveInfinite :: RealFloat a => a -> Bool

Check if a floating-point number is positive infinite.

fpIsNegativeInfinite :: RealFloat a => a -> Bool

Check if a floating-point number is negative infinite.

fpIsPositive :: RealFloat a => a -> Bool

Check if a floating-point number is positive. +0, +inf are considered positive. nan, -0, -inf are not positive.

fpIsNegative :: RealFloat a => a -> Bool

Check if a floating-point number is negative. -0, -inf are considered negative. nan, +0, +inf are not negative.

fpIsInfinite :: RealFloat a => a -> Bool

Check if a floating-point number is infinite.

fpIsZero :: RealFloat a => a -> Bool

Check if a floating-point number is zero.

fpIsNormal :: RealFloat a => a -> Bool

Check if a floating-point number is normal, i.e., not 0, not inf, not nan, and not denormalized.

fpIsSubnormal :: RealFloat a => a -> Bool

Check if a floating-point number is subnormal, i.e., denormalized. 0, inf, or nan are not subnormal.

fpIsPoint :: RealFloat a => a -> Bool

Check if a floating-point number is a point, i.e., not inf, not nan.

class IEEEFPConstants a where

Constants for IEEE floating-point numbers.

Methods

fpPositiveInfinite :: a

Positive infinity.

fpNegativeInfinite :: a

Negative infinity.

fpNaN :: a

Not-a-number.

fpNegativeZero :: a

Negative zero.

fpPositiveZero :: a

Positive zero.

fpMinNormalized :: a

Smallest positive normalized number.

fpMinSubnormal :: a

Smallest positive subnormal number.

fpMaxNormalized :: a

Largest positive normalized number.

fpMaxSubnormal :: a

Largest positive subnormal number.

class IEEEFPRoundingMode mode where

Rounding modes for floating-point operations.

Methods

rne :: mode

Round to nearest, ties to even.

rna :: mode

Round to nearest, ties to away from zero.

rtp :: mode

Round towards positive infinity.

rtn :: mode

Round towards negative infinity.

rtz :: mode

Round towards zero.

class IEEEFPOp a where

Operations on IEEE floating-point numbers, without rounding mode.

Methods

fpAbs :: a -> a

IEEE754-2019 abs operation.

fpNeg :: a -> a

IEEE754-2019 negate operation.

fpRem :: a -> a -> a

IEEE754-2019 remainder operation.

fpMinimum :: a -> a -> a

IEEE754-2019 minimum operation.

The comparison for zeros follows -0 < 0

Returns NaN if one operand is NaN.

fpMinimumNumber :: a -> a -> a

IEEE754-2019 minimumNumber operation.

The comparison for zeros follows -0 < 0

Returns the other operand if one operand is NaN.

fpMaximum :: a -> a -> a

IEEE754-2019 maximum operation.

The comparison for zeros follows -0 < 0

Returns NaN if one operand is NaN.

fpMaximumNumber :: a -> a -> a

IEEE754-2019 maximumNumber operation.

The comparison for zeros follows -0 < 0

Returns the other operand if one operand is NaN.

class IEEEFPRoundingMode mode => IEEEFPRoundingOp a mode | a -> mode where

Operations on IEEE floating-point numbers, with rounding mode.

Methods

fpAdd :: mode -> a -> a -> a

fpSub :: mode -> a -> a -> a

fpMul :: mode -> a -> a -> a

fpDiv :: mode -> a -> a -> a

fpFMA :: mode -> a -> a -> a -> a

fpSqrt :: mode -> a -> a

fpRoundToIntegral :: mode -> a -> a

class IEEEFPConvertible a fp mode | fp -> mode where

Conversion from and to FPs.

Methods

fromFPOr

Arguments

:: a
Default value when converting non-representable FPs. For example, when converting to non-FP types, the NaN and infinities are not representable. Additionally, when converting to bit-vectors, out-of-bound FPs are not representable.

Note that out-of-bound means that the value after conversion is out of bound, not the value before conversion, meaning that converting from 3.5 to 2-bit unsigned bit-vector is out-of-bound when rounding to positive, but not when rounding to negative.

-> mode
Rounding mode. Ignored when converting to AlgReal because every representable FP value is converted to an exact AlgReal.

-> fp
FP value.

-> a

toFP :: mode -> a -> fp

class (IEEEFPConvertible a fp mode, IEEEFPRoundingMode mode) => IEEEFPToAlgReal a fp mode | fp -> mode where

Converting FP to real numbers.

Minimal complete definition

Nothing

Methods

fpToAlgReal :: a -> fp -> a

Similar to fromFPOr for AlgReal, but dropped the ignored rounding mode.