Grisette.Internal.Core.Data.Class.ModelOps

Description

Model and symbolic set operations

class Monoid symbolSet => SymbolSetOps symbolSet (typedSymbol :: Type -> Type) | symbolSet -> typedSymbol where

The operations on symbolic constant sets

Note that symbolic constants with different types are considered different.

>>> let aBool = "a" :: TypedAnySymbol Bool
>>> let bBool = "b" :: TypedAnySymbol Bool
>>> let cBool = "c" :: TypedAnySymbol Bool
>>> let aInteger = "a" :: TypedAnySymbol Integer
>>> emptySet :: AnySymbolSet
SymbolSet {}
>>> containsSymbol aBool (buildSymbolSet aBool :: AnySymbolSet)
True
>>> containsSymbol bBool (buildSymbolSet aBool :: AnySymbolSet)
False
>>> insertSymbol aBool (buildSymbolSet aBool :: AnySymbolSet)
SymbolSet {a :: Bool}
>>> insertSymbol aInteger (buildSymbolSet aBool :: AnySymbolSet)
SymbolSet {a :: Bool, a :: Integer}
>>> let abSet = buildSymbolSet (aBool, bBool) :: AnySymbolSet
>>> let acSet = buildSymbolSet (aBool, cBool) :: AnySymbolSet
>>> intersectionSet abSet acSet
SymbolSet {a :: Bool}
>>> unionSet abSet acSet
SymbolSet {a :: Bool, b :: Bool, c :: Bool}
>>> differenceSet abSet acSet
SymbolSet {b :: Bool}

Methods

emptySet :: symbolSet

Construct an empty set

isEmptySet :: symbolSet -> Bool

Check if the set is empty

containsSymbol :: typedSymbol a -> symbolSet -> Bool

Check if the set contains the given symbol

insertSymbol :: typedSymbol a -> symbolSet -> symbolSet

Insert a symbol into the set

intersectionSet :: symbolSet -> symbolSet -> symbolSet

Set intersection

unionSet :: symbolSet -> symbolSet -> symbolSet

Set union

differenceSet :: symbolSet -> symbolSet -> symbolSet

Set difference

class SymbolSetOps symbolSet typedSymbol => SymbolSetRep rep symbolSet (typedSymbol :: Type -> Type) where

A type class for building a symbolic constant set manually from a symbolic constant set representation

>>> buildSymbolSet ("a" :: TypedAnySymbol Bool, "b" :: TypedAnySymbol Bool) :: AnySymbolSet
SymbolSet {a :: Bool, b :: Bool}

Methods

buildSymbolSet :: rep -> symbolSet

Build a symbolic constant set

class SymbolSetOps symbolSet typedSymbol => ModelOps model symbolSet (typedSymbol :: Type -> Type) | model -> symbolSet typedSymbol where

The operations on Models.

Note that symbolic constants with different types are considered different.

>>> let aBool = "a" :: TypedAnySymbol Bool
>>> let bBool = "b" :: TypedAnySymbol Bool
>>> let cBool = "c" :: TypedAnySymbol Bool
>>> let aInteger = "a" :: TypedAnySymbol Integer
>>> emptyModel :: Model
Model {}
>>> valueOf aBool (buildModel (aBool ::= True) :: Model)
Just True
>>> valueOf bBool (buildModel (aBool ::= True) :: Model)
Nothing
>>> insertValue bBool False (buildModel (aBool ::= True) :: Model)
Model {a -> true :: Bool, b -> false :: Bool}
>>> let abModel = buildModel (aBool ::= True, bBool ::= False) :: Model
>>> let acSet = buildSymbolSet (aBool, cBool) :: AnySymbolSet
>>> exceptFor acSet abModel
Model {b -> false :: Bool}
>>> restrictTo acSet abModel
Model {a -> true :: Bool}
>>> extendTo acSet abModel
Model {a -> true :: Bool, b -> false :: Bool, c -> false :: Bool}
>>> exact acSet abModel
Model {a -> true :: Bool, c -> false :: Bool}

Minimal complete definition

emptyModel, isEmptyModel, modelContains, valueOf, insertValue, exceptFor, exceptFor', restrictTo, extendTo

Methods

emptyModel :: model

Construct an empty model

isEmptyModel :: model -> Bool

Check if the model is empty

modelContains :: typedSymbol a -> model -> Bool

Check if the model contains the given symbol

valueOf :: typedSymbol t -> model -> Maybe t

Extract the assigned value for a given symbolic constant

insertValue :: typedSymbol t -> t -> model -> model

Insert an assignment into the model

exceptFor :: symbolSet -> model -> model

Returns a model that removed all the assignments for the symbolic constants in the set

exceptFor' :: typedSymbol t -> model -> model

Returns a model that removed the assignments for the symbolic constants

restrictTo :: symbolSet -> model -> model

Returns a model that only keeps the assignments for the symbolic constants in the set

extendTo :: symbolSet -> model -> model

Returns a model that extends the assignments for the symbolic constants in the set by assigning default values to them

exact :: symbolSet -> model -> model

Returns a model that contains the assignments for exactly the symbolic constants in the set by removing assignments for the symbolic constants that are not in the set and add assignments for the missing symbolic constants by assigning default values to them.

class ModelRep rep model | rep -> model where

A type class for building a model manually from a model representation

Methods

buildModel :: rep -> model

Build a model

>>> let aBool = "a" :: TypedAnySymbol Bool
>>> let bBool = "b" :: TypedAnySymbol Bool
>>> buildModel (aBool ::= True, bBool ::= False) :: Model
Model {a -> true :: Bool, b -> false :: Bool}