Grisette.Internal.Core.Data.Class.PPrint

Description

Pretty printing

class PPrint a where

Pretty printing of values.

This class is similar to the Pretty class from the Prettyprinter package, but it also provides pretty printing with a given precedence level.

We are able to derive instances of this class for algebraic data types. You may need the DerivingVia and DerivingStrategies extensions.

data X = ... deriving Generic deriving PPrint via (Default X)

The derived instance will pretty print the value with a format similar to the one used by ormolu.

Minimal complete definition

pformat | pformatPrec

Methods

pformat :: a -> Doc ann

pformatPrec :: Int -> a -> Doc ann

pformatList :: [a] -> Doc ann

docToTextWith :: LayoutOptions -> Doc ann -> Text

Convenience function to layout and render a Doc to Text.

You can control the layout with LayoutOptions.

docToTextWithWidth :: Int -> Doc ann -> Text

Convenience function to layout and render a Doc to Text.

You can control the layout with a single number of the width limit.

docToText :: Doc ann -> Text

Convenience function to layout and render a Doc to Text.

The default layout options defaultLayoutOptions are used.

pformatTextWith :: PPrint a => LayoutOptions -> a -> Text

Convenience function to format a value to Text.

You can control the layout with LayoutOptions.

pformatTextWithWidth :: PPrint a => Int -> a -> Text

Convenience function to format a value to Text.

You can control the layout with a single number of the width limit.

pformatText :: PPrint a => a -> Text

Convenience function to format a value to Text.

The default layout options defaultLayoutOptions are used.

pprint :: PPrint a => a -> IO ()

Pretty print a value to the standard output.

class (forall a. PPrint a => PPrint (f a)) => PPrint1 (f :: Type -> Type) where

Lifting of the PPrint class to unary type constructors.

Minimal complete definition

liftPFormatPrec

Methods

liftPFormatPrec :: (Int -> a -> Doc ann) -> ([a] -> Doc ann) -> Int -> f a -> Doc ann

Lift a pretty-printer to a unary type constructor.

liftPFormatList :: (Int -> a -> Doc ann) -> ([a] -> Doc ann) -> [f a] -> Doc ann

Lift a pretty-printer to list of values with unary type constructors.

pformatPrec1 :: (PPrint1 f, PPrint a) => Int -> f a -> Doc ann

Lift the standard pretty-printer (pformatPrec, pformatList) to unary type constructors.

pformatList1 :: (PPrint1 f, PPrint a) => [f a] -> Doc ann

Lift the standard pretty-printer (pformatPrec, pformatList) to list of values with unary type constructors.

class (forall a. PPrint a => PPrint1 (f a), forall a b. (PPrint a, PPrint b) => PPrint (f a b)) => PPrint2 (f :: Type -> Type -> Type) where

Lifting of the PPrint class to binary type constructors.

Minimal complete definition

liftPFormatPrec2

Methods

liftPFormatPrec2 :: (Int -> a -> Doc ann) -> ([a] -> Doc ann) -> (Int -> b -> Doc ann) -> ([b] -> Doc ann) -> Int -> f a b -> Doc ann

Lift two pretty-printers to a binary type constructor.

liftPFormatList2 :: (Int -> a -> Doc ann) -> ([a] -> Doc ann) -> (Int -> b -> Doc ann) -> ([b] -> Doc ann) -> [f a b] -> Doc ann

Lift two pretty-printers to list of values with binary type constructors.

pformatPrec2 :: (PPrint2 f, PPrint a, PPrint b) => Int -> f a b -> Doc ann

Lift the standard pretty-printer (pformatPrec, pformatList) to binary type constructors.

pformatList2 :: (PPrint2 f, PPrint a, PPrint b) => [f a b] -> Doc ann

Lift the standard pretty-printer (pformatPrec, pformatList) to list of values with binary type constructors.

Generic PPrint

genericPFormatPrec :: (Generic a, GPPrint Arity0 (Rep a)) => Int -> a -> Doc ann

Generic pformatPrec function.

genericLiftPFormatPrec :: (Generic1 f, GPPrint Arity1 (Rep1 f)) => (Int -> a -> Doc ann) -> ([a] -> Doc ann) -> Int -> f a -> Doc ann

Generic liftPFormatPrec function.

genericPFormatList :: (Generic a, GPPrint Arity0 (Rep a)) => [a] -> Doc ann

Generic pformatList function.

genericLiftPFormatList :: (Generic1 f, GPPrint Arity1 (Rep1 f)) => (Int -> a -> Doc ann) -> ([a] -> Doc ann) -> [f a] -> Doc ann

Generic liftPFormatList function.

data family PPrintArgs arity a ann

The arguments to the generic PPrint class.

class GPPrint arity (f :: Type -> Type) where

Generic PPrint class.

Minimal complete definition

gpformatPrec

Methods

gpformatPrec :: PPrintArgs arity a ann -> PPrintType -> Int -> f a -> Doc ann

gpformatList :: HasCallStack => PPrintArgs arity a ann -> [f a] -> Doc ann

gisNullary :: HasCallStack => PPrintArgs arity a ann -> f a -> Bool

data PPrintType

Controls how to pretty-print a generic representation.

Constructors

Rec

Tup

Pref

Inf String Int

Helpers

groupedEnclose :: Doc ann -> Doc ann -> Doc ann -> Doc ann

Enclose a document with left and right documents.

The pretty printer will try to layout the document in a single line, but the right document may be split to a newline.

condEnclose :: Bool -> Doc ann -> Doc ann -> Doc ann -> Doc ann

Conditionally enclose a document with left and right documents.

If the condition is True, then this function is equivalent to groupedEnclose.

pformatWithConstructor :: Int -> Doc ann -> [Doc ann] -> Doc ann

Pretty print a list of fields with a constructor.

Aligns the fields and nests them by 2 spaces.

pformatWithConstructorNoAlign :: Int -> Doc ann -> [Doc ann] -> Doc ann

Pretty print a list of fields with a constructor without alignment.

viaShowsPrec :: (Int -> a -> ShowS) -> Int -> a -> Doc ann

Pretty print a value using showsPrec.

Re-exports

module Prettyprinter