Grisette.Internal.Core.Data.Class.SafeDiv

Description

Documentation

data ArithException

Arithmetic exceptions.

Constructors

Overflow

Underflow

DivideByZero

class (MonadError e m, TryMerge m, Mergeable a, DivOr a) => SafeDiv e a (m :: Type -> Type) where

Safe division with monadic error handling in multi-path execution. These procedures throw an exception when the divisor is zero. The result should be able to handle errors with MonadError.

Minimal complete definition

(safeDiv, safeMod | safeDivMod), (safeQuot, safeRem | safeQuotRem)

Methods

safeDiv :: a -> a -> m a

Safe div with monadic error handling in multi-path execution.

>>> safeDiv "a" "b" :: ExceptT ArithException Union SymInteger
ExceptT {If (= b 0) (Left divide by zero) (Right (div a b))}

safeMod :: a -> a -> m a

Safe mod with monadic error handling in multi-path execution.

>>> safeMod "a" "b" :: ExceptT ArithException Union SymInteger
ExceptT {If (= b 0) (Left divide by zero) (Right (mod a b))}

safeDivMod :: a -> a -> m (a, a)

Safe divMod with monadic error handling in multi-path execution.

>>> safeDivMod "a" "b" :: ExceptT ArithException Union (SymInteger, SymInteger)
ExceptT {If (= b 0) (Left divide by zero) (Right ((div a b),(mod a b)))}

safeQuot :: a -> a -> m a

Safe quot with monadic error handling in multi-path execution.

safeRem :: a -> a -> m a

Safe rem with monadic error handling in multi-path execution.

safeQuotRem :: a -> a -> m (a, a)

Safe quotRem with monadic error handling in multi-path execution.

class DivOr a where

Safe division handling with default values returned on exception.

Methods

divOr :: a -> a -> a -> a

Safe div with default value returned on exception.

>>> divOr "d" "a" "b" :: SymInteger
(ite (= b 0) d (div a b))

modOr :: a -> a -> a -> a

Safe mod with default value returned on exception.

>>> modOr "d" "a" "b" :: SymInteger
(ite (= b 0) d (mod a b))

divModOr :: (a, a) -> a -> a -> (a, a)

Safe divMod with default value returned on exception.

>>> divModOr ("d", "m") "a" "b" :: (SymInteger, SymInteger)
((ite (= b 0) d (div a b)),(ite (= b 0) m (mod a b)))

quotOr :: a -> a -> a -> a

Safe quot with default value returned on exception.

remOr :: a -> a -> a -> a

Safe rem with default value returned on exception.

quotRemOr :: (a, a) -> a -> a -> (a, a)

Safe quotRem with default value returned on exception.

divOrZero :: (DivOr a, Num a) => a -> a -> a

Safe div with 0 returned on exception.

modOrDividend :: (DivOr a, Num a) => a -> a -> a

Safe mod with dividend returned on exception.

quotOrZero :: (DivOr a, Num a) => a -> a -> a

Safe quot with 0 returned on exception.

remOrDividend :: (DivOr a, Num a) => a -> a -> a

Safe rem with dividend returned on exception.

divModOrZeroDividend :: (DivOr a, Num a) => a -> a -> (a, a)

Safe divMod with 0 returned on exception.

quotRemOrZeroDividend :: (DivOr a, Num a) => a -> a -> (a, a)

Safe quotRem with 0 returned on exception.