Grisette.Internal.Core.Data.Class.SafeFdiv

Description

Documentation

class (MonadError e m, TryMerge m, Mergeable a) => SafeFdiv e a (m :: Type -> Type) where

Safe fractional division with monadic error handling in multi-path execution. These procedures throw an exception when the denominator is zero. The result should be able to handle errors with MonadError.

Minimal complete definition

safeFdiv

Methods

safeFdiv :: a -> a -> m a

Safe fractional division with monadic error handling in multi-path execution.

>>> safeFdiv "a" "b" :: ExceptT ArithException Union SymAlgReal
ExceptT {If (= b 0.0) (Left Ratio has zero denominator) (Right (fdiv a b))}

safeRecip :: a -> m a

Safe fractional reciprocal with monadic error handling in multi-path execution.

>>> safeRecip "a" :: ExceptT ArithException Union SymAlgReal
ExceptT {If (= a 0.0) (Left Ratio has zero denominator) (Right (recip a))}

default safeRecip :: Fractional a => a -> m a

class FdivOr a where

Safe fractional with default values returned on exception.

Methods

fdivOr :: a -> a -> a -> a

Safe / with default values returned on exception.

>>> fdivOr "d" "a" "b" :: SymAlgReal
(ite (= b 0.0) d (fdiv a b))

recipOr :: a -> a -> a

Safe recip with default values returned on exception.

>>> recipOr "d" "a" :: SymAlgReal
(ite (= a 0.0) d (recip a))

fdivOrZero :: (FdivOr a, Num a) => a -> a -> a

Safe / with 0 returned on exception.

recipOrZero :: (FdivOr a, Num a) => a -> a

Safe recip with 0 returned on exception.