Grisette.Internal.Core.Data.Class.SafeLogBase

Description

Documentation

class (MonadError e m, TryMerge m, Mergeable a) => SafeLogBase e a (m :: Type -> Type) where

Safe logBase with monadic error handling in multi-path execution. These procedures throw an exception when the base is 1. The result should be able to handle errors with MonadError.

Minimal complete definition

Nothing

Methods

safeLogBase :: a -> a -> m a

Safe logBase with monadic error handling in multi-path execution.

>>> safeLogBase (ssym "base") (ssym "val") :: ExceptT ArithException Union SymAlgReal
ExceptT {If (= base 1.0) (Left Ratio has zero denominator) (Right (fdiv (log val) (log base)))}

class LogBaseOr a where

Safe logBase with default values returned on exception.

Methods

logBaseOr :: a -> a -> a -> a

Safe logBase with default values returned on exception.

>>> logBaseOr "d" "base" "val" :: SymAlgReal
(ite (= base 1.0) d (fdiv (log val) (log base)))

logBaseOrZero :: (LogBaseOr a, Num a) => a -> a -> a

Safe logBase with 0 returned on exception.