Grisette.Internal.Core.Data.Class.SafeSymShift

Description

Documentation

class (MonadError e m, TryMerge m, Mergeable a) => SafeSymShift e a (m :: Type -> Type) where

Safe version for shiftL or shiftR.

The safeSymShiftL and safeSymShiftR are defined for all non-negative shift amounts.

Shifting by negative shift amounts is an error.

The result is defined to be 0 when shifting left by more than or equal to the bit size of the number.

The result is defined to be 0 when shifting right by more than or equal to the bit size of the number and the number is unsigned or signed non-negative.

The result is defined to be -1 when shifting right by more than or equal to the bit size of the number and the number is signed negative.

The safeSymStrictShiftL and safeSymStrictShiftR are defined for all non-negative shift amounts that is less than the bit size. Shifting by more than or equal to the bit size is an error, otherwise they are the same as the non-strict versions.

Methods

safeSymShiftL :: a -> a -> m a

safeSymShiftR :: a -> a -> m a

safeSymStrictShiftL :: a -> a -> m a

safeSymStrictShiftR :: a -> a -> m a