Grisette.Internal.Core.Data.Class.SymEq

Description

Symbolic equality

class SymEq a where

Symbolic equality. Note that we can't use Haskell's Eq class since symbolic comparison won't necessarily return a concrete Bool value.

>>> let a = 1 :: SymInteger
>>> let b = 2 :: SymInteger
>>> a .== b
false
>>> a ./= b
true

>>> let a = "a" :: SymInteger
>>> let b = "b" :: SymInteger
>>> a .== b
(= a b)
>>> a ./= b
(distinct a b)

Note: This type class can be derived for algebraic data types. You may need the DerivingVia and DerivingStrategies extensions.

data X = ... deriving Generic deriving SymEq via (Default X)

Minimal complete definition

(.==) | (./=)

Methods

(.==) :: a -> a -> SymBool infix 4

(./=) :: a -> a -> SymBool infix 4

symDistinct :: [a] -> SymBool

Check if all elements in a list are distinct, under the symbolic equality semantics.

class (forall a. SymEq a => SymEq (f a)) => SymEq1 (f :: Type -> Type) where

Lifting of the SymEq class to unary type constructors.

Any instance should be subject to the following law that canonicity is preserved:

liftSymEq (.==) should be equivalent to (.==), under the symbolic semantics.

This class therefore represents the generalization of SymEq by decomposing its main method into a canonical lifting on a canonical inner method, so that the lifting can be reused for other arguments than the canonical one.

Methods

liftSymEq :: (a -> b -> SymBool) -> f a -> f b -> SymBool

Lift a symbolic equality test through the type constructor.

The function will usually be applied to an symbolic equality function, but the more general type ensures that the implementation uses it to compare elements of the first container with elements of the second.

symEq1 :: (SymEq a, SymEq1 f) => f a -> f a -> SymBool

Lift the standard (.==) function through the type constructor.

class (forall a. SymEq a => SymEq1 (f a)) => SymEq2 (f :: Type -> Type -> Type) where

Lifting of the SymEq class to binary type constructors.

Methods

liftSymEq2 :: (a -> b -> SymBool) -> (c -> d -> SymBool) -> f a c -> f b d -> SymBool

Lift symbolic equality tests through the type constructor.

The function will usually be applied to an symbolic equality function, but the more general type ensures that the implementation uses it to compare elements of the first container with elements of the second.

symEq2 :: (SymEq a, SymEq b, SymEq2 f) => f a b -> f a b -> SymBool

Lift the standard (.==) function through the type constructor.

pairwiseSymDistinct :: SymEq a => [a] -> SymBool

Default pairwise symbolic distinct implementation.

More Eq helper

distinct :: Eq a => [a] -> Bool

Check if all elements in a list are distinct.

Note that empty or singleton lists are always distinct.

>>> distinct []
True
>>> distinct [1]
True
>>> distinct [1, 2, 3]
True
>>> distinct [1, 2, 2]
False

Generic SymEq

data family SymEqArgs arity a b

The arguments to the generic equality function.

class GSymEq arity (f :: Type -> Type) where

The class of types that can be generically compared for symbolic equality.

Methods

gsymEq :: SymEqArgs arity a b -> f a -> f b -> SymBool

genericSymEq :: (Generic a, GSymEq Arity0 (Rep a)) => a -> a -> SymBool

Generic (.==) function.

genericLiftSymEq :: (Generic1 f, GSymEq Arity1 (Rep1 f)) => (a -> b -> SymBool) -> f a -> f b -> SymBool

Generic liftSymEq function.