Grisette.Internal.Core.Data.Class.SymFiniteBits

Description

Documentation

lsb :: Bits a => a -> Bool

Extract the least significant bit of a concrete value.

msb :: FiniteBits a => a -> Bool

Extract the most significant bit of a concrete value.

setBitTo :: Bits a => a -> Int -> Bool -> a

Set a bit in a concrete value to a specific value.

bitBlast :: FiniteBits a => a -> [Bool]

Bit-blast a concrete value into a list of concrete bits. The first element in the resulting list corresponds to the least significant bit.

class FiniteBits a => FromBits a where

Type class for assembling concrete bits to a bit-vector.

Minimal complete definition

Nothing

Methods

fromBits :: [Bool] -> a

Assembling concrete bits to a bit-vector. The first boolean value in the list corresponding to the least signification value.

class (FiniteBits a, ITEOp a) => SymFiniteBits a where

A class for symbolic finite bit operations.

Minimal complete definition

symTestBit, symFromBits

Methods

symTestBit :: a -> Int -> SymBool

Test a symbolic bit in a symbolic bit-vector.

symSetBitTo :: a -> Int -> SymBool -> a

Set a bit in a symbolic value to a specific value.

symFromBits :: [SymBool] -> a

Assembling symbolic bits to a symbolic bit-vector. The first symbolic boolean value in the list corresponding to the least signification value.

symBitBlast :: SymFiniteBits a => a -> [SymBool]

Bit-blast a symbolic value into a list of symbolic bits. The first element in the resulting list corresponds to the least significant bit.

symLsb :: SymFiniteBits a => a -> SymBool

Extract the least significant bit of a symbolic value.

symMsb :: SymFiniteBits a => a -> SymBool

Extract the most significant bit of a symbolic value.

symPopCount :: (Num a, ITEOp a, SymFiniteBits a) => a -> a

Count the number of set bits in a symbolic value.

symCountLeadingZeros :: (Num a, ITEOp a, SymFiniteBits a) => a -> a

Count the number of leading zeros in a symbolic value.

symCountTrailingZeros :: (Num a, ITEOp a, SymFiniteBits a) => a -> a

Count the number of trailing zeros in a symbolic value.