Grisette.Internal.Core.Data.Class.SymOrd

Description

Symbolic total order relation

class SymEq a => SymOrd a where

Symbolic total order. Note that we can't use Haskell's Ord class since symbolic comparison won't necessarily return a concrete Bool or Ordering value.

>>> let a = 1 :: SymInteger
>>> let b = 2 :: SymInteger
>>> a .< b
true
>>> a .> b
false

>>> let a = "a" :: SymInteger
>>> let b = "b" :: SymInteger
>>> a .< b
(< a b)
>>> a .<= b
(<= a b)
>>> a .> b
(< b a)
>>> a .>= b
(<= b a)

For symCompare, Ordering is not a solvable type, and the result would be wrapped in a union-like monad. See Union and PlainUnion for more information.

>>> a `symCompare` b :: Union Ordering
{If (< a b) LT (If (= a b) EQ GT)}

Note: This type class can be derived for algebraic data types. You may need the DerivingVia and DerivingStrategies extensions.

data X = ... deriving Generic deriving SymOrd via (Default X)

Minimal complete definition

(.<) | symCompare

Methods

(.<) :: a -> a -> SymBool infix 4

(.<=) :: a -> a -> SymBool infix 4

(.>) :: a -> a -> SymBool infix 4

(.>=) :: a -> a -> SymBool infix 4

symCompare :: a -> a -> Union Ordering

class (SymEq1 f, forall a. SymOrd a => SymOrd (f a)) => SymOrd1 (f :: Type -> Type) where

Lifting of the SymOrd class to unary type constructors.

Any instance should be subject to the following law that canonicity is preserved:

liftSymCompare symCompare should be equivalent to symCompare, under the symbolic semantics.

This class therefore represents the generalization of SymOrd by decomposing its main method into a canonical lifting on a canonical inner method, so that the lifting can be reused for other arguments than the canonical one.

Methods

liftSymCompare :: (a -> b -> Union Ordering) -> f a -> f b -> Union Ordering

Lift a symCompare function through the type constructor.

The function will usually be applied to an symbolic comparison function, but the more general type ensures that the implementation uses it to compare elements of the first container with elements of the second.

symCompare1 :: (SymOrd1 f, SymOrd a) => f a -> f a -> Union Ordering

Lift the standard symCompare function to binary type constructors.

class (SymEq2 f, forall a. SymOrd a => SymOrd1 (f a)) => SymOrd2 (f :: Type -> Type -> Type) where

Lifting of the SymOrd class to binary type constructors.

Methods

liftSymCompare2 :: (a -> b -> Union Ordering) -> (c -> d -> Union Ordering) -> f a c -> f b d -> Union Ordering

Lift a symCompare function through the type constructor.

The function will usually be applied to an symbolic comparison function, but the more general type ensures that the implementation uses it to compare elements of the first container with elements of the second.

symCompare2 :: (SymOrd2 f, SymOrd a, SymOrd b) => f a b -> f a b -> Union Ordering

Lift the standard symCompare function through the type constructors.

Min and max

symMax :: (SymOrd a, ITEOp a) => a -> a -> a

Symbolic maximum.

symMin :: (SymOrd a, ITEOp a) => a -> a -> a

Symbolic minimum.

mrgMax :: (SymOrd a, Mergeable a, SymBranching m, Applicative m) => a -> a -> m a

Symbolic maximum, with a union-like monad.

mrgMin :: (SymOrd a, Mergeable a, SymBranching m, Applicative m) => a -> a -> m a

Symbolic minimum, with a union-like monad.

Generic SymOrd

data family SymOrdArgs arity a b

The arguments to the generic comparison function.

class GSymOrd arity (f :: Type -> Type) where

The class of types that can be generically symbolically compared.

Methods

gsymCompare :: SymOrdArgs arity a b -> f a -> f b -> Union Ordering

genericSymCompare :: (Generic a, GSymOrd Arity0 (Rep a)) => a -> a -> Union Ordering

Generic symCompare function.

genericLiftSymCompare :: (Generic1 f, GSymOrd Arity1 (Rep1 f)) => (a -> b -> Union Ordering) -> f a -> f b -> Union Ordering

Generic liftSymCompare function.