Grisette.Internal.Core.Data.Class.SymShift

Description

Documentation

class Bits a => SymShift a where

A class for shifting operations.

The symShift function shifts the value to the left if the shift amount is positive, and to the right if the shift amount is negative. If shifting beyond the bit width of the value, the result is the same as shifting with the bit width.

The symShiftNegated function shifts the value to the right if the shift amount is positive, and to the left if the shift amount is negative. This function is introduced to handle the asymmetry of the range of values.

Methods

symShift :: a -> a -> a

symShiftNegated :: a -> a -> a

newtype DefaultFiniteBitsSymShift a

A newtype wrapper. Use this to derive SymShift for types that have FiniteBits instances.

Constructors

DefaultFiniteBitsSymShift

Fields

unDefaultFiniteBitsSymShift :: a