Grisette.Internal.Core.Data.Class.ToCon

Description

Converting to concrete values

class ToCon a b where

Convert a symbolic value to concrete value if possible.

Methods

toCon :: a -> Maybe b

Convert a symbolic value to concrete value if possible. If the symbolic value cannot be converted to concrete, the result will be Nothing.

>>> toCon (ssym "a" :: SymInteger) :: Maybe Integer
Nothing

>>> toCon (con 1 :: SymInteger) :: Maybe Integer
Just 1

toCon works on complex types too.

>>> toCon ([con 1, con 2] :: [SymInteger]) :: Maybe [Integer]
Just [1,2]

>>> toCon ([con 1, ssym "a"] :: [SymInteger]) :: Maybe [Integer]
Nothing

class (forall a b. ToCon a b => ToCon (f1 a) (f2 b)) => ToCon1 (f1 :: Type -> Type) (f2 :: Type -> Type) where

Lifting of ToCon to unary type constructors.

Methods

liftToCon :: (a -> Maybe b) -> f1 a -> Maybe (f2 b)

Lift a conversion to concrete function to unary type constructors.

toCon1 :: (ToCon1 f1 f2, ToCon a b) => f1 a -> Maybe (f2 b)

Lift the standard toCon to unary type constructors.

class (forall a b. ToCon a b => ToCon1 (f1 a) (f2 b)) => ToCon2 (f1 :: Type -> Type -> Type) (f2 :: Type -> Type -> Type) where

Lifting of ToCon to binary type constructors.

Methods

liftToCon2 :: (a -> Maybe b) -> (c -> Maybe d) -> f1 a c -> Maybe (f2 b d)

Lift conversion to concrete functions to binary type constructors.

toCon2 :: (ToCon2 f1 f2, ToCon a b, ToCon c d) => f1 a c -> Maybe (f2 b d)

Lift the standard toCon to binary type constructors.

Generic ToCon

data family ToConArgs arity a b

The arguments to the generic toCon function.

class GToCon arity (f1 :: Type -> Type) (f2 :: Type -> Type) where

The class of types that can be generically converted to concrete values.

Methods

gtoCon :: ToConArgs arity a b -> f1 a -> Maybe (f2 b)

genericToCon :: (Generic a, Generic b, GToCon Arity0 (Rep a) (Rep b)) => a -> Maybe b

Generic toCon function.

genericLiftToCon :: (Generic1 f1, Generic1 f2, GToCon Arity1 (Rep1 f1) (Rep1 f2)) => (a -> Maybe b) -> f1 a -> Maybe (f2 b)

Generic liftToCon function.