Grisette.Internal.Core.Data.Class.ToSym

Description

Converting to symbolic values

class ToSym a b where

Convert a concrete value to symbolic value.

Methods

toSym :: a -> b

Convert a concrete value to symbolic value.

>>> toSym False :: SymBool
false

>>> toSym [False, True] :: [SymBool]
[false,true]

class (forall a b. ToSym a b => ToSym (f1 a) (f2 b)) => ToSym1 (f1 :: Type -> Type) (f2 :: Type -> Type) where

Lifting of ToSym to unary type constructors.

Methods

liftToSym :: (a -> b) -> f1 a -> f2 b

Lift a conversion to symbolic function to unary type constructors.

toSym1 :: (ToSym1 f1 f2, ToSym a b) => f1 a -> f2 b

Lift the standard toSym to unary type constructors.

class (forall a b. ToSym a b => ToSym1 (f1 a) (f2 b)) => ToSym2 (f1 :: Type -> Type -> Type) (f2 :: Type -> Type -> Type) where

Lifting of ToSym to binary type constructors.

Methods

liftToSym2 :: (a -> b) -> (c -> d) -> f1 a c -> f2 b d

Lift conversion to symbolic functions to binary type constructors.

toSym2 :: (ToSym2 f1 f2, ToSym a b, ToSym c d) => f1 a c -> f2 b d

Lift the standard toSym to binary type constructors.

Generic ToSym

data family ToSymArgs arity a b

The arguments to the generic toSym function.

class GToSym arity (f1 :: Type -> Type) (f2 :: Type -> Type) where

The class of types that can be generically converted to symbolic values.

Methods

gtoSym :: ToSymArgs arity a b -> f1 a -> f2 b

genericToSym :: (Generic a, Generic b, GToSym Arity0 (Rep a) (Rep b)) => a -> b

Generic toSym function.

genericLiftToSym :: (Generic1 f1, Generic1 f2, GToSym Arity1 (Rep1 f1) (Rep1 f2)) => (a -> b) -> f1 a -> f2 b

Generic liftToSym function.