Grisette.Internal.Core.Data.MemoUtils

Description

Memoization

stableMemo :: (a -> b) -> a -> b

Memoize a unary function.

stableMemo2 :: (a -> b -> c) -> a -> b -> c

Curried memoization to share partial evaluation

stableMemo3 :: (a -> b -> c -> d) -> a -> b -> c -> d

Curried memoization to share partial evaluation

stableMup :: (b -> c) -> (a -> b) -> a -> c

Lift a memoizer to work with one more argument.

stableMemoFix :: ((a -> b) -> a -> b) -> a -> b

Memoizing recursion. Use like fix.

weakStableMemo :: (a -> b) -> a -> b

Memoize a unary function.

weakStableMemo2 :: (a -> b -> c) -> a -> b -> c

Curried memoization to share partial evaluation

weakStableMemo3 :: (a -> b -> c -> d) -> a -> b -> c -> d

Curried memoization to share partial evaluation

weakStableMup :: (b -> c) -> (a -> b) -> a -> c

Lift a memoizer to work with one more argument.

weakStableMemoFix :: ((a -> b) -> a -> b) -> a -> b

Memoizing recursion. Use like fix.

htmemo :: (Eq k, Hashable k) => (k -> a) -> k -> a

Function memoizer with mutable hash table.

htmemo2 :: (Eq k1, Hashable k1, Eq k2, Hashable k2) => (k1 -> k2 -> a) -> k1 -> k2 -> a

Function memoizer with mutable hash table. Works on binary functions.

htmemo3 :: (Eq k1, Hashable k1, Eq k2, Hashable k2, Eq k3, Hashable k3) => (k1 -> k2 -> k3 -> a) -> k1 -> k2 -> k3 -> a

Function memoizer with mutable hash table. Works on ternary functions.

htmemoFix :: (Eq k, Hashable k) => ((k -> a) -> k -> a) -> k -> a

Memoizing recursion. Use like fix.

htmup :: (Eq k, Hashable k) => (b -> c) -> (k -> b) -> k -> c

Lift a memoizer to work with one more argument.