Grisette.Internal.Core.Data.UnionBase

Description

The union data structure.

Please consider using Union instead.

data UnionBase a where

The base union implementation, which is an if-then-else tree structure.

Constructors

UnionSingle :: forall a. a -> UnionBase a
A single value

UnionIf
A if value

Fields

:: forall a. a

Cached leftmost value

-> !Bool

Is merged invariant already maintained?

-> !SymBool

If condition

-> UnionBase a

True branch

-> UnionBase a

False branch

-> UnionBase a

ifWithLeftMost :: Bool -> SymBool -> UnionBase a -> UnionBase a -> UnionBase a

Build UnionIf with leftmost cache correctly maintained.

Usually you should never directly try to build a UnionIf with its constructor.

ifWithStrategy :: MergingStrategy a -> SymBool -> UnionBase a -> UnionBase a -> UnionBase a

Use a specific strategy to build a UnionIf value.

The merged invariant will be maintained in the result.

fullReconstruct :: MergingStrategy a -> UnionBase a -> UnionBase a

Fully reconstruct a Union to maintain the merged invariant.