Grisette.Internal.SymPrim.BV

Description

Documentation

newtype IntN (n :: Nat)

Signed bit vector type. Indexed with the bit width. Signedness affects the semantics of the operations, including comparison/extension, etc.

>>> 3 + 5 :: IntN 5
0b01000
>>> sizedBVConcat (0b101 :: IntN 3) (0b110 :: IntN 3)
0b101110
>>> sizedBVExt (Proxy @6) (0b101 :: IntN 3)
0b111101
>>> (8 :: IntN 4) < (7 :: IntN 4)
True

More operations are available. Please refer to Grisette.Core for more information.

Constructors

IntN

Fields

unIntN :: Integer

type IntN8 = IntN 8

8-bit signed bit-vector

type IntN16 = IntN 16

16-bit signed bit-vector

type IntN32 = IntN 32

32-bit signed bit-vector

type IntN64 = IntN 64

64-bit signed bit-vector

newtype WordN (n :: Nat)

Unsigned bit vector type. Indexed with the bit width. Signedness affect the semantics of the operations, including comparison/extension, etc.

>>> 3 + 5 :: WordN 5
0b01000
>>> sizedBVConcat (0b101 :: WordN 3) (0b110 :: WordN 3)
0b101110
>>> sizedBVExt (Proxy @6) (0b101 :: WordN 3)
0b000101
>>> (8 :: WordN 4) < (7 :: WordN 4)
False

More operations are available. Please refer to Grisette.Core for more information.

Constructors

WordN

Fields

unWordN :: Integer

type WordN8 = WordN 8

8-bit unsigned bit-vector

type WordN16 = WordN 16

16-bit unsigned bit-vector

type WordN32 = WordN 32

32-bit unsigned bit-vector

type WordN64 = WordN 64

64-bit unsigned bit-vector

readBinary :: Num a => ReadPrec a

Read a binary number.