Grisette.Internal.SymPrim.FP

Description

Documentation

type ValidFP (eb :: Nat) (sb :: Nat) = ValidFloat eb sb

A type-level proof that the given bit-widths are valid for a floating-point number.

newtype FP (eb :: Nat) (sb :: Nat)

IEEE 754 floating-point number with eb exponent bits and sb significand bits.

>>> 1.0 + 2.0 :: FP 11 53
3.0

More operations are available. Please refer to Grisette.Core for more information.

Constructors

FP

Fields

unFP :: FloatingPoint eb sb

type FP16 = FP 5 11

IEEE 754 half-precision floating-point number.

type FP32 = FP 8 24

IEEE 754 single-precision floating-point number.

type FP64 = FP 11 53

IEEE 754 double-precision floating-point number.

withValidFPProofs :: forall (eb :: Nat) (sb :: Nat) r. ValidFP eb sb => ((KnownNat (eb + sb), BVIsNonZero (eb + sb), 1 <= (eb + sb), 1 <= eb, 1 <= sb) => r) -> r

Some type-level witnesses that could be derived from ValidFP.

data FPRoundingMode

Rounding mode for floating-point operations.

Constructors

RNE
Round to nearest, ties to even.

RNA
Round to nearest, ties to away from zero.

RTP
Round towards positive infinity.

RTN
Round towards negative infinity.

RTZ
Round towards zero.

allFPRoundingMode :: [FPRoundingMode]

All IEEE 754 rounding modes.

data NotRepresentableFPError

An error thrown when bitcasting or converting FP NaN to other types.

Constructors

NaNError

FPUnderflowError

FPOverflowError

class ConvertibleBound (bv :: Nat -> Type) where

Bounds for converting bit vectors to floating-point numbers. Out-of-range FP values cannot be converted to a representable bit-vector.

Methods

convertibleLowerBound :: forall (eb :: Nat) (sb :: Nat) (n :: Nat). (ValidFP eb sb, KnownNat n, 1 <= n) => bv n -> FPRoundingMode -> FP eb sb

convertibleUpperBound :: forall (eb :: Nat) (sb :: Nat) (n :: Nat). (ValidFP eb sb, KnownNat n, 1 <= n) => bv n -> FPRoundingMode -> FP eb sb

nextFP :: forall (eb :: Nat) (sb :: Nat). ValidFP eb sb => FP eb sb -> FP eb sb

Next representable floating-point number.

Note:

nextFP(+inf) = +inf
nextFP(-inf) = -maxNormalized
nextFP(NaN) = NaN

The function do not distinguish between -0 and +0.

prevFP :: forall (eb :: Nat) (sb :: Nat). ValidFP eb sb => FP eb sb -> FP eb sb

Previous representable floating-point number.

Note:

prevFP(+inf) = +maxNormalized
prevFP(-inf) = -inf
prevFP(NaN) = NaN

The function do not distinguish between -0 and +0.

withUnsafeValidFP :: forall (eb :: Nat) (sb :: Nat) r. (KnownNat eb, KnownNat sb) => (ValidFP eb sb => r) -> r

Provide an (unsafe) type-level proof that the given floating-point type is valid.

checkDynamicValidFP :: Natural -> Natural -> Bool

Check if the given floating-point type is valid.

invalidFPMessage :: String

A message thrown when the floating-point type is invalid.