Grisette.Internal.SymPrim.FunInstanceGen

Description

Documentation

supportedPrimFun :: ExpQ -> ExpQ -> ExpQ -> ([TypeQ] -> ExpQ) -> String -> String -> Name -> Int -> DecsQ

Generate an instance of SupportedPrim for a function with a given number of arguments.

supportedPrimFunUpTo :: ExpQ -> ExpQ -> ExpQ -> ([TypeQ] -> ExpQ) -> String -> String -> Name -> Int -> DecsQ

Generate instances of SupportedPrim for functions with up to a given number of arguments.