Grisette.Internal.SymPrim.ModelRep

Description

Documentation

data ModelSymPair ct st where

A pair of a symbolic constant and its value. This is used to build a model from a list of symbolic constants and their values.

>>> buildModel ("a" := (1 :: Integer), "b" := True) :: Model
Model {a -> 1 :: Integer, b -> true :: Bool}

Constructors

(:=) :: forall ct st. LinkedRep ct st => st -> ct -> ModelSymPair ct st