Grisette.Internal.SymPrim.Prim.Internal.Instances.PEvalDivModIntegralTerm

Description

Documentation

pevalDefaultDivIntegralTerm :: (PEvalDivModIntegralTerm a, Integral a) => Term a -> Term a -> Term a

Default partial evaluation of division operation for integral types.

pevalDefaultDivBoundedIntegralTerm :: (PEvalDivModIntegralTerm a, Bounded a, Integral a) => Term a -> Term a -> Term a

Default partial evaluation of division operation for bounded integral types.

pevalDefaultModIntegralTerm :: (PEvalDivModIntegralTerm a, Integral a) => Term a -> Term a -> Term a

Default partial evaluation of modulo operation for integral types.

pevalDefaultQuotIntegralTerm :: (PEvalDivModIntegralTerm a, Integral a) => Term a -> Term a -> Term a

Default partial evaluation of quotient operation for integral types.

pevalDefaultQuotBoundedIntegralTerm :: (PEvalDivModIntegralTerm a, Bounded a, Integral a) => Term a -> Term a -> Term a

Default partial evaluation of quotient operation for bounded integral types.

pevalDefaultRemIntegralTerm :: (PEvalDivModIntegralTerm a, Integral a) => Term a -> Term a -> Term a

Default partial evaluation of remainder operation for integral types.

Orphan instances

PEvalDivModIntegralTerm Integer

Instance details

Methods

pevalDivIntegralTerm :: Term Integer -> Term Integer -> Term Integer

pevalModIntegralTerm :: Term Integer -> Term Integer -> Term Integer

pevalQuotIntegralTerm :: Term Integer -> Term Integer -> Term Integer

pevalRemIntegralTerm :: Term Integer -> Term Integer -> Term Integer

withSbvDivModIntegralTermConstraint :: (SDivisible (SBVType Integer) => r) -> r

sbvDivIntegralTerm :: SBVType Integer -> SBVType Integer -> SBVType Integer

sbvModIntegralTerm :: SBVType Integer -> SBVType Integer -> SBVType Integer

sbvQuotIntegralTerm :: SBVType Integer -> SBVType Integer -> SBVType Integer

sbvRemIntegralTerm :: SBVType Integer -> SBVType Integer -> SBVType Integer

(KnownNat n, 1 <= n) => PEvalDivModIntegralTerm (IntN n)

Instance details

Methods

pevalDivIntegralTerm :: Term (IntN n) -> Term (IntN n) -> Term (IntN n)

pevalModIntegralTerm :: Term (IntN n) -> Term (IntN n) -> Term (IntN n)

pevalQuotIntegralTerm :: Term (IntN n) -> Term (IntN n) -> Term (IntN n)

pevalRemIntegralTerm :: Term (IntN n) -> Term (IntN n) -> Term (IntN n)

withSbvDivModIntegralTermConstraint :: (SDivisible (SBVType (IntN n)) => r) -> r

sbvDivIntegralTerm :: SBVType (IntN n) -> SBVType (IntN n) -> SBVType (IntN n)

sbvModIntegralTerm :: SBVType (IntN n) -> SBVType (IntN n) -> SBVType (IntN n)

sbvQuotIntegralTerm :: SBVType (IntN n) -> SBVType (IntN n) -> SBVType (IntN n)

sbvRemIntegralTerm :: SBVType (IntN n) -> SBVType (IntN n) -> SBVType (IntN n)

(KnownNat n, 1 <= n) => PEvalDivModIntegralTerm (WordN n)

Instance details

Methods

pevalDivIntegralTerm :: Term (WordN n) -> Term (WordN n) -> Term (WordN n)

pevalModIntegralTerm :: Term (WordN n) -> Term (WordN n) -> Term (WordN n)

pevalQuotIntegralTerm :: Term (WordN n) -> Term (WordN n) -> Term (WordN n)

pevalRemIntegralTerm :: Term (WordN n) -> Term (WordN n) -> Term (WordN n)

withSbvDivModIntegralTermConstraint :: (SDivisible (SBVType (WordN n)) => r) -> r

sbvDivIntegralTerm :: SBVType (WordN n) -> SBVType (WordN n) -> SBVType (WordN n)

sbvModIntegralTerm :: SBVType (WordN n) -> SBVType (WordN n) -> SBVType (WordN n)

sbvQuotIntegralTerm :: SBVType (WordN n) -> SBVType (WordN n) -> SBVType (WordN n)

sbvRemIntegralTerm :: SBVType (WordN n) -> SBVType (WordN n) -> SBVType (WordN n)