Grisette.Internal.SymPrim.Prim.Internal.Instances.PEvalRotateTerm

Description

Documentation

pevalFiniteBitsSymRotateRotateLeftTerm :: (Integral a, SymRotate a, FiniteBits a, PEvalRotateTerm a) => Term a -> Term a -> Term a

Partial evaluation of symbolic rotate left term for finite bits types.

pevalFiniteBitsSymRotateRotateRightTerm :: (Integral a, SymRotate a, FiniteBits a, PEvalRotateTerm a) => Term a -> Term a -> Term a

Partial evaluation of symbolic rotate right term for finite bits types.

Orphan instances

(KnownNat n, 1 <= n) => PEvalRotateTerm (IntN n)

Instance details

Methods

pevalRotateLeftTerm :: Term (IntN n) -> Term (IntN n) -> Term (IntN n)

pevalRotateRightTerm :: Term (IntN n) -> Term (IntN n) -> Term (IntN n)

withSbvRotateTermConstraint :: (SIntegral (NonFuncSBVBaseType (IntN n)) => r) -> r

sbvRotateLeftTerm :: SBVType (IntN n) -> SBVType (IntN n) -> SBVType (IntN n)

sbvRotateRightTerm :: SBVType (IntN n) -> SBVType (IntN n) -> SBVType (IntN n)

(KnownNat n, 1 <= n) => PEvalRotateTerm (WordN n)

Instance details

Methods

pevalRotateLeftTerm :: Term (WordN n) -> Term (WordN n) -> Term (WordN n)

pevalRotateRightTerm :: Term (WordN n) -> Term (WordN n) -> Term (WordN n)

withSbvRotateTermConstraint :: (SIntegral (NonFuncSBVBaseType (WordN n)) => r) -> r

sbvRotateLeftTerm :: SBVType (WordN n) -> SBVType (WordN n) -> SBVType (WordN n)

sbvRotateRightTerm :: SBVType (WordN n) -> SBVType (WordN n) -> SBVType (WordN n)