Grisette.Internal.SymPrim.Prim.Internal.Instances.PEvalShiftTerm

Description

Documentation

pevalFiniteBitsSymShiftShiftLeftTerm :: (Integral a, SymShift a, FiniteBits a, PEvalShiftTerm a) => Term a -> Term a -> Term a

Partial evaluation of symbolic shift left term for finite bits types.

pevalFiniteBitsSymShiftShiftRightTerm :: (Integral a, SymShift a, FiniteBits a, PEvalShiftTerm a) => Term a -> Term a -> Term a

Partial evaluation of symbolic shift right term for finite bits types.

Orphan instances

(KnownNat n, 1 <= n) => PEvalShiftTerm (IntN n)

Instance details

Methods

pevalShiftLeftTerm :: Term (IntN n) -> Term (IntN n) -> Term (IntN n)

pevalShiftRightTerm :: Term (IntN n) -> Term (IntN n) -> Term (IntN n)

withSbvShiftTermConstraint :: (SIntegral (NonFuncSBVBaseType (IntN n)) => r) -> r

sbvShiftLeftTerm :: SBVType (IntN n) -> SBVType (IntN n) -> SBVType (IntN n)

sbvShiftRightTerm :: SBVType (IntN n) -> SBVType (IntN n) -> SBVType (IntN n)

(KnownNat n, 1 <= n) => PEvalShiftTerm (WordN n)

Instance details

Methods

pevalShiftLeftTerm :: Term (WordN n) -> Term (WordN n) -> Term (WordN n)

pevalShiftRightTerm :: Term (WordN n) -> Term (WordN n) -> Term (WordN n)

withSbvShiftTermConstraint :: (SIntegral (NonFuncSBVBaseType (WordN n)) => r) -> r

sbvShiftLeftTerm :: SBVType (WordN n) -> SBVType (WordN n) -> SBVType (WordN n)

sbvShiftRightTerm :: SBVType (WordN n) -> SBVType (WordN n) -> SBVType (WordN n)