Grisette.Internal.SymPrim.Prim.Internal.Utils

Description

Documentation

pattern Dyn :: (Typeable a, Typeable b) => a -> b

Pattern synonym for dynamic type casting.

cmpHetero :: (Typeable a, Typeable b) => (a -> a -> Bool) -> a -> b -> Bool

Compare two values of different types.

eqHetero :: (Typeable a, Typeable b, Eq a) => a -> b -> Bool

Compare two values of different types for equality.

cmpHeteroRep :: TypeRep a -> TypeRep b -> (a -> a -> Bool) -> a -> b -> Bool

Compare two values of different types, resolve the type equality using the type representation.

eqHeteroRep :: Eq a => TypeRep a -> TypeRep b -> a -> b -> Bool

Compare two values of different types for equality, resolve the type equality using the type representation.

eqTypeRepBool :: forall ka kb (a :: ka) (b :: kb). TypeRep a -> TypeRep b -> Bool

Compare two type representations for equality.

type WeakThreadId = Word64

A weak identifier to a thread id that doesn't prevent its garbage collection.

weakThreadId :: ThreadId -> Word64

Get an id of a thread that doesn't prevent its garbage collection.

type WeakThreadIdRef = Weak ThreadId

A weak reference to a thread id that doesn't prevent its garbage collection.

myWeakThreadId :: IO WeakThreadId

Get a weak identifier to the current thread id.

weakThreadRefAlive :: WeakThreadIdRef -> IO Bool

Check if a weak reference to a thread id is still alive.

mkWeakThreadIdRefWithFinalizer :: ThreadId -> IO () -> IO (Weak ThreadId)

Create a weak reference to a thread id with a finalizer.

addStableNameFinalizer :: StableName a -> IO () -> IO ()

Add a finalizer to a stable name.

addThreadIdFinalizer :: ThreadId -> IO () -> IO ()

Add a finalizer to a thread id.

mkWeakStableNameRefWithFinalizer :: StableName a -> IO () -> IO (Weak (StableName a))

Create a weak reference to a stable name with a finalizer.

data SomeStableName where

A type-erased stable name.

Constructors

SomeStableName :: forall a. StableName a -> SomeStableName

mkWeakSomeStableNameRefWithFinalizer :: SomeStableName -> IO () -> IO (Weak SomeStableName)

Create a weak reference to a stable name with a finalizer.

mkWeakSomeStableNameRef :: SomeStableName -> IO (Weak SomeStableName)

Create a weak reference to a stable name.