Grisette.Internal.SymPrim.Prim.Model

Description

Documentation

newtype SymbolSet (knd :: SymbolKind)

Set of symbols.

Check SymbolSetOps for operations, and SymbolSetRep for manual constructions.

Constructors

SymbolSet

Fields

unSymbolSet :: HashSet (SomeTypedSymbol knd)

type ConstantSymbolSet = SymbolSet 'ConstantKind

Set of constant symbols. Excluding unintepreted functions.

type AnySymbolSet = SymbolSet 'AnyKind

Set of any symbols.

newtype Model

Model returned by the solver.

Check ModelOps for operations, and ModelRep for manual constructions.

Constructors

Model

Fields

unModel :: HashMap SomeTypedAnySymbol ModelValue

data ModelValuePair t

A type used for building a model by hand.

>>> buildModel ("x" ::= (1 :: Integer), "y" ::= True) :: Model
Model {x -> 1 :: Integer, y -> true :: Bool}

Constructors

(TypedAnySymbol t) ::= t

equation :: TypedAnySymbol a -> Model -> Maybe (Term Bool)

Given a typed symbol and a model, return the equation (symbol = value) encoded in the model.

evalTerm :: SupportedPrim a => Bool -> Model -> HashSet SomeTypedConstantSymbol -> Term a -> Term a

Evaluate a term in the given model.

Orphan instances

(ModelRep a Model, ModelRep b Model) => ModelRep (a, b) Model

Instance details

Methods

buildModel :: (a, b) -> Model

(ModelRep a Model, ModelRep b Model, ModelRep c Model) => ModelRep (a, b, c) Model

Instance details

Methods

buildModel :: (a, b, c) -> Model

(ModelRep a Model, ModelRep b Model, ModelRep c Model, ModelRep d Model) => ModelRep (a, b, c, d) Model

Instance details

Methods

buildModel :: (a, b, c, d) -> Model

(ModelRep a Model, ModelRep b Model, ModelRep c Model, ModelRep d Model, ModelRep e Model) => ModelRep (a, b, c, d, e) Model

Instance details

Methods

buildModel :: (a, b, c, d, e) -> Model

(ModelRep a Model, ModelRep b Model, ModelRep c Model, ModelRep d Model, ModelRep e Model, ModelRep f Model) => ModelRep (a, b, c, d, e, f) Model

Instance details

Methods

buildModel :: (a, b, c, d, e, f) -> Model

(ModelRep a Model, ModelRep b Model, ModelRep c Model, ModelRep d Model, ModelRep e Model, ModelRep f Model, ModelRep g Model) => ModelRep (a, b, c, d, e, f, g) Model

Instance details

Methods

buildModel :: (a, b, c, d, e, f, g) -> Model

(ModelRep a Model, ModelRep b Model, ModelRep c Model, ModelRep d Model, ModelRep e Model, ModelRep f Model, ModelRep g Model, ModelRep h Model) => ModelRep (a, b, c, d, e, f, g, h) Model

Instance details

Methods

buildModel :: (a, b, c, d, e, f, g, h) -> Model