Grisette.Internal.SymPrim.SymBV

Description

Documentation

newtype SymWordN (n :: Nat)

Symbolic unsigned bit vector type. Indexed with the bit width. Signedness affects the semantics of the operations, including comparison/extension, etc.

>>> "a" + 5 :: SymWordN 5
(+ 0b00101 a)
>>> sizedBVConcat (con 0b101 :: SymWordN 3) (con 0b110 :: SymWordN 3)
0b101110
>>> sizedBVExt (Proxy @6) (con 0b101 :: SymWordN 3)
0b000101
>>> (8 :: SymWordN 4) .< (7 :: SymWordN 4)
false

More operations are available. Please refer to Grisette.Core for more information.

Constructors

SymWordN (Term (WordN n))

type SymWordN8 = SymWordN 8

Symbolic 8-bit unsigned bit-vector.

type SymWordN16 = SymWordN 16

Symbolic 16-bit unsigned bit-vector.

type SymWordN32 = SymWordN 32

Symbolic 32-bit unsigned bit-vector.

type SymWordN64 = SymWordN 64

Symbolic 64-bit unsigned bit-vector.

newtype SymIntN (n :: Nat)

Symbolic signed bit vector type. Indexed with the bit width. Signedness affects the semantics of the operations, including comparison/extension, etc.

>>> "a" + 5 :: SymIntN 5
(+ 0b00101 a)
>>> sizedBVConcat (con 0b101 :: SymIntN 3) (con 0b110 :: SymIntN 3)
0b101110
>>> sizedBVExt (Proxy @6) (con 0b101 :: SymIntN 3)
0b111101
>>> (8 :: SymIntN 4) .< (7 :: SymIntN 4)
true

More operations are available. Please refer to Grisette.Core for more information.

Constructors

SymIntN (Term (IntN n))

type SymIntN8 = SymIntN 8

Symbolic 8-bit signed bit-vector.

type SymIntN16 = SymIntN 16

Symbolic 16-bit signed bit-vector.

type SymIntN32 = SymIntN 32

Symbolic 32-bit signed bit-vector.

type SymIntN64 = SymIntN 64

Symbolic 64-bit signed bit-vector.

Orphan instances

(KnownNat n, 1 <= n) => SymRep (IntN n)

Instance details

Associated Types

type SymType (IntN n)

Instance details

Defined in Grisette.Internal.SymPrim.SymBV

type SymType (IntN n) = SymIntN n

(KnownNat n, 1 <= n) => SymRep (WordN n)

Instance details

Associated Types

type SymType (WordN n)

Instance details

Defined in Grisette.Internal.SymPrim.SymBV

type SymType (WordN n) = SymWordN n

(KnownNat n, 1 <= n) => LinkedRep (IntN n) (SymIntN n)

Instance details

Methods

underlyingTerm :: SymIntN n -> Term (IntN n)

wrapTerm :: Term (IntN n) -> SymIntN n

(KnownNat n, 1 <= n) => LinkedRep (WordN n) (SymWordN n)

Instance details

Methods

underlyingTerm :: SymWordN n -> Term (WordN n)

wrapTerm :: Term (WordN n) -> SymWordN n