Grisette.Internal.SymPrim.SymPrim

Description

Documentation

type Prim a = (Show a, Binary a, Serial a, Serialize a, NFData a, Eq a, EvalSym a, ExtractSym a, Mergeable a, PPrint a, SubstSym a, SymEq a, SymOrd a, AllSyms a, Hashable a, Lift a, Typeable a)

A type that is used as a constraint for all the primitive types (including concrete primitives) in Grisette.

type SymPrim a = (Prim a, ITEOp a, GenSymSimple a a)

A type that is used as a constraint for all the symbolic primitive types in Grisette.

type BasicSymPrim a = (SymPrim a, SimpleMergeable a, GenSymSimple () a, Apply a, Solvable (ConType a) a, ConRep a, LinkedRep (ConType a) a, ToCon a (ConType a), ToSym (ConType a) a)

A type that is used as a constraint for all the basic symbolic primitive types in Grisette.

SomeSymWordN is not considered as a basic symbolic primitive type.