Grisette.Internal.TH.Ctor.Common

Description

Documentation

withNameTransformer

Arguments

:: ([String] -> Name -> Q [Dec])
A function that generates decs given a list of constructor names and a type name

-> (String -> String)
A function that transforms constructor names

-> Name
The type to generate the wrappers for

-> Q [Dec]

Generate smart constructor given a type name, using a name transformer to transform constructor names.

prefixTransformer :: String -> String -> String

A name transformer that prefixes a string to the constructor name

decapitalizeTransformer :: String -> String

A name transformer that converts the first character to lowercase