Grisette.Internal.TH.Derivation.DeriveMergeable

Description

Documentation

deriveMergeable :: DeriveConfig -> Name -> Q [Dec]

Derive Mergeable instance for GADT.

deriveMergeable1 :: DeriveConfig -> Name -> Q [Dec]

Derive Mergeable1 instance for GADT.

deriveMergeable2 :: DeriveConfig -> Name -> Q [Dec]

Derive Mergeable2 instance for GADT.

deriveMergeable3 :: DeriveConfig -> Name -> Q [Dec]

Derive Mergeable3 instance for GADT.

genMergeableAndGetMergingInfoResult :: DeriveConfig -> Name -> Int -> Q (MergingInfoResult, [Dec])

Generate Mergeable instance and merging information for a data type.

genMergeable :: DeriveConfig -> Name -> Int -> Q [Dec]

Generate Mergeable instance for a data type.

genMergeable' :: DeriveConfig -> MergingInfoResult -> Name -> Int -> Q (Name, [Dec])

Generate Mergeable instance for a data type, using a given merging info result.

genMergeableNoExistential :: DeriveConfig -> Name -> Int -> Q [Dec]

Generate Mergeable instance for a data type without existential variables.

genMergeableNoStrategy :: DeriveConfig -> Name -> Int -> Q [Dec]

Generate Mergeable instance for a data type, using NoStrategy.

genMergeableList :: DeriveConfig -> Name -> [Int] -> Q [Dec]

Generate multiple Mergeable instances for a data type.