Grisette.Internal.TH.GADT.BinaryOpCommon

Description

Documentation

data BinaryOpClassConfig

Configuration for a binary operation type class generation on a GADT.

Constructors

BinaryOpClassConfig

Fields

binaryOpFieldConfigs :: [BinaryOpFieldConfig]

binaryOpInstanceNames :: [Name]

binaryOpAllowSumType :: Bool

binaryOpAllowExistential :: Bool

data BinaryOpFieldConfig

Configuration for a binary operation field generation on a GADT.

Constructors

BinaryOpFieldConfig

Fields

extraPatNames :: [String]

fieldResFun :: [Exp] -> (Exp, Exp) -> Exp -> Q (Exp, [Bool])

fieldCombineFun :: Name -> [Exp] -> Q (Exp, [Bool])

fieldDifferentExistentialFun :: Exp -> Q Exp

fieldLMatchResult :: Q Exp

fieldRMatchResult :: Q Exp

fieldFunExp :: FieldFunExp

fieldFunNames :: [Name]

type FieldFunExp = Map Name Name -> Type -> Q Exp

Type of field function expression generator.

defaultFieldFunExp :: [Name] -> FieldFunExp

Default field function expression generator.

genBinaryOpClause :: BinaryOpFieldConfig -> [(Type, Kind)] -> [(Type, Kind)] -> Bool -> ConstructorInfo -> ConstructorInfo -> Q [Clause]

Generate a clause for a binary operation on a GADT.

genBinaryOpClass :: DeriveConfig -> BinaryOpClassConfig -> Int -> Name -> Q [Dec]

Generate a type class instance for a binary operation on a GADT.