Grisette.Internal.TH.GADT.Common

Description

Documentation

data CheckArgsResult

Result of checkArgs for a GADT.

Constructors

CheckArgsResult

Fields

constructors :: [ConstructorInfo]

keptVars :: [(Type, Kind)]

argVars :: [(Type, Kind)]

checkArgs :: String -> Int -> Name -> Bool -> Int -> Q CheckArgsResult

Check if the number of type parameters is valid for a GADT, and return new names for the type variables, split into kept and arg parts.

ctxForVar :: [Type] -> Type -> Kind -> Q (Maybe Pred)

Generate a context for a variable in a GADT.

data EvalModeConfig

Configuration for constraints for evaluation modes tag.

EvalModeConstraints specifies a list of constraints for the tag, for example, we may use EvalModeBase and EvalModeBV to specify that the evaluation mode must support both base (boolean and data types) and bit vectors. This should be used when the data type uses bit vectors.

EvalModeSpecified specifies a that an evaluation mode tag should be specialized to a specific tag for all the instances.

Constructors

EvalModeConstraints [Name]

EvalModeSpecified EvalModeTag

data DeriveConfig

Configuration for deriving instances for a GADT.

Constructors

DeriveConfig

Fields

evalModeConfig :: [(Int, EvalModeConfig)]

bitSizePositions :: [Int]

fpBitSizePositions :: [(Int, Int)]

unconstrainedPositions :: [Int]

needExtraMergeableUnderEvalMode :: Bool

needExtraMergeableWithConcretizedEvalMode :: Bool

useNoStrategy :: Bool

useSerialForCerealAndBinary :: Bool

extraEvalModeConstraint :: Name -> Name -> [(Type, Kind)] -> (Int, EvalModeConfig) -> Q [Pred]

Generate extra constraints for evaluation modes.

extraBitSizeConstraint :: Name -> Name -> [(Type, Kind)] -> Int -> Q [Pred]

Generate extra constraints for bit vectors.

extraFpBitSizeConstraint :: Name -> Name -> [(Type, Kind)] -> (Int, Int) -> Q [Pred]

Generate extra constraints for floating point exponents and significands.

extraExtraMergeableConstraint :: DeriveConfig -> [ConstructorInfo] -> [(Type, Kind)] -> Q [Pred]

Generate extra constraints for Mergeable instances.

extraConstraint :: DeriveConfig -> Name -> Name -> [(Type, Kind)] -> [(Type, Kind)] -> [ConstructorInfo] -> Q [Pred]

Generate extra constraints for a GADT.

specializeResult :: [(Int, EvalModeTag)] -> CheckArgsResult -> Q CheckArgsResult

Specialize the evaluation mode tags for the CheckArgsResult.

evalModeSpecializeList :: DeriveConfig -> [(Int, EvalModeTag)]

Get all the evaluation modes to specialize in the DeriveConfig.

isVarUsedInFields :: CheckArgsResult -> Name -> Bool

Check if a variable is used in the fields of a constructor.

freshenCheckArgsResult :: Bool -> CheckArgsResult -> Q CheckArgsResult

Freshen the type variables in the CheckArgsResult.