Grisette.Internal.TH.GADT.DeriveGADT

Description

Documentation

deriveGADT :: [Name] -> [Name] -> Q [Dec]

Derive the specified classes for a GADT with the given name.

See deriveGADTWith for more details.

deriveGADTWith :: DeriveConfig -> [Name] -> [Name] -> Q [Dec]

Derive the specified classes for a GADT with the given name.

Support the following classes.

Mergeable

Mergeable1

Mergeable2

Mergeable3

EvalSym

EvalSym1

EvalSym2

ExtractSym

ExtractSym1

ExtractSym2

SubstSym

SubstSym1

SubstSym2

NFData

NFData1

NFData2

Hashable

Hashable1

Hashable2

Show

Show1

Show2

PPrint

PPrint1

PPrint2

AllSyms

AllSyms1

AllSyms2

Eq

Eq1

Eq2

Ord

Ord1

Ord2

SymOrd

SymOrd1

SymOrd2

SymEq

SymEq1

SymEq2

UnifiedSymEq

UnifiedSymEq1

UnifiedSymEq2

UnifiedSymOrd

UnifiedSymOrd1

UnifiedSymOrd2

ToSym

ToSym1

ToSym2

ToCon

ToCon1

ToCon2

Serial

Serial1

Serial2

SimpleMergeable

SimpleMergeable1

SimpleMergeable2

Binary

Serialize

Note that the following type classes cannot be derived for GADTs with existential type variables.

ToCon

ToCon1

ToCon2

ToSym

ToSym1

ToSym2

Serial

Serial1

Serial2

Binary

Serialize

allClasses0 :: [Name]

All the classes that can be derived for GADTs.

This includes:

Mergeable

EvalSym

ExtractSym

SubstSym

NFData

Hashable

Show

PPrint

AllSyms

Eq

SymEq

SymOrd

UnifiedSymEq

Ord

UnifiedSymOrd

Serial

ToCon

ToSym

allClasses01 :: [Name]

All the classes that can be derived for GADT functors.

This includes all the classes in allClasses0 and allClasses1.

allClasses012 :: [Name]

All the classes that can be derived for GADTfunctors.

This includes all the classes in allClasses0, allClasses1, and allClasses2.

basicClasses0 :: [Name]

Basic classes for GADTs.

This includes:

Mergeable

EvalSym

ExtractSym

SubstSym

NFData

Hashable

Show

PPrint

AllSyms

Eq

SymEq

SymOrd

UnifiedSymEq

These classes can be derived for most GADTs.

noExistentialClasses0 :: [Name]

Classes that can only be derived for GADTs without existential type variables.

This includes:

Serial

Serialize

Binary

ToCon

ToSym

concreteOrdClasses0 :: [Name]

Concrete ordered classes that can be derived for GADTs that

uses unified evaluation mode, or

does not contain any symbolic variables.

This includes:

Ord

UnifiedSymOrd

basicClasses1 :: [Name]

Basic classes for GADT functors.

This includes:

Mergeable1

EvalSym1

ExtractSym1

SubstSym1

NFData1

Hashable1

Show1

PPrint1

AllSyms1

Eq1

SymEq1

SymOrd1

UnifiedSymEq1

noExistentialClasses1 :: [Name]

*1 classes that can only be derived for GADT functors without existential type variables.

This includes:

Serial1

ToCon1

ToSym1

concreteOrdClasses1 :: [Name]

*1 concrete ordered classes that can be derived for GADT functors that

uses unified evaluation mode, or

does not contain any symbolic variables.

This includes:

Ord1

UnifiedSymOrd1

basicClasses2 :: [Name]

Basic classes for GADT functors.

This includes:

Mergeable2

EvalSym2

ExtractSym2

SubstSym2

NFData2

Hashable2

Show2

PPrint2

AllSyms2

Eq2

SymEq2

SymOrd2

UnifiedSymEq2

noExistentialClasses2 :: [Name]

*2 classes that can only be derived for GADT functors without existential type variables.

This includes:

Serial2

ToCon2

ToSym2

concreteOrdClasses2 :: [Name]

*2 concrete ordered classes that can be derived for GADT functors that

uses unified evaluation mode, or

does not contain any symbolic variables.

This includes:

Ord2

UnifiedSymOrd2

showClasses :: [Name]

Show classes that can be derived for GADTs.

This includes:

Show

Show1

Show2

pprintClasses :: [Name]

PPrint classes that can be derived for GADTs.

This includes:

PPrint

PPrint1

PPrint2

evalSymClasses :: [Name]

EvalSym classes that can be derived for GADTs.

This includes:

EvalSym

EvalSym1

EvalSym2

extractSymClasses :: [Name]

ExtractSym classes that can be derived for GADTs.

This includes:

ExtractSym

ExtractSym1

ExtractSym2

substSymClasses :: [Name]

SubstSym classes that can be derived for GADTs.

This includes:

SubstSym

SubstSym1

SubstSym2

allSymsClasses :: [Name]

AllSyms classes that can be derived for GADTs.

This includes:

AllSyms

AllSyms1

AllSyms2

eqClasses :: [Name]

Eq classes that can be derived for GADTs.

This includes:

Eq

Eq1

Eq2

ordClasses :: [Name]

Ord classes that can be derived for GADTs.

This includes:

Ord

Ord1

Ord2

symOrdClasses :: [Name]

SymOrd classes that can be derived for GADTs.

This includes:

SymOrd

SymOrd1

SymOrd2

symEqClasses :: [Name]

SymEq classes that can be derived for GADTs.

This includes:

SymEq

SymEq1

SymEq2

unifiedSymOrdClasses :: [Name]

UnifiedSymOrd classes that can be derived for GADTs.

This includes:

UnifiedSymOrd

UnifiedSymOrd1

UnifiedSymOrd2

unifiedSymEqClasses :: [Name]

UnifiedSymEq classes that can be derived for GADTs.

This includes:

UnifiedSymEq

UnifiedSymEq1

UnifiedSymEq2

mergeableClasses :: [Name]

Mergeable classes that can be derived for GADTs.

This includes:

Mergeable

Mergeable1

Mergeable2

Mergeable3

nfDataClasses :: [Name]

NFData classes that can be derived for GADTs.

This includes:

NFData

NFData1

NFData2

hashableClasses :: [Name]

Hashable classes that can be derived for GADTs.

This includes:

Hashable

Hashable1

Hashable2

toSymClasses :: [Name]

ToSym classes that can be derived for GADTs.

This includes:

ToSym

ToSym1

ToSym2

toConClasses :: [Name]

ToCon classes that can be derived for GADTs.

This includes:

ToCon

ToCon1

ToCon2

serialClasses :: [Name]

Serial classes that can be derived for GADTs.

This includes:

Serial

Serial1

Serial2

simpleMergeableClasses :: [Name]

SimpleMergeable classes that can be derived for GADTs.

This includes:

SimpleMergeable

SimpleMergeable1

SimpleMergeable2

unifiedSimpleMergeableClasses :: [Name]

UnifiedSimpleMergeable classes that can be derived for GADTs.

This includes:

UnifiedSimpleMergeable

UnifiedSimpleMergeable1

UnifiedSimpleMergeable2

filterExactNumArgs :: Int -> [Name] -> [Name]

Filter classes that accepts type constructors with exactly n arguments.

filterLeqNumArgs :: Int -> [Name] -> [Name]

Filter classes that accepts type constructors with at most n arguments.