Grisette.Internal.TH.GADT.UnaryOpCommon

Description

Documentation

data UnaryOpClassConfig

Configuration for a unary operation type class generation on a GADT.

Constructors

UnaryOpClassConfig

Fields

unaryOpConfigs :: [UnaryOpConfig]

unaryOpInstanceNames :: [Name]

unaryOpContextNames :: Maybe [Name]

unaryOpExtraVars :: DeriveConfig -> Q [(Type, Kind)]

unaryOpInstanceTypeFromConfig :: DeriveConfig -> [(Type, Kind)] -> [(Type, Kind)] -> Name -> Q Type

unaryOpAllowExistential :: Bool

data UnaryOpFieldConfig

Configuration for the derivation rules for a unary operation that can be derived by transforming each field and then combining the results.

Constructors

UnaryOpFieldConfig

Fields

extraPatNames :: [String]

extraLiftedPatNames :: Int -> [String]

fieldResFun :: ConstructorVariant -> Name -> [Exp] -> Int -> Exp -> Exp -> Q (Exp, [Bool])

fieldCombineFun :: Int -> Int -> ConstructorVariant -> Name -> [Exp] -> [Exp] -> Q (Exp, [Bool])

fieldFunExp :: FieldFunExp

data UnaryOpConfig where

Configuration for a unary function field expression generation on a GADT.

Constructors

UnaryOpConfig :: forall config. UnaryOpFunConfig config => config -> [Name] -> UnaryOpConfig

class UnaryOpFunConfig config where

Configuration for the derivation rules for a unary operation.

Methods

genUnaryOpFun

Arguments

:: DeriveConfig
Derive configuration

-> config
Configuration

-> [Name]
Function names

-> Int
Number of functor arguments to the class

-> [(Type, Kind)]
Extra variables

-> [(Type, Kind)]
Kept variables

-> [(Type, Kind)]
Argument variables

-> (Name -> Bool)
Whether the variable is used in fields

-> [ConstructorInfo]
Constructor infos

-> Q Dec

type FieldFunExp = Map Name Name -> Map Name [Name] -> Type -> Q Exp

Type of field function expression generator.

defaultFieldResFun :: ConstructorVariant -> Name -> [Exp] -> Int -> Exp -> Exp -> Q (Exp, [Bool])

Default field result function.

defaultFieldFunExp :: [Name] -> FieldFunExp

Default field function expression generator.

genUnaryOpClass :: DeriveConfig -> UnaryOpClassConfig -> Int -> Name -> Q [Dec]

Generate a unary operation type class instance for a GADT.

defaultUnaryOpInstanceTypeFromConfig :: DeriveConfig -> [(Type, Kind)] -> [(Type, Kind)] -> Name -> Q Type

Default unary operation instance type generator.