Grisette.Internal.TH.GADT.UnifiedOpCommon

Description

Documentation

newtype UnaryOpUnifiedConfig

Configuration for the derivation rules for a unified operation.

Constructors

UnaryOpUnifiedConfig

Fields

unifiedFun :: Type -> (Type, Kind) -> Q (Maybe Exp)

defaultUnaryOpUnifiedFun :: [Name] -> Type -> (Type, Kind) -> Q (Maybe Exp)

Default implementation for the derivation rules for a unified operation.