Grisette.Internal.TH.Util

Description

Documentation

constructorInfoToType :: DatatypeInfo -> ConstructorInfo -> Q Type

Convert a ConstructorInfo to a Type of the constructor.

tvIsMode :: TyVarBndr_ flag -> Bool

Check if a type variable is of kind EvalModeTag.

tvIsNat :: TyVarBndr_ flag -> Bool

Check if a type variable is of kind Nat.

tvIsStar :: TyVarBndr_ flag -> Bool

Check if a type variable is of kind Type.

tvIsStarToStar :: TyVarBndr_ flag -> Bool

Check if a type variable is of kind 'Data.Kind.Type -> Data.Kind.Type'.

substDataType :: DatatypeInfo -> Map Name Type -> DatatypeInfo

Substitute the type variables in a DatatypeInfo with the given substitution map.

reifyDatatypeWithFreshNames :: Name -> Q DatatypeInfo

Reify a datatype with fresh type variable names.

singleParamClassParamKind :: Name -> Q Kind

Get the kind of the single type parameter of a class.

binaryClassParamKind :: Name -> Q Kind

Get the kind of the binary type parameter of a class.

getTypeWithMaybeSubst :: TyVarBndrUnit -> Maybe Type -> Q Type

Get a type with a possible substitution.

dropLastTypeParam :: Type -> Q Type

Drop the last instantiated type parameter of a type.

dropNTypeParam :: Int -> Type -> Q Type

Drop the last N instantiated type parameters of a type.

classParamKinds :: Name -> Q [Kind]

Get the kinds of the type parameters of a class.

allSameKind :: [TyVarBndrUnit] -> Bool

Check if all type variables have the same kind.

classNumParam :: Name -> Q Int

Get the number of type parameters of a class.

kindNumParam :: Kind -> Q Int

Get the number of type parameters of a kind.

concatPreds :: Maybe [Pred] -> Maybe [Pred] -> Maybe [Pred]

Concatenate two 'Maybe [Pred]'.

putHaddock :: Name -> String -> Q ()

Put a haddock comment on a declaration.

allUsedNamesMaybe :: Maybe Exp -> Set Name

Get the names used in an expression.

allUsedNames :: Exp -> Set Name

Get the names used in an expression.

isNonUnitTupleString :: String -> Bool

Check if a string is the data constructor name of a non-unit tuple.

isNonUnitTuple :: Name -> Bool

Check if a name is the data constructor name of a non-unit tuple.

integerE :: Integral a => a -> Q Exp

Convert an integer to an Exp.

mangleName :: Name -> String

Mangle a name string to contain only alphanumeric characters and underscores.

dataTypeHasExistential :: Name -> Q Bool

Check if a data type has existential variables in constructors.