Grisette.Internal.Unified.BaseMonad

Description

Documentation

type family BaseMonad (mode :: EvalModeTag) = (r :: Type -> Type) | r -> mode where ...

A type family that specifies the base monad for the evaluation mode.

Resolves to Identity for C mode, and Union for S mode.

Equations

BaseMonad 'C = Identity

BaseMonad 'S = Union