Grisette.Internal.Unified.Class.UnifiedRep

Description

Documentation

class UnifiedConRep a

A class that gives the concrete type of a unified primitive type.

Associated Types

type ConType a

class UnifiedSymRep a

A class that gives the symbolic type of a unified primitive type.

Associated Types

type SymType a