Grisette.Internal.Unified.Class.UnifiedSafeBitCast

Description

Documentation

safeBitCast :: forall (mode :: EvalModeTag) e a b m. (MonadError e m, UnifiedSafeBitCast mode e a b m) => a -> m b

Unified safeSub operation.

This function isn't able to infer the mode, so you need to provide the mode explicitly. For example:

safeSub @mode a b

class UnifiedSafeBitCast (mode :: EvalModeTag) e a b (m :: Type -> Type) where

A class that provides unified safe bitcast operations.

We use this type class to help resolve the constraints for SafeBitCast.

Methods

withBaseSafeBitCast :: (SafeBitCast e a b m => r) -> r