Grisette.Internal.Unified.Class.UnifiedSafeFdiv

Description

Documentation

safeFdiv :: forall (mode :: EvalModeTag) e a m. (MonadError e m, UnifiedSafeFdiv mode e a m) => a -> a -> m a

Unified safeFdiv operation.

This function isn't able to infer the mode, so you need to provide the mode explicitly. For example:

safeFdiv @mode a b

class UnifiedSafeFdiv (mode :: EvalModeTag) e a (m :: Type -> Type) where

A class that provides unified floating division operations.

We use this type class to help resolve the constraints for SafeFdiv.

Methods

withBaseUnifiedSafeFdiv :: (SafeFdiv e a m => r) -> r