Grisette.Internal.Unified.Class.UnifiedSafeFromFP

Description

Documentation

class UnifiedSafeFromFP (mode :: EvalModeTag) e a fp fprd (m :: Type -> Type) where

A class that provides unified safe conversion from floating points.

We use this type class to help resolve the constraints for SafeFromFP.

Methods

withBaseSafeFromFP :: (SafeFromFP e a fp fprd m => r) -> r

safeFromFP :: forall (mode :: EvalModeTag) e a fp fprd m. (UnifiedSafeFromFP mode e a fp fprd m, MonadError e m) => fprd -> fp -> m a

Unified safeFromFP operation.

This function isn't able to infer the mode, so you need to provide the mode explicitly. For example:

safeFromFP @mode mode fp