Grisette.Internal.Unified.Class.UnifiedSafeSymRotate

Description

Documentation

safeSymRotateL :: forall (mode :: EvalModeTag) e a m. (MonadError e m, UnifiedSafeSymRotate mode e a m) => a -> a -> m a

Unified safeSymRotateL operation.

This function isn't able to infer the mode, so you need to provide the mode explicitly. For example:

safeSymRotateL @mode a b

safeSymRotateR :: forall (mode :: EvalModeTag) e a m. (MonadError e m, UnifiedSafeSymRotate mode e a m) => a -> a -> m a

Unified safeSymRotateR operation.

This function isn't able to infer the mode, so you need to provide the mode explicitly. For example:

safeSymRotateR @mode a b

class UnifiedSafeSymRotate (mode :: EvalModeTag) e a (m :: Type -> Type) where

A class that provides unified safe symbolic rotation operations.

We use this type class to help resolve the constraints for SafeSymRotate.

Methods

withBaseSafeSymRotate :: (SafeSymRotate e a m => r) -> r