Grisette.Internal.Unified.Class.UnifiedSafeSymShift

Description

Documentation

safeSymShiftL :: forall (mode :: EvalModeTag) e a m. (MonadError e m, UnifiedSafeSymShift mode e a m) => a -> a -> m a

Unified safeSymShiftL operation.

This function isn't able to infer the mode, so you need to provide the mode explicitly. For example:

safeSymShiftL @mode a b

safeSymShiftR :: forall (mode :: EvalModeTag) e a m. (MonadError e m, UnifiedSafeSymShift mode e a m) => a -> a -> m a

Unified safeSymShiftR operation.

This function isn't able to infer the mode, so you need to provide the mode explicitly. For example:

safeSymShiftR @mode a b

safeSymStrictShiftL :: forall (mode :: EvalModeTag) e a m. (MonadError e m, UnifiedSafeSymShift mode e a m) => a -> a -> m a

Unified safeSymStrictShiftL operation.

This function isn't able to infer the mode, so you need to provide the mode explicitly. For example:

safeSymStrictShiftL @mode a b

safeSymStrictShiftR :: forall (mode :: EvalModeTag) e a m. (MonadError e m, UnifiedSafeSymShift mode e a m) => a -> a -> m a

Unified safeSymStrictShiftR operation.

This function isn't able to infer the mode, so you need to provide the mode explicitly. For example:

safeSymStrictShiftR @mode a b

class UnifiedSafeSymShift (mode :: EvalModeTag) e a (m :: Type -> Type) where

A class that provides unified safe symbolic rotation operations.

We use this type class to help resolve the constraints for SafeSymShift.

Methods

withBaseSafeSymShift :: (SafeSymShift e a m => r) -> r