Grisette.Internal.Unified.Class.UnionViewMode

Documentation

class UnionViewMode (mode :: EvalModeTag) (u :: Type -> Type) | u -> mode

This class is used to determine the type of the Boolean used in the UnionView class.

For Identity, we use Bool as the Boolean type, and this ensures that the ifView function will return Nothing.

For Union, we use SymBool as the Boolean type, and ifView function can return Nothing or Just.