Grisette.Internal.Unified.Theories

Description

Documentation

data TheoryToUnify

This data type is used to represent the theories that is unified.

The UFun constructor is used to represent a specific uninterpreted function type. The type is uncurried.

Constructors

UBool

UIntN

UWordN

UInteger

UAlgReal

UFP

UFun [TheoryToUnify]

isUFun :: TheoryToUnify -> Bool

Check if the theory is a uninterpreted function.