Grisette.Internal.Unified.UnifiedData

Description

Documentation

class (UnifiedSimpleMergeable1 mode (GetData mode), UnifiedBranching mode (GetData mode), UnionView (GetData mode), UnionViewMode mode (GetData mode), Monad (GetData mode), TryMerge (GetData mode)) => UnifiedDataBase (mode :: EvalModeTag)

type family GetData (mode :: EvalModeTag) = (r :: Type -> Type) | r -> mode

Get a unified data type. Resolves to Identity in C mode, and Union in S mode.

type BaseMonad (mode :: EvalModeTag) = GetData mode

A type family that specifies the base monad for the evaluation mode.

Resolves to Identity for C mode, and Union for S mode.

wrapData :: UnifiedDataImpl mode v u => v -> u

Wraps a value into the unified data type.

extractData :: (UnifiedDataImpl mode v u, Mergeable v, Monad m, UnifiedBranching mode m) => u -> m v

Extracts a value from the unified data type.

class UnifiedDataImpl mode v (GetData mode v) => UnifiedData (mode :: EvalModeTag) v

This class is needed as constraint in user code prior to GHC 9.2.1. See the notes in IsMode.

class (forall v. UnifiedData bool v, forall v. Mergeable v => UnifiedDataSimpleMergeable v) => AllUnifiedData (bool :: EvalModeTag)

Evaluation mode with unified data types.

symCompare :: forall (mode :: EvalModeTag) a. (DecideEvalMode mode, UnifiedSymOrd mode a) => a -> a -> GetData mode Ordering

Unified symCompare.

liftSymCompare :: forall (mode :: EvalModeTag) f a b. (DecideEvalMode mode, UnifiedSymOrd1 mode f) => (a -> b -> GetData mode Ordering) -> f a -> f b -> GetData mode Ordering

Unified liftSymCompare.

symCompare1 :: forall (mode :: EvalModeTag) f a. (DecideEvalMode mode, UnifiedSymOrd mode a, UnifiedSymOrd1 mode f) => f a -> f a -> GetData mode Ordering

Unified symCompare1.

liftSymCompare2 :: forall (mode :: EvalModeTag) f a b c d. (DecideEvalMode mode, UnifiedSymOrd2 mode f) => (a -> b -> GetData mode Ordering) -> (c -> d -> GetData mode Ordering) -> f a c -> f b d -> GetData mode Ordering

Unified liftSymCompare2.

symCompare2 :: forall (mode :: EvalModeTag) f a b. (DecideEvalMode mode, UnifiedSymOrd mode a, UnifiedSymOrd mode b, UnifiedSymOrd2 mode f) => f a b -> f a b -> GetData mode Ordering

Unified symCompare2.