Grisette.Internal.Unified.UnifiedPrim

Description

Documentation

type UnifiedPrim (mode :: EvalModeTag) a = (Prim a, UnifiedITEOp mode a, UnifiedSymEq mode a, UnifiedSymOrd mode a)

A type that is used as a constraint for all the (unified) primitive types in Grisette.

type UnifiedBasicPrim (mode :: EvalModeTag) a = (UnifiedPrim mode a, UnifiedSimpleMergeable mode a, UnifiedConRep a, UnifiedSymRep a, UnifiedSolvable mode a (ConType a), ConSymConversion (ConType a) (SymType a) a)

A type that is used as a constraint for all the basic (unified) primitive types in Grisette.

GetSomeWordN is not considered as a basic (unified) primitive type.