Grisette.Internal.Unified.Util

Description

Documentation

class DecideEvalMode (mode :: EvalModeTag) where

A class that provides the mode tag at runtime.

Methods

decideEvalMode :: EvalModeTag

withMode :: forall (mode :: EvalModeTag) r. DecideEvalMode mode => (mode ~ 'C => r) -> (mode ~ 'S => r) -> r

Case analysis on the mode.

class (DecideEvalMode c, DecideEvalMode s) => EvalModeConvertible (c :: EvalModeTag) (s :: EvalModeTag) where

A class saying that we can convert a value with one mode to another mode.

Allowed conversions:

C - C

S - S

C - S

Conversion from left to right uses ToSym class, and conversion from right to left uses ToCon class.

Methods

withModeConvertible :: (c ~ 'C => r) -> (s ~ 'S => r) -> r

withModeConvertible' :: ((c ~ 'C, s ~ 'C) => r) -> ((c ~ 'C, s ~ 'S) => r) -> ((c ~ 'S, s ~ 'S) => r) -> r