Grisette.Internal.Utils.Parameterized

Description

Unsafe axiom

unsafeAxiom :: forall {k} (a :: k) (b :: k). a :~: b

Assert a proof of equality between two types. This is unsafe if used improperly, so use this with caution!

Unparameterized type

data SomeNatRepr where

Existential wrapper for NatRepr.

Constructors

SomeNatRepr :: forall (n :: Nat). KnownNat n => NatRepr n -> SomeNatRepr

data SomePositiveNatRepr where

Existential wrapper for NatRepr with the constraint that the natural number is greater than 0.

Constructors

SomePositiveNatRepr :: forall (n :: Nat). (KnownNat n, 1 <= n) => NatRepr n -> SomePositiveNatRepr

Runtime representation of type-level natural numbers

data NatRepr (n :: Nat)

A runtime representation of type-level natural numbers. This can be used for performing dynamic checks on type-level natural numbers.

withKnownNat :: forall (n :: Nat) r. NatRepr n -> (KnownNat n => r) -> r

Construct the KnownNat constraint when the runtime value is known.

natValue :: forall (n :: Nat). NatRepr n -> Natural

The underlying runtime natural number value of a type-level natural number.

mkNatRepr :: Natural -> SomeNatRepr

Turn a Natural into the corresponding NatRepr with the KnownNat constraint.

mkPositiveNatRepr :: Natural -> SomePositiveNatRepr

Turn a NatRepr into the corresponding NatRepr with the KnownNat constraint and asserts that its greater than 0.

natRepr :: forall (n :: Nat). KnownNat n => NatRepr n

Construct a runtime representation of a type-level natural number when its runtime value is known.

decNat :: forall (n :: Natural). 1 <= n => NatRepr n -> NatRepr (n - 1)

Decrement a NatRepr by 1.

predNat :: forall (n :: Natural). NatRepr (n + 1) -> NatRepr n

Predecessor of a NatRepr

incNat :: forall (n :: Nat). NatRepr n -> NatRepr (n + 1)

Increment a NatRepr by 1.

addNat :: forall (m :: Nat) (n :: Nat). NatRepr m -> NatRepr n -> NatRepr (m + n)

Addition of two NatReprs.

subNat :: forall (n :: Nat) (m :: Nat). n <= m => NatRepr m -> NatRepr n -> NatRepr (m - n)

Subtraction of two NatReprs.

divNat :: forall (n :: Natural) (m :: Nat). 1 <= n => NatRepr m -> NatRepr n -> NatRepr (Div m n)

Division of two NatReprs.

halfNat :: forall (n :: Natural). NatRepr (n + n) -> NatRepr n

Half of a NatRepr.

Proof of KnownNat

data KnownProof (n :: Nat) where

'KnownProof n' is a type whose values are only inhabited when n has a known runtime value.

Constructors

KnownProof :: forall (n :: Nat). KnownNat n => KnownProof n

hasRepr :: forall (n :: Nat). NatRepr n -> KnownProof n

Construct a KnownProof given the runtime representation.

withKnownProof :: forall (n :: Nat) r. KnownProof n -> (KnownNat n => r) -> r

Introduces the KnownNat constraint when it's proven.

unsafeKnownProof :: forall (n :: Nat). Natural -> KnownProof n

Construct a KnownProof given the runtime value.

Note: This function is unsafe, as it does not check that the runtime representation is consistent with the type-level representation. You should ensure the consistency yourself or the program can crash or generate incorrect results.

knownAdd :: forall (m :: Nat) (n :: Nat). KnownProof m -> KnownProof n -> KnownProof (m + n)

Adding two type-level natural numbers with known runtime values gives a type-level natural number with a known runtime value.

Proof of CmpNat

data CmpNatProof (m :: Nat) (n :: Nat) (o :: Ordering) where

Proof that the comparison of two type-level natural numbers is consistent with the runtime comparison.

Constructors

CmpNatProof :: forall (m :: Nat) (n :: Nat) (o :: Ordering). (CmpNat m n == o) ~ 'True => CmpNatProof m n o

unsafeCmpNatProof :: forall (m :: Nat) (n :: Nat) (o :: Ordering). CmpNatProof m n o

Construct a CmpNatProof.

withCmpNatProof :: forall (m :: Nat) (n :: Nat) (o :: Ordering) r. CmpNatProof m n o -> ((CmpNat m n == o) ~ 'True => r) -> r

Introduces the CmpNat m n o constraint when it's proven.

Proof of (<=) for type-level natural numbers

data LeqProof (m :: Nat) (n :: Nat) where

'LeqProof m n' is a type whose values are only inhabited when m <= n.

Constructors

LeqProof :: forall (m :: Nat) (n :: Nat). m <= n => LeqProof m n

withLeqProof :: forall (m :: Nat) (n :: Nat) r. LeqProof m n -> (m <= n => r) -> r

Introduces the m <= n constraint when it's proven.

unsafeLeqProof :: forall (m :: Nat) (n :: Nat). LeqProof m n

Construct a LeqProof.

Note: This function is unsafe, as it does not check that the left-hand side is less than or equal to the right-hand side. You should ensure the consistency yourself or the program can crash or generate incorrect results.

testLeq :: forall (m :: Nat) (n :: Nat). NatRepr m -> NatRepr n -> Maybe (LeqProof m n)

Checks if a NatRepr is less than or equal to another NatRepr.

leqRefl :: forall f (n :: Nat). f n -> LeqProof n n

Apply reflexivity to LeqProof.

leqSucc :: forall f (n :: Nat). f n -> LeqProof n (n + 1)

A natural number is less than or equal to its successor.

leqTrans :: forall (a :: Nat) (b :: Nat) (c :: Nat). LeqProof a b -> LeqProof b c -> LeqProof a c

Apply transitivity to LeqProof.

leqZero :: forall (n :: Nat). LeqProof 0 n

Zero is less than or equal to any natural number.

leqAdd2 :: forall (xl :: Nat) (xh :: Nat) (yl :: Nat) (yh :: Nat). LeqProof xl xh -> LeqProof yl yh -> LeqProof (xl + yl) (xh + yh)

Add both sides of two inequalities.

leqAdd :: forall (m :: Nat) (n :: Nat) f (o :: Natural). LeqProof m n -> f o -> LeqProof m (n + o)

Produce proof that adding a value to the larger element in an LeqProof is larger.

leqAddPos :: forall (m :: Natural) (n :: Natural) p q. (1 <= m, 1 <= n) => p m -> q n -> LeqProof 1 (m + n)

Adding two positive natural numbers is positive.