Grisette.Lib.Data.Maybe

Description

Documentation

mrgNothing :: (Mergeable (Maybe a), Applicative m, TryMerge m) => m (Maybe a)

Smart constructor for Nothing to construct values wrapped and possibly merged in a container.

mrgJust :: (Mergeable (Maybe a), Applicative m, TryMerge m) => a -> m (Maybe a)

Smart constructor for Just to construct values wrapped and possibly merged in a container.