Grisette.Lib.Data.Traversable

Description

The Traversable class

mrgTraverse :: forall a b t f. (Mergeable b, Mergeable1 t, TryMerge f, Applicative f, Traversable t) => (a -> f b) -> t a -> f (t b)

traverse with MergingStrategy knowledge propagation.

mrgSequenceA :: (Mergeable a, Mergeable1 t, Applicative f, TryMerge f, Traversable t) => t (f a) -> f (t a)

sequenceA with MergingStrategy knowledge propagation.

mrgMapM :: forall a b t f. (Mergeable b, Mergeable1 t, MonadTryMerge f, Traversable t) => (a -> f b) -> t a -> f (t b)

mapM with MergingStrategy knowledge propagation.

mrgSequence :: (Mergeable a, Mergeable1 t, MonadTryMerge f, Traversable t) => t (f a) -> f (t a)

sequence with MergingStrategy knowledge propagation.

Utility functions

mrgFor :: (Mergeable b, Mergeable1 t, Traversable t, TryMerge m, Applicative m) => t a -> (a -> m b) -> m (t b)

for with MergingStrategy knowledge propagation.

mrgForM :: (Mergeable b, Mergeable1 t, Traversable t, MonadTryMerge m) => t a -> (a -> m b) -> m (t b)

forM with MergingStrategy knowledge propagation.

mrgMapAccumM :: (MonadTryMerge m, Traversable t, Mergeable s, Mergeable b, Mergeable1 t) => (s -> a -> m (s, b)) -> s -> t a -> m (s, t b)

mapAccumM with MergingStrategy knowledge propagation.

mrgForAccumM :: (MonadTryMerge m, Traversable t, Mergeable s, Mergeable b, Mergeable1 t) => s -> t a -> (s -> a -> m (s, b)) -> m (s, t b)

forAccumM and MergingStrategy knowledge propagation.