Grisette.Lib.Data.Tuple

Description

Documentation

mrgUnit :: (Mergeable (), Applicative m, TryMerge m) => m ()

Smart constructor for v'GHC.Tuple.()' to construct values wrapped and possibly merged in a container.

mrgTuple2 :: (Mergeable (a, b), Applicative m, TryMerge m) => a -> b -> m (a, b)

Smart constructor for v'GHC.Tuple.(,)' to construct values wrapped and possibly merged in a container.

mrgTuple3 :: (Mergeable (a, b, c), Applicative m, TryMerge m) => a -> b -> c -> m (a, b, c)

Smart constructor for v'GHC.Tuple.(,,)' to construct values wrapped and possibly merged in a container.

mrgTuple4 :: (Mergeable (a, b, c, d), Applicative m, TryMerge m) => a -> b -> c -> d -> m (a, b, c, d)

Smart constructor for v'GHC.Tuple.(,,,)' to construct values wrapped and possibly merged in a container.

mrgTuple5 :: (Mergeable (a, b, c, d, e), Applicative m, TryMerge m) => a -> b -> c -> d -> e -> m (a, b, c, d, e)

Smart constructor for v'GHC.Tuple.(,,,,)' to construct values wrapped and possibly merged in a container.

mrgTuple6 :: (Mergeable (a, b, c, d, e, f), Applicative m, TryMerge m) => a -> b -> c -> d -> e -> f -> m (a, b, c, d, e, f)

Smart constructor for v'GHC.Tuple.(,,,,,)' to construct values wrapped and possibly merged in a container.

mrgTuple7 :: (Mergeable (a, b, c, d, e, f, g), Applicative m, TryMerge m) => a -> b -> c -> d -> e -> f -> g -> m (a, b, c, d, e, f, g)

Smart constructor for v'GHC.Tuple.(,,,,,,)' to construct values wrapped and possibly merged in a container.

mrgTuple8 :: (Mergeable (a, b, c, d, e, f, g, h), Applicative m, TryMerge m) => a -> b -> c -> d -> e -> f -> g -> h -> m (a, b, c, d, e, f, g, h)

Smart constructor for v'GHC.Tuple.(,,,,,,,)' to construct values wrapped and possibly merged in a container.