Grisette.Unified

Description

Evaluation mode

data EvalModeTag

Evaluation mode for unified types. C means concrete evaluation, S means symbolic evaluation.

Constructors

C

S

type family IsConMode (mode :: EvalModeTag) = (r :: Bool) | r -> mode where ...

Type family to check if a mode is C.

Equations

IsConMode 'C = 'True

IsConMode 'S = 'False

type family BaseMonad (mode :: EvalModeTag) = (r :: Type -> Type) | r -> mode where ...

A type family that specifies the base monad for the evaluation mode.

Resolves to Identity for C mode, and Union for S mode.

Equations

BaseMonad 'C = Identity

BaseMonad 'S = Union

Aggregated constraints

genEvalMode :: String -> [TheoryToUnify] -> DecsQ

This template haskell function generates an EvalMode constraint on demand.

For example, if in your system, you are only working on bit-vectors and booleans, but not floating points, integers, or real numbers, you can use this function to generate a constraint that only includes the necessary constraints:

genEvalMode "MyEvalMode" [UWordN, UIntN, UBool]
f :: MyEvalMode mode => GetBool mode -> GetWordN mode 8 -> GetWordN mode 8
f = ...

This may help with faster compilation times.

Another usage of this custom constraint is to working with uninterpreted functions. The uninterpreted functions aren't available even with EvalModeAll, and is only available with the constraint generated by this function. Note that you need to explicitly list all the uninterpreted function types you need in your system.

genEvalMode "MyEvalModeUF" [UFun [UWordN, UIntN], UFun [UBool, UBool, UWordN]]

This will give us a constraint that allows us to work with booleans and bit-vectors, and also the uninterpreted functions that

maps an unsigned bit-vector (any bitwidth) to an unsigned integer (any bitwidth), and

maps two booleans to an unsigned bit-vector (any bitwidth).

You can then use them in your code like this:

f :: MyEvalModeUF mode => GetFun mode (GetWordN mode 8) (GetIntN mode 8) -> GetIntN mode 8
f fun = f # 1

The function will also provide the constraint MonadMyEvalModeUF, which includes the constraints for the monad and the unified branching, similar to MonadEvalModeAll.

For compilers older than GHC 9.2.1, see the notes for EvalModeAll. This function will also generate constraints like MyEvalModeUFFunUWordNUIntN, which can be used to resolve the constraints for older compilers.

The naming conversion is the concatenation of the three parts:

The base name provided by the user (i.e., MyEvalModeUF),

Fun,

The concatenation of all the types in the uninterpreted function (i.e., UWordNUIntN).

The arguments to the type class is as follows:

The first argument is the mode,

The second to the end arguments are the natural number arguments for all the types. Here the second argument is the bitwidth of the unsigned bit-vector argument, and the third argument is the bitwidth of the signed bit-vector result.

data TheoryToUnify

This data type is used to represent the theories that is unified.

The UFun constructor is used to represent a specific uninterpreted function type. The type is uncurried.

Constructors

UBool

UIntN

UWordN

UInteger

UAlgReal

UFP

UFun [TheoryToUnify]

class (DecideEvalMode mode, UnifiedBool mode, UnifiedBasicPrim mode (GetBool mode), Monad (BaseMonad mode), TryMerge (BaseMonad mode), UnifiedBranching mode (BaseMonad mode), AllUnifiedData mode) => EvalModeBase (mode :: EvalModeTag)

Provide the constraint that the mode is a valid evaluation mode, and provides the support for GetBool and GetData.

For compilers prior to GHC 9.2.1, see the notes for EvalModeAll.

type EvalModeInteger = UnifiedInteger

Provide the support for GetInteger.

For compilers prior to GHC 9.2.1, see the notes for EvalModeAll.

class (AllUnifiedBV mode, AllUnifiedBVBVConversion mode) => EvalModeBV (mode :: EvalModeTag)

Provide the support for GetIntN, GetWordN, GetSomeIntN, and GetSomeWordN.

For compilers prior to GHC 9.2.1, see the notes for EvalModeAll.

class (AllUnifiedFP mode, AllUnifiedFPFPConversion mode, AllUnifiedBVFPConversion mode) => EvalModeFP (mode :: EvalModeTag)

Provide the support for GetFP and GetFPRoundingMode.

For compilers prior to GHC 9.2.1, see the notes for EvalModeAll.

type EvalModeAlgReal = UnifiedAlgReal

Provide the support for GetAlgReal.

For compilers prior to GHC 9.2.1, see the notes for EvalModeAll.

class (EvalModeBase mode, EvalModeInteger mode, EvalModeAlgReal mode, EvalModeBV mode, EvalModeFP mode) => EvalModeAll (mode :: EvalModeTag)

A constraint that specifies that the mode is valid, and provide all the corresponding constraints for the operaions for the types.

Note for users with GHC prior to 9.2.1: the GHC compiler isn't able to resolve the operations for sized bitvectors and data types. In this case, you may need to provide UnifiedBV, SafeUnifiedBV, SafeUnifiedSomeBV, and UnifiedData constraints manually.

For example, the following code is valid for GHC 9.2.1 and later:

fbv ::
  forall mode n.
  (EvalMode mode, KnownNat n, 1 <= n) =>
  GetIntN mode n ->
  GetIntN mode n ->
  GetIntN mode n
fbv l r =
  mrgIte @mode
    (l .== r)
    (l + r)
    (symIte @mode (l .< r) l r)

But with older GHCs, you need to write:

fbv ::
  forall mode n.
  (EvalMode mode, KnownNat n, 1 <= n, UnifiedBV mode n) =>
  GetIntN mode n ->
  GetIntN mode n ->
  GetIntN mode n
fbv l r =
  mrgIte @mode
    (l .== r)
    (l + r)
    (symIte @mode (l .< r) l r)

type MonadEvalModeAll (mode :: EvalModeTag) (m :: Type -> Type) = (EvalModeAll mode, Monad m, TryMerge m, UnifiedBranching mode m)

A constraint that specifies that the mode is valid, and provide all the corresponding constraints for the operations for the types.

This also provide the branching constraints for the monad, and the safe operations: for example, SafeUnifiedInteger provides safeDiv for the integer type with in ExceptT ArithException m.

For users with GHC prior to 9.2.1, see notes in EvalModeAll.

class DecideEvalMode (mode :: EvalModeTag) where

A class that provides the mode tag at runtime.

Methods

decideEvalMode :: EvalModeTag

withMode :: forall (mode :: EvalModeTag) r. DecideEvalMode mode => (mode ~ 'C => r) -> (mode ~ 'S => r) -> r

Case analysis on the mode.

class (DecideEvalMode c, DecideEvalMode s) => EvalModeConvertible (c :: EvalModeTag) (s :: EvalModeTag) where

A class saying that we can convert a value with one mode to another mode.

Allowed conversions:

C - C

S - S

C - S

Conversion from left to right uses ToSym class, and conversion from right to left uses ToCon class.

Methods

withModeConvertible :: (c ~ 'C => r) -> (s ~ 'S => r) -> r

withModeConvertible' :: ((c ~ 'C, s ~ 'C) => r) -> ((c ~ 'C, s ~ 'S) => r) -> ((c ~ 'S, s ~ 'S) => r) -> r

Unified operations

Unified simple mergeable

class (DecideEvalMode mode, TryMerge m) => UnifiedBranching (mode :: EvalModeTag) (m :: Type -> Type) where

A class that provides a unified branching operation.

We use this type class to help resolve the constraints for SymBranching.

Methods

withBaseBranching :: (If (IsConMode mode) (TryMerge m) (SymBranching m) => r) -> r

class Mergeable a => UnifiedSimpleMergeable (mode :: EvalModeTag) a where

A class that provides a unified simple merging.

We use this type class to help resolve the constraints for SimpleMergeable.

Methods

withBaseSimpleMergeable :: (If (IsConMode mode) () (SimpleMergeable a) => r) -> r

class UnifiedSimpleMergeable1 (mode :: EvalModeTag) (f :: Type -> Type) where

A class that provides lifting of unified simple merging.

We use this type class to help resolve the constraints for SimpleMergeable1.

Methods

withBaseSimpleMergeable1 :: (If (IsConMode mode) () (SimpleMergeable1 f) => r) -> r

class UnifiedSimpleMergeable2 (mode :: EvalModeTag) (f :: Type -> Type -> Type) where

A class that provides lifting of unified simple merging.

We use this type class to help resolve the constraints for SimpleMergeable2.

Methods

withBaseSimpleMergeable2 :: (If (IsConMode mode) () (SimpleMergeable2 f) => r) -> r

mrgIf :: forall (mode :: EvalModeTag) a m. (DecideEvalMode mode, Mergeable a, UnifiedBranching mode m) => GetBool mode -> m a -> m a -> m a

Unified mrgIf.

This function isn't able to infer the mode of the boolean variable, so you need to provide the mode explicitly. For example:

mrgIf @mode (a .== b) ...
mrgIf (a .== b :: SymBool) ...
mrgIf (a .== b :: GetBool mode) ...

liftBaseMonad :: forall (mode :: EvalModeTag) a m. (Applicative m, UnifiedBranching mode m, Mergeable a) => BaseMonad mode a -> m a

Unified lifting of a base monad.

mrgIte :: forall (mode :: EvalModeTag) a. (DecideEvalMode mode, UnifiedSimpleMergeable mode a) => GetBool mode -> a -> a -> a

Unified mrgIte.

mrgIte1 :: forall (mode :: EvalModeTag) f a. (DecideEvalMode mode, UnifiedSimpleMergeable1 mode f, UnifiedSimpleMergeable mode a) => GetBool mode -> f a -> f a -> f a

Unified mrgIte1.

liftMrgIte :: forall (mode :: EvalModeTag) f a. (DecideEvalMode mode, UnifiedSimpleMergeable1 mode f) => (GetBool mode -> a -> a -> a) -> GetBool mode -> f a -> f a -> f a

Unified liftMrgIte.

mrgIte2 :: forall (mode :: EvalModeTag) f a b. (DecideEvalMode mode, UnifiedSimpleMergeable2 mode f, UnifiedSimpleMergeable mode a, UnifiedSimpleMergeable mode b) => GetBool mode -> f a b -> f a b -> f a b

Unified mrgIte2.

liftMrgIte2 :: forall (mode :: EvalModeTag) f a b. (DecideEvalMode mode, UnifiedSimpleMergeable2 mode f) => (GetBool mode -> a -> a -> a) -> (GetBool mode -> b -> b -> b) -> GetBool mode -> f a b -> f a b -> f a b

Unified liftMrgIte2.

simpleMerge :: forall (mode :: EvalModeTag) a. (DecideEvalMode mode, UnifiedSimpleMergeable mode a) => BaseMonad mode a -> a

Unified merge of simply mergeable values in the base monad.

Unified ITE operator

class UnifiedITEOp (mode :: EvalModeTag) v where

A class that provides unified equality comparison.

We use this type class to help resolve the constraints for ITEOp.

Methods

withBaseITEOp :: (If (IsConMode mode) () (ITEOp v) => r) -> r

symIte :: forall (mode :: EvalModeTag) v. (DecideEvalMode mode, UnifiedITEOp mode v) => GetBool mode -> v -> v -> v

Unified symIte operation.

This function isn't able to infer the mode of the boolean variable, so you need to provide the mode explicitly. For example:

symIte @mode (a .== b) ...
symIte (a .== b :: SymBool) ...
symIte (a .== b :: GetBool mode) ...

symIteMerge :: forall (mode :: EvalModeTag) v. (DecideEvalMode mode, UnifiedITEOp mode v, Mergeable v) => BaseMonad mode v -> v

Unified symIteMerge operation.

This function isn't able to infer the mode of the base monad from the result, so you need to provide the mode explicitly. For example:

symIteMerge @mode ...
symIteMerge (... :: BaseMonad mode v) ...

Unified SymEq

class UnifiedSymEq (mode :: EvalModeTag) a where

A class that provides unified equality comparison.

We use this type class to help resolve the constraints for Eq and SymEq.

Methods

withBaseSymEq :: (If (IsConMode mode) (Eq a) (SymEq a) => r) -> r

class (forall a. UnifiedSymEq mode a => UnifiedSymEq mode (f a)) => UnifiedSymEq1 (mode :: EvalModeTag) (f :: Type -> Type) where

A class that provides unified lifting of equality comparison.

We use this type class to help resolve the constraints for Eq1 and SymEq1.

Methods

withBaseSymEq1 :: (If (IsConMode mode) (Eq1 f) (SymEq1 f) => r) -> r

class (forall a. UnifiedSymEq mode a => UnifiedSymEq1 mode (f a)) => UnifiedSymEq2 (mode :: EvalModeTag) (f :: Type -> Type -> Type) where

A class that provides unified lifting of equality comparison.

We use this type class to help resolve the constraints for Eq2 and SymEq2.

Methods

withBaseSymEq2 :: (If (IsConMode mode) (Eq2 f) (SymEq2 f) => r) -> r

(.==) :: forall (mode :: EvalModeTag) a. (DecideEvalMode mode, UnifiedSymEq mode a) => a -> a -> GetBool mode

Unified (.==).

Note that you may sometimes need to write type annotations for the result when the mode isn't clear:

a .== b :: GetBool mode

One example when it isn't clear is when this is used in unified mrgIf.

(./=) :: forall (mode :: EvalModeTag) a. (DecideEvalMode mode, UnifiedSymEq mode a) => a -> a -> GetBool mode

Unified (./=).

Note that you may sometimes need to write type annotations for the result when the mode isn't clear:

a ./= b :: GetBool mode

One example when it isn't clear is when this is used in unified mrgIf.

symDistinct :: forall (mode :: EvalModeTag) a. (DecideEvalMode mode, UnifiedSymEq mode a) => [a] -> GetBool mode

Unified symDistinct.

Note that you may sometimes need to write type annotations for the result when the mode isn't clear:

symDistinct l :: GetBool mode

One example when it isn't clear is when this is used in unified mrgIf.

liftSymEq :: forall (mode :: EvalModeTag) f a b. (DecideEvalMode mode, UnifiedSymEq1 mode f) => (a -> b -> GetBool mode) -> f a -> f b -> GetBool mode

Unified liftSymEq.

symEq1 :: forall (mode :: EvalModeTag) f a. (DecideEvalMode mode, UnifiedSymEq mode a, UnifiedSymEq1 mode f) => f a -> f a -> GetBool mode

Unified symEq1.

liftSymEq2 :: forall (mode :: EvalModeTag) f a b c d. (DecideEvalMode mode, UnifiedSymEq2 mode f) => (a -> b -> GetBool mode) -> (c -> d -> GetBool mode) -> f a c -> f b d -> GetBool mode

Unified liftSymEq2.

symEq2 :: forall (mode :: EvalModeTag) f a b. (DecideEvalMode mode, UnifiedSymEq mode a, UnifiedSymEq mode b, UnifiedSymEq2 mode f) => f a b -> f a b -> GetBool mode

Unified symEq2.

Unified SymOrd

class UnifiedSymOrd (mode :: EvalModeTag) a where

A class that provides unified comparison.

We use this type class to help resolve the constraints for Ord and SymOrd.

Methods

withBaseSymOrd :: (If (IsConMode mode) (Ord a) (SymOrd a) => r) -> r

class UnifiedSymOrd1 (mode :: EvalModeTag) (f :: Type -> Type) where

A class that provides unified lifting of comparison.

We use this type class to help resolve the constraints for Ord1 and SymOrd1.

Methods

withBaseSymOrd1 :: (If (IsConMode mode) (Ord1 f) (SymOrd1 f) => r) -> r

class UnifiedSymOrd2 (mode :: EvalModeTag) (f :: Type -> Type -> Type) where

A class that provides unified lifting of comparison.

We use this type class to help resolve the constraints for Ord2 and SymOrd2.

Methods

withBaseSymOrd2 :: (If (IsConMode mode) (Ord2 f) (SymOrd2 f) => r) -> r

(.<=) :: forall (mode :: EvalModeTag) a. (DecideEvalMode mode, UnifiedSymOrd mode a) => a -> a -> GetBool mode

Unified (.<=).

Note that you may sometimes need to write type annotations for the result when the mode isn't clear:

a .<= b :: GetBool mode

One example when it isn't clear is when this is used in unified mrgIf.

(.<) :: forall (mode :: EvalModeTag) a. (DecideEvalMode mode, UnifiedSymOrd mode a) => a -> a -> GetBool mode

Unified (.<).

(.>=) :: forall (mode :: EvalModeTag) a. (DecideEvalMode mode, UnifiedSymOrd mode a) => a -> a -> GetBool mode

Unified (.>=).

(.>) :: forall (mode :: EvalModeTag) a. (DecideEvalMode mode, UnifiedSymOrd mode a) => a -> a -> GetBool mode

Unified (.>).

symCompare :: forall (mode :: EvalModeTag) a (ctx :: Type -> Type). (DecideEvalMode mode, UnifiedSymOrd mode a, Monad ctx) => a -> a -> BaseMonad mode Ordering

Unified symCompare.

liftSymCompare :: forall (mode :: EvalModeTag) f a b. (DecideEvalMode mode, UnifiedSymOrd1 mode f) => (a -> b -> BaseMonad mode Ordering) -> f a -> f b -> BaseMonad mode Ordering

Unified liftSymCompare.

symCompare1 :: forall (mode :: EvalModeTag) f a. (DecideEvalMode mode, UnifiedSymOrd mode a, UnifiedSymOrd1 mode f) => f a -> f a -> BaseMonad mode Ordering

Unified symCompare1.

liftSymCompare2 :: forall (mode :: EvalModeTag) f a b c d. (DecideEvalMode mode, UnifiedSymOrd2 mode f) => (a -> b -> BaseMonad mode Ordering) -> (c -> d -> BaseMonad mode Ordering) -> f a c -> f b d -> BaseMonad mode Ordering

Unified liftSymCompare2.

symCompare2 :: forall (mode :: EvalModeTag) f a b. (DecideEvalMode mode, UnifiedSymOrd mode a, UnifiedSymOrd mode b, UnifiedSymOrd2 mode f) => f a b -> f a b -> BaseMonad mode Ordering

Unified symCompare2.

symMax :: forall (mode :: EvalModeTag) a. (UnifiedSymOrd mode a, UnifiedITEOp mode a, DecideEvalMode mode) => a -> a -> a

Unified symMax.

symMin :: forall (mode :: EvalModeTag) a. (UnifiedSymOrd mode a, UnifiedITEOp mode a, DecideEvalMode mode) => a -> a -> a

Unified symMin.

mrgMax :: forall (mode :: EvalModeTag) a m. (UnifiedSymOrd mode a, UnifiedBranching mode m, DecideEvalMode mode, Applicative m, Mergeable a) => a -> a -> m a

Unified mrgMax.

mrgMin :: forall (mode :: EvalModeTag) a m. (UnifiedSymOrd mode a, UnifiedBranching mode m, DecideEvalMode mode, Applicative m, Mergeable a) => a -> a -> m a

Unified mrgMin.

Unified finite bits

class UnifiedFiniteBits (mode :: EvalModeTag) a where

A class that provides unified equality comparison.

We use this type class to help resolve the constraints for FiniteBits, FromBits and SymFiniteBits.

Methods

withBaseFiniteBits :: (If (IsConMode mode) (FiniteBits a, FromBits a) (SymFiniteBits a) => r) -> r

symTestBit :: forall (mode :: EvalModeTag) a. (DecideEvalMode mode, UnifiedFiniteBits mode a) => a -> Int -> GetBool mode

Unified symTestBit.

symSetBitTo :: forall (mode :: EvalModeTag) a. (DecideEvalMode mode, UnifiedFiniteBits mode a) => a -> Int -> GetBool mode -> a

Unified symSetBitTo.

symFromBits :: forall (mode :: EvalModeTag) a. (DecideEvalMode mode, UnifiedFiniteBits mode a) => [GetBool mode] -> a

Unified symFromBits.

symBitBlast :: forall (mode :: EvalModeTag) a. (DecideEvalMode mode, UnifiedFiniteBits mode a) => a -> [GetBool mode]

Unified symBitBlast.

symLsb :: forall (mode :: EvalModeTag) a. (DecideEvalMode mode, UnifiedFiniteBits mode a) => a -> GetBool mode

Unified symLsb.

symMsb :: forall (mode :: EvalModeTag) a. (DecideEvalMode mode, UnifiedFiniteBits mode a) => a -> GetBool mode

Unified symMsb.

symPopCount :: forall (mode :: EvalModeTag) a. (DecideEvalMode mode, UnifiedFiniteBits mode a, Num a, UnifiedITEOp mode a) => a -> a

Unified symPopCount.

symCountLeadingZeros :: forall (mode :: EvalModeTag) a. (DecideEvalMode mode, UnifiedFiniteBits mode a, Num a, UnifiedITEOp mode a) => a -> a

Unified symCountLeadingZeros.

symCountTrailingZeros :: forall (mode :: EvalModeTag) a. (DecideEvalMode mode, UnifiedFiniteBits mode a, Num a, UnifiedITEOp mode a) => a -> a

Unified symCountTrailingZeros.

Unified conversions

class UnifiedFromIntegral (mode :: EvalModeTag) a b where

A class that provides unified conversion from integral types.

We use this type class to help resolve the constraints for SymFromIntegral.

Methods

withBaseFromIntegral :: (If (IsConMode mode) (Integral a, Num b) (SymFromIntegral a b) => r) -> r

symFromIntegral :: forall (mode :: EvalModeTag) a b. (DecideEvalMode mode, UnifiedFromIntegral mode a b) => a -> b

Unified symFromIntegral operation.

This function isn't able to infer the mode, so you need to provide the mode explicitly. For example:

symFromIntegral @mode a

class UnifiedSafeBitCast (mode :: EvalModeTag) e a b (m :: Type -> Type) where

A class that provides unified safe bitcast operations.

We use this type class to help resolve the constraints for SafeBitCast.

Methods

withBaseSafeBitCast :: (SafeBitCast e a b m => r) -> r

safeBitCast :: forall (mode :: EvalModeTag) e a b m. (MonadError e m, UnifiedSafeBitCast mode e a b m) => a -> m b

Unified safeSub operation.

This function isn't able to infer the mode, so you need to provide the mode explicitly. For example:

safeSub @mode a b

class UnifiedSafeFromFP (mode :: EvalModeTag) e a fp fprd (m :: Type -> Type) where

A class that provides unified safe conversion from floating points.

We use this type class to help resolve the constraints for SafeFromFP.

Methods

withBaseSafeFromFP :: (SafeFromFP e a fp fprd m => r) -> r

safeFromFP :: forall (mode :: EvalModeTag) e a fp fprd m. (UnifiedSafeFromFP mode e a fp fprd m, MonadError e m) => fprd -> fp -> m a

Unified safeFromFP operation.

This function isn't able to infer the mode, so you need to provide the mode explicitly. For example:

safeFromFP @mode mode fp

Unified safe ops

class UnifiedSafeDiv (mode :: EvalModeTag) e a (m :: Type -> Type) where

A class that provides unified division operations.

We use this type class to help resolve the constraints for SafeDiv.

Methods

withBaseSafeDiv :: (SafeDiv e a m => r) -> r

safeDiv :: forall (mode :: EvalModeTag) e a m. (MonadError e m, UnifiedSafeDiv mode e a m) => a -> a -> m a

Unified safeDiv operation.

This function isn't able to infer the mode, so you need to provide the mode explicitly. For example:

safeDiv @mode a b

safeMod :: forall (mode :: EvalModeTag) e a m. (MonadError e m, UnifiedSafeDiv mode e a m) => a -> a -> m a

Unified safeMod operation.

This function isn't able to infer the mode, so you need to provide the mode explicitly. For example:

safeMod @mode a b

safeDivMod :: forall (mode :: EvalModeTag) e a m. (MonadError e m, UnifiedSafeDiv mode e a m) => a -> a -> m (a, a)

Unified safeDivMod operation.

This function isn't able to infer the mode, so you need to provide the mode explicitly. For example:

safeDivMod @mode a b

safeQuot :: forall (mode :: EvalModeTag) e a m. (MonadError e m, UnifiedSafeDiv mode e a m) => a -> a -> m a

Unified safeQuot operation.

This function isn't able to infer the mode, so you need to provide the mode explicitly. For example:

safeQuot @mode a b

safeRem :: forall (mode :: EvalModeTag) e a m. (MonadError e m, UnifiedSafeDiv mode e a m) => a -> a -> m a

Unified safeRem operation.

This function isn't able to infer the mode, so you need to provide the mode explicitly. For example:

safeRem @mode a b

safeQuotRem :: forall (mode :: EvalModeTag) e a m. (MonadError e m, UnifiedSafeDiv mode e a m) => a -> a -> m (a, a)

Unified safeQuotRem operation.

This function isn't able to infer the mode, so you need to provide the mode explicitly. For example:

safeQuotRem @mode a b

class UnifiedSafeLinearArith (mode :: EvalModeTag) e a (m :: Type -> Type) where

A class that provides unified linear arithmetic operations.

We use this type class to help resolve the constraints for SafeLinearArith.

Methods

withBaseSafeLinearArith :: (SafeLinearArith e a m => r) -> r

safeAdd :: forall (mode :: EvalModeTag) e a m. (MonadError e m, UnifiedSafeLinearArith mode e a m) => a -> a -> m a

Unified safeAdd operation.

This function isn't able to infer the mode, so you need to provide the mode explicitly. For example:

safeAdd @mode a b

safeNeg :: forall (mode :: EvalModeTag) e a m. (MonadError e m, UnifiedSafeLinearArith mode e a m) => a -> m a

Unified safeNeg operation.

This function isn't able to infer the mode, so you need to provide the mode explicitly. For example:

safeNeg @mode a

safeSub :: forall (mode :: EvalModeTag) e a m. (MonadError e m, UnifiedSafeLinearArith mode e a m) => a -> a -> m a

Unified safeSub operation.

This function isn't able to infer the mode, so you need to provide the mode explicitly. For example:

safeSub @mode a b

class UnifiedSafeSymRotate (mode :: EvalModeTag) e a (m :: Type -> Type) where

A class that provides unified safe symbolic rotation operations.

We use this type class to help resolve the constraints for SafeSymRotate.

Methods

withBaseSafeSymRotate :: (SafeSymRotate e a m => r) -> r

safeSymRotateL :: forall (mode :: EvalModeTag) e a m. (MonadError e m, UnifiedSafeSymRotate mode e a m) => a -> a -> m a

Unified safeSymRotateL operation.

This function isn't able to infer the mode, so you need to provide the mode explicitly. For example:

safeSymRotateL @mode a b

safeSymRotateR :: forall (mode :: EvalModeTag) e a m. (MonadError e m, UnifiedSafeSymRotate mode e a m) => a -> a -> m a

Unified safeSymRotateR operation.

This function isn't able to infer the mode, so you need to provide the mode explicitly. For example:

safeSymRotateR @mode a b

class UnifiedSafeSymShift (mode :: EvalModeTag) e a (m :: Type -> Type) where

A class that provides unified safe symbolic rotation operations.

We use this type class to help resolve the constraints for SafeSymShift.

Methods

withBaseSafeSymShift :: (SafeSymShift e a m => r) -> r

safeSymShiftL :: forall (mode :: EvalModeTag) e a m. (MonadError e m, UnifiedSafeSymShift mode e a m) => a -> a -> m a

Unified safeSymShiftL operation.

This function isn't able to infer the mode, so you need to provide the mode explicitly. For example:

safeSymShiftL @mode a b

safeSymShiftR :: forall (mode :: EvalModeTag) e a m. (MonadError e m, UnifiedSafeSymShift mode e a m) => a -> a -> m a

Unified safeSymShiftR operation.

This function isn't able to infer the mode, so you need to provide the mode explicitly. For example:

safeSymShiftR @mode a b

safeSymStrictShiftL :: forall (mode :: EvalModeTag) e a m. (MonadError e m, UnifiedSafeSymShift mode e a m) => a -> a -> m a

Unified safeSymStrictShiftL operation.

This function isn't able to infer the mode, so you need to provide the mode explicitly. For example:

safeSymStrictShiftL @mode a b

safeSymStrictShiftR :: forall (mode :: EvalModeTag) e a m. (MonadError e m, UnifiedSafeSymShift mode e a m) => a -> a -> m a

Unified safeSymStrictShiftR operation.

This function isn't able to infer the mode, so you need to provide the mode explicitly. For example:

safeSymStrictShiftR @mode a b

class UnifiedSafeFdiv (mode :: EvalModeTag) e a (m :: Type -> Type) where

A class that provides unified floating division operations.

We use this type class to help resolve the constraints for SafeFdiv.

Methods

withBaseUnifiedSafeFdiv :: (SafeFdiv e a m => r) -> r

safeFdiv :: forall (mode :: EvalModeTag) e a m. (MonadError e m, UnifiedSafeFdiv mode e a m) => a -> a -> m a

Unified safeFdiv operation.

This function isn't able to infer the mode, so you need to provide the mode explicitly. For example:

safeFdiv @mode a b

Shared constraints

type UnifiedPrim (mode :: EvalModeTag) a = (Prim a, UnifiedITEOp mode a, UnifiedSymEq mode a, UnifiedSymOrd mode a)

A type that is used as a constraint for all the (unified) primitive types in Grisette.

type UnifiedBasicPrim (mode :: EvalModeTag) a = (UnifiedPrim mode a, UnifiedSimpleMergeable mode a, UnifiedConRep a, UnifiedSymRep a, UnifiedSolvable mode a (ConType a), ConSymConversion (ConType a) (SymType a) a)

A type that is used as a constraint for all the basic (unified) primitive types in Grisette.

GetSomeWordN is not considered as a basic (unified) primitive type.

Unified types

Boolean

type family GetBool (mode :: EvalModeTag) = (bool :: Type) | bool -> mode

Get a unified Boolean type. Resolves to Bool in C mode, and SymBool in S mode.

Bit-vector

type family GetIntN (mode :: EvalModeTag) = (i :: Nat -> Type) | i -> mode

Get a unified signed size-tagged bit vector type. Resolves to IntN in C mode, and SymIntN in S mode.

type family GetWordN (mode :: EvalModeTag) = (w :: Nat -> Type) | w -> mode

Get a unified unsigned size-tagged bit vector type. Resolves to WordN in C mode, and SymWordN in S mode.

type family GetSomeWordN (mode :: EvalModeTag) = (sw :: Type) | sw -> mode where ...

Get a unified unsigned dynamic bit width bit vector type. Resolves to SomeWordN in C mode, and SomeSymWordN in S mode.

Equations

GetSomeWordN mode = SomeBV (GetWordN mode)

type family GetSomeIntN (mode :: EvalModeTag) = (sw :: Type) | sw -> mode where ...

Get a unified signed dynamic bit width bit vector type. Resolves to SomeIntN in C mode, and SomeSymIntN in S mode.

Equations

GetSomeIntN mode = SomeBV (GetIntN mode)

class UnifiedBVImpl mode (GetWordN mode) (GetIntN mode) n (GetWordN mode n) (GetIntN mode n) => UnifiedBV (mode :: EvalModeTag) (n :: Nat)

This class is needed as constraint in user code prior to GHC 9.2.1. See the notes in EvalMode.

class SafeUnifiedBVImpl mode (GetWordN mode) (GetIntN mode) n (GetWordN mode n) (GetIntN mode n) m => SafeUnifiedBV (mode :: EvalModeTag) (n :: Nat) (m :: Type -> Type)

This class is needed as constraint in user code prior to GHC 9.2.1. See the notes in EvalMode.

class SafeUnifiedSomeBVImpl mode (SomeBV (GetWordN mode)) (SomeBV (GetIntN mode)) m => SafeUnifiedSomeBV (mode :: EvalModeTag) (m :: Type -> Type)

This class is needed as constraint in user code prior to GHC 9.2.1. See the notes in EvalMode.

Integer

type family GetInteger (mode :: EvalModeTag) = (int :: Type) | int -> mode

Get a unified Integer type. Resolves to Integer in C mode, and SymInteger in S mode.

class UnifiedIntegerImpl mode (GetInteger mode) => UnifiedInteger (mode :: EvalModeTag)

Evaluation mode with unified Integer type.

FP

type family GetFP (mode :: EvalModeTag) = (f :: Nat -> Nat -> Type) | f -> mode

Get a unified floating point type. Resolves to FP in C mode, and SymFP in S mode.

type family GetFPRoundingMode (mode :: EvalModeTag) = (r :: Type) | r -> mode

Get a unified floating point rounding mode type. Resolves to FPRoundingMode in C mode, and SymFPRoundingMode in S mode.

class UnifiedFPImpl mode (GetFP mode) eb sb (GetFP mode eb sb) (GetFPRoundingMode mode) => UnifiedFP (mode :: EvalModeTag) (eb :: Nat) (sb :: Nat)

Evaluation mode with unified FP type.

class (SafeUnifiedFPImpl mode (GetFP mode) eb sb (GetFP mode eb sb) (GetFPRoundingMode mode) m, UnifiedSafeFromFP mode NotRepresentableFPError (GetInteger mode) (GetFP mode eb sb) (GetFPRoundingMode mode) m) => SafeUnifiedFP (mode :: EvalModeTag) (eb :: Nat) (sb :: Nat) (m :: Type -> Type)

This class is needed as constraint in user code prior to GHC 9.2.1. See the notes in EvalMode.

AlgReal

type family GetAlgReal (mode :: EvalModeTag) = (real :: Type) | real -> mode

Get a unified algebraic real type. Resolves to AlgReal in C mode, and SymAlgReal in S mode.

Floating, LogBaseOr and SafeLogBase for SymAlgReal are not provided as they are not available for AlgReal.

class UnifiedAlgRealImpl mode (GetAlgReal mode) => UnifiedAlgReal (mode :: EvalModeTag)

Evaluation mode with unified AlgReal type.

Data

type family GetData (mode :: EvalModeTag) v = (r :: Type) | r -> mode v

Get a unified data type. Resolves to v in C mode, and Union v in S mode.

class UnifiedDataImpl mode v (GetData mode v) => UnifiedData (mode :: EvalModeTag) v

This class is needed as constraint in user code prior to GHC 9.2.1. See the notes in IsMode.

extractData :: (UnifiedDataImpl mode v u, Mergeable v, Monad m, UnifiedBranching mode m) => u -> m v

Extracts a value from the unified data type.

wrapData :: UnifiedDataImpl mode v u => v -> u

Wraps a value into the unified data type.

Functions

type family GetFun (mode :: EvalModeTag) = (fun :: Type -> Type -> Type) | fun -> mode

Get a unified function type. Resolves to =-> in C mode, and =~> in S mode.

type GetFun2 (mode :: EvalModeTag) a b = GetFun mode a b

The unified function type with 2 arguments.

type GetFun3 (mode :: EvalModeTag) a b c = GetFun mode a (GetFun mode b c)

The unified function type with 3 arguments.

type GetFun4 (mode :: EvalModeTag) a b c d = GetFun mode a (GetFun mode b (GetFun mode c d))

The unified function type with 4 arguments.

type GetFun5 (mode :: EvalModeTag) a b c d e = GetFun mode a (GetFun mode b (GetFun mode c (GetFun mode d e)))

The unified function type with 5 arguments.

type GetFun6 (mode :: EvalModeTag) a b c d e f = GetFun mode a (GetFun mode b (GetFun mode c (GetFun mode d (GetFun mode e f))))

The unified function type with 6 arguments.

type GetFun7 (mode :: EvalModeTag) a b c d e f g = GetFun mode a (GetFun mode b (GetFun mode c (GetFun mode d (GetFun mode e (GetFun mode f g)))))

The unified function type with 7 arguments.

type GetFun8 (mode :: EvalModeTag) a b c d e f g h = GetFun mode a (GetFun mode b (GetFun mode c (GetFun mode d (GetFun mode e (GetFun mode f (GetFun mode g h))))))

The unified function type with 8 arguments.

class UnifiedFun (mode :: EvalModeTag)

Provide unified function types.

type UnifiedFunConstraint (mode :: EvalModeTag) a b ca cb sa sb = (Show (GetFun mode a b), Binary (GetFun mode a b), Serial (GetFun mode a b), Serialize (GetFun mode a b), NFData (GetFun mode a b), Eq (GetFun mode a b), EvalSym (GetFun mode a b), ExtractSym (GetFun mode a b), Mergeable (GetFun mode a b), PPrint (GetFun mode a b), SubstSym (GetFun mode a b), Hashable (GetFun mode a b), Lift (GetFun mode a b), Typeable (GetFun mode a b), ToCon (GetFun mode a b) (ca =-> cb), ToCon (sa =~> sb) (GetFun mode a b), ToSym (GetFun mode a b) (sa =~> sb), ToSym (ca =-> cb) (GetFun mode a b), Function (GetFun mode a b) a b, Apply (GetFun mode a b), FunType (GetFun mode a b) ~ (a -> b))

The constraint for a unified function.

unifiedFunInstanceName :: String -> [TheoryToUnify] -> String

Generate unified function instance names.

genUnifiedFunInstance :: String -> [TheoryToUnify] -> DecsQ

Generate unified function instances.

Supplemental conversions

class (UnifiedBVBVConversionImpl mode (GetWordN mode) (GetWordN mode) n0 n1 (GetWordN mode n0) (GetWordN mode n1), UnifiedBVBVConversionImpl mode (GetWordN mode) (GetIntN mode) n0 n1 (GetWordN mode n0) (GetIntN mode n1), UnifiedBVBVConversionImpl mode (GetIntN mode) (GetWordN mode) n0 n1 (GetIntN mode n0) (GetWordN mode n1), UnifiedBVBVConversionImpl mode (GetIntN mode) (GetIntN mode) n0 n1 (GetIntN mode n0) (GetIntN mode n1)) => UnifiedBVBVConversion (mode :: EvalModeTag) (n0 :: Nat) (n1 :: Nat)

Unified constraints for conversion between bit-vectors.

class UnifiedBVFPConversionImpl mode (GetWordN mode) (GetIntN mode) (GetFP mode) n eb sb (GetWordN mode n) (GetIntN mode n) (GetFP mode eb sb) (GetFPRoundingMode mode) => UnifiedBVFPConversion (mode :: EvalModeTag) (n :: Nat) (eb :: Nat) (sb :: Nat)

Unified constraints for conversion from bit-vectors to floating point numbers.

class SafeUnifiedBVFPConversionImpl mode (GetWordN mode) (GetIntN mode) (GetFP mode) n eb sb (GetWordN mode n) (GetIntN mode n) (GetFP mode eb sb) (GetFPRoundingMode mode) m => SafeUnifiedBVFPConversion (mode :: EvalModeTag) (n :: Nat) (eb :: Nat) (sb :: Nat) (m :: Type -> Type)

Unified constraints for safe conversion from bit-vectors to floating point numbers.

class UnifiedFPFPConversionImpl mode (GetFP mode) eb0 sb0 eb1 sb1 (GetFP mode eb0 sb0) (GetFP mode eb1 sb1) (GetFPRoundingMode mode) => UnifiedFPFPConversion (mode :: EvalModeTag) (eb0 :: Nat) (sb0 :: Nat) (eb1 :: Nat) (sb1 :: Nat)

Unified constraints for conversion from floating point numbers to floating point numbers.