Grisette.Unified.Lib.Control.Monad

Description

Functor and Monad classes

mrgFmap :: (TryMerge f, Mergeable a, Mergeable b, Functor f) => (a -> b) -> f a -> f b

fmap with MergingStrategy knowledge propagation.

(.<$) :: (TryMerge f, Mergeable a, Mergeable b, Functor f) => b -> f a -> f b infixl 4

<$ with MergingStrategy knowledge propagation.

mrgReturnWithStrategy :: MonadTryMerge u => MergingStrategy a -> a -> u a

return with MergingStrategy knowledge propagation.

mrgBindWithStrategy :: MonadTryMerge u => MergingStrategy a -> MergingStrategy b -> u a -> (a -> u b) -> u b

>>= with MergingStrategy knowledge propagation.

mrgReturn :: (MonadTryMerge u, Mergeable a) => a -> u a

return with MergingStrategy knowledge propagation.

(.>>=) :: (MonadTryMerge u, Mergeable a, Mergeable b) => u a -> (a -> u b) -> u b infixl 1

>>= with MergingStrategy knowledge propagation.

(.>>) :: (MonadTryMerge m, Mergeable a, Mergeable b) => m a -> m b -> m b infixl 1

>> with MergingStrategy knowledge propagation.

This is usually more efficient than calling the original >> and merge the results.

mrgFail :: (MonadTryMerge m, Mergeable a, MonadFail m) => String -> m a

fail with MergingStrategy knowledge propagation.

mrgMzero :: (MonadTryMerge m, Mergeable a, MonadPlus m) => m a

mzero with MergingStrategy knowledge propagation.

mrgMplus :: (MonadTryMerge m, Mergeable a, MonadPlus m) => m a -> m a -> m a

mplus with MergingStrategy knowledge propagation.

Functions

Basic Monad functions

mrgMapM :: forall a b t f. (Mergeable b, Mergeable1 t, MonadTryMerge f, Traversable t) => (a -> f b) -> t a -> f (t b)

mapM with MergingStrategy knowledge propagation.

mrgMapM_ :: (MonadTryMerge m, Foldable t) => (a -> m b) -> t a -> m ()

mapM_ with MergingStrategy knowledge propagation.

mrgForM :: (Mergeable b, Mergeable1 t, Traversable t, MonadTryMerge m) => t a -> (a -> m b) -> m (t b)

forM with MergingStrategy knowledge propagation.

mrgForM_ :: (MonadTryMerge m, Foldable t) => t a -> (a -> m b) -> m ()

forM_ with MergingStrategy knowledge propagation.

mrgSequence :: (Mergeable a, Mergeable1 t, MonadTryMerge f, Traversable t) => t (f a) -> f (t a)

sequence with MergingStrategy knowledge propagation.

mrgSequence_ :: (Foldable t, MonadTryMerge m) => t (m a) -> m ()

sequence_ with MergingStrategy knowledge propagation.

(.=<<) :: (MonadTryMerge m, Mergeable a, Mergeable b) => (a -> m b) -> m a -> m b infixr 1

=<< with MergingStrategy knowledge propagation.

(.>=>) :: (MonadTryMerge m, Mergeable a, Mergeable b, Mergeable c) => (a -> m b) -> (b -> m c) -> a -> m c infixr 1

>=> with MergingStrategy knowledge propagation.

(.<=<) :: (MonadTryMerge m, Mergeable a, Mergeable b, Mergeable c) => (b -> m c) -> (a -> m b) -> a -> m c infixr 1

<=< with MergingStrategy knowledge propagation.

mrgForever :: (Applicative m, TryMerge m, Mergeable b, Mergeable a) => m a -> m b

forever with MergingStrategy knowledge propagation.

mrgVoid :: (TryMerge f, Functor f) => f a -> f ()

void with MergingStrategy knowledge propagation.

Generalisations of list functions

mrgJoin :: (MonadTryMerge m, Mergeable a) => m (m a) -> m a

join with MergingStrategy knowledge propagation.

mrgMsum :: (MonadTryMerge m, Mergeable a, MonadPlus m, Foldable t) => t (m a) -> m a

msum with MergingStrategy knowledge propagation.

mrgMfilter :: (MonadTryMerge m, MonadPlus m, Mergeable a) => (a -> Bool) -> m a -> m a

mfilter with MergingStrategy knowledge propagation.

symMfilter :: forall (mode :: EvalModeTag) m a. (MonadTryMerge m, MonadPlus m, UnifiedBranching mode m, Mergeable a) => (a -> GetBool mode) -> m a -> m a

mfilter with MergingStrategy knowledge propagation and symbolic conditions.

mrgFilterM :: (TryMerge m, Applicative m, Mergeable a, Foldable t) => (a -> m Bool) -> t a -> m [a]

filterM with MergingStrategy knowledge propagation.

symFilterM :: forall (mode :: EvalModeTag) m t a. (TryMerge m, UnifiedBranching mode m, MonadTryMerge m, EvalModeBase mode, Mergeable a, Foldable t) => (a -> m (GetBool mode)) -> t a -> m [a]

filterM with MergingStrategy knowledge propagation and symbolic conditions.

mrgMapAndUnzipM :: (Applicative m, TryMerge m, Mergeable b, Mergeable c) => (a -> m (b, c)) -> [a] -> m ([b], [c])

mapAndUnzipM with MergingStrategy knowledge propagation.

mrgZipWithM :: (Applicative m, TryMerge m, Mergeable c) => (a -> b -> m c) -> [a] -> [b] -> m [c]

zipWithM with MergingStrategy knowledge propagation.

mrgZipWithM_ :: (Applicative m, TryMerge m, Mergeable c) => (a -> b -> m c) -> [a] -> [b] -> m ()

zipWithM_ with MergingStrategy knowledge propagation.

mrgFoldM :: (MonadTryMerge m, Mergeable b, Foldable t) => (b -> a -> m b) -> b -> t a -> m b

foldM with MergingStrategy knowledge propagation.

mrgFoldM_ :: (MonadTryMerge m, Foldable t, Mergeable b) => (b -> a -> m b) -> b -> t a -> m ()

foldM_ with MergingStrategy knowledge propagation.

mrgReplicateM :: (Applicative m, TryMerge m, Mergeable a) => Int -> m a -> m [a]

replicateM with MergingStrategy knowledge propagation.

symReplicateM :: forall (mode :: EvalModeTag) m a int. (EvalModeBase mode, TryMerge m, Applicative m, Mergeable a, Num int, UnifiedBranching mode m, UnifiedSymOrd mode Int, UnifiedSymOrd mode int) => Int -> int -> m a -> m [a]

replicateM with MergingStrategy knowledge propagation and symbolic number of elements.

mrgReplicateM_ :: (Applicative m, TryMerge m, Mergeable a) => Int -> m a -> m ()

replicateM_ with MergingStrategy knowledge propagation.

symReplicateM_ :: forall (mode :: EvalModeTag) m a int. (EvalModeBase mode, TryMerge m, Applicative m, Mergeable a, Num int, UnifiedBranching mode m, UnifiedSymOrd mode Int, UnifiedSymOrd mode int) => Int -> int -> m a -> m ()

replicateM_ with MergingStrategy knowledge propagation and symbolic number of elements.

Conditional execution of monadic expressions

mrgGuard :: (Alternative m, TryMerge m) => Bool -> m ()

guard with MergingStrategy knowledge propagation.

symGuard :: (SymBranching m, TryMerge m, Alternative m) => SymBool -> m ()

guard with MergingStrategy knowledge propagation and symbolic conditions.

mrgWhen :: (Applicative m, TryMerge m) => Bool -> m () -> m ()

when with MergingStrategy knowledge propagation.

symWhen :: (Applicative m, TryMerge m, SymBranching m) => SymBool -> m () -> m ()

when with MergingStrategy knowledge propagation and symbolic conditions.

mrgUnless :: (Applicative m, TryMerge m) => Bool -> m () -> m ()

unless with MergingStrategy knowledge propagation.

symUnless :: (Applicative m, TryMerge m, SymBranching m) => SymBool -> m () -> m ()

unless with MergingStrategy knowledge propagation and symbolic conditions.

Monadic lifting operators

mrgLiftM :: (MonadTryMerge m, Mergeable a, Mergeable b) => (a -> b) -> m a -> m b

liftM with MergingStrategy knowledge propagation.

mrgLiftM2 :: (MonadTryMerge m, Mergeable a, Mergeable b, Mergeable c) => (a -> b -> c) -> m a -> m b -> m c

liftM2 with MergingStrategy knowledge propagation.

mrgLiftM3 :: (MonadTryMerge m, Mergeable a, Mergeable b, Mergeable c, Mergeable d) => (a -> b -> c -> d) -> m a -> m b -> m c -> m d

liftM3 with MergingStrategy knowledge propagation.

mrgLiftM4 :: (MonadTryMerge m, Mergeable a, Mergeable b, Mergeable c, Mergeable d, Mergeable e) => (a -> b -> c -> d -> e) -> m a -> m b -> m c -> m d -> m e

liftM4 with MergingStrategy knowledge propagation.

mrgLiftM5 :: (MonadTryMerge m, Mergeable a, Mergeable b, Mergeable c, Mergeable d, Mergeable e, Mergeable f) => (a -> b -> c -> d -> e -> f) -> m a -> m b -> m c -> m d -> m e -> m f

liftM5 with MergingStrategy knowledge propagation.

mrgAp :: (MonadTryMerge m, Mergeable a, Mergeable b) => m (a -> b) -> m a -> m b

<*> with MergingStrategy knowledge propagation.

Strict monadic functions

(.<$!>) :: (MonadTryMerge m, Mergeable a, Mergeable b) => (a -> b) -> m a -> m b infixl 4

<$!> with MergingStrategy knowledge propagation. Merging is always strict so we can directly use .<$>.