Grisette.Internal.SymPrim.Prim.Internal.Term

Description

Supported primitive types

class SupportedPrimConstraint t

Type class for resolving the constraint for a supported primitive type.

Associated Types

type PrimConstraint t

type PrimConstraint t = ()

class (Lift t, NFData t, SupportedPrimConstraint t, SBVRep t) => SupportedPrim t where

Indicates that a type is supported, can be represented as a symbolic term, and can be lowered to an SBV term.

Minimal complete definition

defaultValue, pevalITETerm, pevalEqTerm, pevalDistinctTerm, conSBVTerm, symSBVName, symSBVTerm, parseSMTModelResult, castTypedSymbol, funcDummyConstraint

Methods

primTypeRep :: TypeRep t

default primTypeRep :: Typeable t => TypeRep t

sameCon :: t -> t -> Bool

default sameCon :: Eq t => t -> t -> Bool

hashConWithSalt :: Int -> t -> Int

default hashConWithSalt :: Hashable t => Int -> t -> Int

pformatCon :: t -> String

default pformatCon :: Show t => t -> String

defaultValue :: t

pevalITETerm :: Term Bool -> Term t -> Term t -> Term t

pevalEqTerm :: Term t -> Term t -> Term Bool

pevalDistinctTerm :: NonEmpty (Term t) -> Term Bool

conSBVTerm :: t -> SBVType t

symSBVName :: TypedSymbol 'AnyKind t -> Int -> String

symSBVTerm :: SBVFreshMonad m => String -> m (SBVType t)

withPrim :: ((PrimConstraint t, SMTDefinable (SBVType t), Mergeable (SBVType t), Typeable (SBVType t)) => a) -> a

default withPrim :: (PrimConstraint t, SMTDefinable (SBVType t), Mergeable (SBVType t), Typeable (SBVType t)) => ((PrimConstraint t, SMTDefinable (SBVType t), Mergeable (SBVType t), Typeable (SBVType t)) => a) -> a

sbvIte :: SBV Bool -> SBVType t -> SBVType t -> SBVType t

sbvEq :: SBVType t -> SBVType t -> SBV Bool

default sbvEq :: EqSymbolic (SBVType t) => SBVType t -> SBVType t -> SBV Bool

sbvDistinct :: NonEmpty (SBVType t) -> SBV Bool

default sbvDistinct :: EqSymbolic (SBVType t) => NonEmpty (SBVType t) -> SBV Bool

parseSMTModelResult :: Int -> ([([CV], CV)], CV) -> t

castTypedSymbol :: forall (knd' :: SymbolKind) (knd :: SymbolKind). IsSymbolKind knd' => TypedSymbol knd t -> Maybe (TypedSymbol knd' t)

funcDummyConstraint :: SBVType t -> SBV Bool

withSupportedPrimTypeable :: SupportedPrim a => (Typeable a => b) -> b

Introduce the Typeable constraint from SupportedPrim.

class SupportedPrim con => SymRep con

Type family to resolve the symbolic type associated with a concrete type.

Associated Types

type SymType con

class ConRep sym

Type family to resolve the concrete type associated with a symbolic type.

Associated Types

type ConType sym

class (ConRep sym, SymRep con, sym ~ SymType con, con ~ ConType sym) => LinkedRep con sym | con -> sym, sym -> con where

One-to-one mapping between symbolic types and concrete types.

Methods

underlyingTerm :: sym -> Term con

wrapTerm :: Term con -> sym

Partial evaluation for the terms

class PEvalApplyTerm f a b | f -> a b where

Partial evaluation and lowering for function application terms.

Methods

pevalApplyTerm :: Term f -> Term a -> Term b

sbvApplyTerm :: SBVType f -> SBVType a -> SBVType b

class PEvalBitwiseTerm t where

Partial evaluation and lowering for bitwise operation terms.

Minimal complete definition

pevalAndBitsTerm, pevalOrBitsTerm, pevalXorBitsTerm, pevalComplementBitsTerm, withSbvBitwiseTermConstraint

Methods

pevalAndBitsTerm :: Term t -> Term t -> Term t

pevalOrBitsTerm :: Term t -> Term t -> Term t

pevalXorBitsTerm :: Term t -> Term t -> Term t

pevalComplementBitsTerm :: Term t -> Term t

withSbvBitwiseTermConstraint :: (Bits (SBVType t) => r) -> r

sbvAndBitsTerm :: SBVType t -> SBVType t -> SBVType t

sbvOrBitsTerm :: SBVType t -> SBVType t -> SBVType t

sbvXorBitsTerm :: SBVType t -> SBVType t -> SBVType t

sbvComplementBitsTerm :: SBVType t -> SBVType t

class PEvalShiftTerm t where

Partial evaluation and lowering for symbolic shifting terms.

Minimal complete definition

pevalShiftLeftTerm, pevalShiftRightTerm, withSbvShiftTermConstraint

Methods

pevalShiftLeftTerm :: Term t -> Term t -> Term t

pevalShiftRightTerm :: Term t -> Term t -> Term t

withSbvShiftTermConstraint :: (SIntegral (NonFuncSBVBaseType t) => r) -> r

sbvShiftLeftTerm :: SBVType t -> SBVType t -> SBVType t

default sbvShiftLeftTerm :: SupportedNonFuncPrim t => SBVType t -> SBVType t -> SBVType t

sbvShiftRightTerm :: SBVType t -> SBVType t -> SBVType t

default sbvShiftRightTerm :: SupportedNonFuncPrim t => SBVType t -> SBVType t -> SBVType t

class PEvalRotateTerm t where

Partial evaluation and lowering for symbolic rotate terms.

Minimal complete definition

pevalRotateLeftTerm, pevalRotateRightTerm, withSbvRotateTermConstraint

Methods

pevalRotateLeftTerm :: Term t -> Term t -> Term t

pevalRotateRightTerm :: Term t -> Term t -> Term t

withSbvRotateTermConstraint :: (SIntegral (NonFuncSBVBaseType t) => r) -> r

sbvRotateLeftTerm :: SBVType t -> SBVType t -> SBVType t

default sbvRotateLeftTerm :: SupportedNonFuncPrim t => SBVType t -> SBVType t -> SBVType t

sbvRotateRightTerm :: SBVType t -> SBVType t -> SBVType t

default sbvRotateRightTerm :: SupportedNonFuncPrim t => SBVType t -> SBVType t -> SBVType t

class Num t => PEvalNumTerm t where

Partial evaluation and lowering for number terms.

Minimal complete definition

pevalAddNumTerm, pevalNegNumTerm, pevalMulNumTerm, pevalAbsNumTerm, pevalSignumNumTerm, withSbvNumTermConstraint

Methods

pevalAddNumTerm :: Term t -> Term t -> Term t

pevalNegNumTerm :: Term t -> Term t

pevalMulNumTerm :: Term t -> Term t -> Term t

pevalAbsNumTerm :: Term t -> Term t

pevalSignumNumTerm :: Term t -> Term t

withSbvNumTermConstraint :: (Num (SBVType t) => r) -> r

sbvAddNumTerm :: SBVType t -> SBVType t -> SBVType t

sbvNegNumTerm :: SBVType t -> SBVType t

sbvMulNumTerm :: SBVType t -> SBVType t -> SBVType t

sbvAbsNumTerm :: SBVType t -> SBVType t

sbvSignumNumTerm :: SBVType t -> SBVType t

pevalSubNumTerm :: PEvalNumTerm a => Term a -> Term a -> Term a

Partial evaluation for subtraction terms.

class PEvalOrdTerm t where

Partial evaluation and lowering for comparison terms.

Minimal complete definition

pevalLtOrdTerm, pevalLeOrdTerm, withSbvOrdTermConstraint

Methods

pevalLtOrdTerm :: Term t -> Term t -> Term Bool

pevalLeOrdTerm :: Term t -> Term t -> Term Bool

withSbvOrdTermConstraint :: (OrdSymbolic (SBVType t) => r) -> r

sbvLtOrdTerm :: SBVType t -> SBVType t -> SBV Bool

sbvLeOrdTerm :: SBVType t -> SBVType t -> SBV Bool

pevalGtOrdTerm :: PEvalOrdTerm a => Term a -> Term a -> Term Bool

Partial evaluation for greater than terms.

pevalGeOrdTerm :: PEvalOrdTerm a => Term a -> Term a -> Term Bool

Partial evaluation for greater than or equal to terms.

pevalNEqTerm :: SupportedPrim a => Term a -> Term a -> Term Bool

Partial evaluation for inequality terms.

class PEvalDivModIntegralTerm t where

Partial evaluation and lowering for integer division and modulo terms.

Minimal complete definition

pevalDivIntegralTerm, pevalModIntegralTerm, pevalQuotIntegralTerm, pevalRemIntegralTerm, withSbvDivModIntegralTermConstraint

Methods

pevalDivIntegralTerm :: Term t -> Term t -> Term t

pevalModIntegralTerm :: Term t -> Term t -> Term t

pevalQuotIntegralTerm :: Term t -> Term t -> Term t

pevalRemIntegralTerm :: Term t -> Term t -> Term t

withSbvDivModIntegralTermConstraint :: (SDivisible (SBVType t) => r) -> r

sbvDivIntegralTerm :: SBVType t -> SBVType t -> SBVType t

sbvModIntegralTerm :: SBVType t -> SBVType t -> SBVType t

sbvQuotIntegralTerm :: SBVType t -> SBVType t -> SBVType t

sbvRemIntegralTerm :: SBVType t -> SBVType t -> SBVType t

class BitCast a b => PEvalBitCastTerm a b where

Partial evaluation and lowering for bitcast terms.

Methods

pevalBitCastTerm :: Term a -> Term b

sbvBitCast :: SBVType a -> SBVType b

class BitCastOr a b => PEvalBitCastOrTerm a b where

Partial evaluation and lowering for bitcast or default value terms.

Methods

pevalBitCastOrTerm :: Term b -> Term a -> Term b

sbvBitCastOr :: SBVType b -> SBVType a -> SBVType b

class SizedBV bv => PEvalBVTerm (bv :: Natural -> Type) where

Partial evaluation and lowering for bit-vector terms.

Methods

pevalBVConcatTerm :: forall (l :: Nat) (r :: Nat). (KnownNat l, KnownNat r, 1 <= l, 1 <= r) => Term (bv l) -> Term (bv r) -> Term (bv (l + r))

pevalBVExtendTerm :: forall (l :: Nat) (r :: Nat) proxy. (KnownNat l, KnownNat r, 1 <= l, 1 <= r, l <= r) => Bool -> proxy r -> Term (bv l) -> Term (bv r)

pevalBVSelectTerm :: forall (n :: Nat) (ix :: Nat) (w :: Nat) p q. (KnownNat n, KnownNat ix, KnownNat w, 1 <= n, 1 <= w, (ix + w) <= n) => p ix -> q w -> Term (bv n) -> Term (bv w)

sbvBVConcatTerm :: forall (l :: Nat) (r :: Nat) p1 p2. (KnownNat l, KnownNat r, 1 <= l, 1 <= r) => p1 l -> p2 r -> SBVType (bv l) -> SBVType (bv r) -> SBVType (bv (l + r))

sbvBVExtendTerm :: forall (l :: Nat) (r :: Nat) p1 p2. (KnownNat l, KnownNat r, 1 <= l, 1 <= r, l <= r) => p1 l -> p2 r -> Bool -> SBVType (bv l) -> SBVType (bv r)

sbvBVSelectTerm :: forall (ix :: Nat) (w :: Nat) (n :: Nat) p1 p2 p3. (KnownNat ix, KnownNat w, KnownNat n, 1 <= n, 1 <= w, (ix + w) <= n) => p1 ix -> p2 w -> p3 n -> SBVType (bv n) -> SBVType (bv w)

class Fractional t => PEvalFractionalTerm t where

Partial evaluation and lowering for fractional terms.

Minimal complete definition

pevalFdivTerm, pevalRecipTerm, withSbvFractionalTermConstraint

Methods

pevalFdivTerm :: Term t -> Term t -> Term t

pevalRecipTerm :: Term t -> Term t

withSbvFractionalTermConstraint :: (Fractional (SBVType t) => r) -> r

sbvFdivTerm :: SBVType t -> SBVType t -> SBVType t

sbvRecipTerm :: SBVType t -> SBVType t

class PEvalFloatingTerm t where

Partial evaluation and lowering for floating point terms.

Minimal complete definition

pevalFloatingUnaryTerm, pevalPowerTerm, withSbvFloatingTermConstraint

Methods

pevalFloatingUnaryTerm :: FloatingUnaryOp -> Term t -> Term t

pevalPowerTerm :: Term t -> Term t -> Term t

withSbvFloatingTermConstraint :: (Floating (SBVType t) => r) -> r

sbvPowerTerm :: SBVType t -> SBVType t -> SBVType t

sbvFloatingUnaryTerm :: FloatingUnaryOp -> SBVType t -> SBVType t

class (Integral a, Num b) => PEvalFromIntegralTerm a b where

Partial evaluation and lowering for integral terms.

Methods

pevalFromIntegralTerm :: Term a -> Term b

sbvFromIntegralTerm :: SBVType a -> SBVType b

class PEvalIEEEFPConvertibleTerm a where

Partial evaluation and lowering for converting from and to IEEE floating point terms.

Methods

pevalFromFPOrTerm :: forall (eb :: Nat) (sb :: Nat). ValidFP eb sb => Term a -> Term FPRoundingMode -> Term (FP eb sb) -> Term a

pevalToFPTerm :: forall (eb :: Nat) (sb :: Nat). ValidFP eb sb => Term FPRoundingMode -> Term a -> Term (FP eb sb)

sbvFromFPOrTerm :: forall (eb :: Nat) (sb :: Nat). ValidFP eb sb => SBVType a -> SBVType FPRoundingMode -> SBVType (FP eb sb) -> SBVType a

sbvToFPTerm :: forall (eb :: Nat) (sb :: Nat). ValidFP eb sb => SBVType FPRoundingMode -> SBVType a -> SBVType (FP eb sb)

Typed symbols

data SymbolKind

The kind of a symbol.

All symbols are AnyKind, and all symbols other than general/tabular functions are ConstantKind.

Constructors

ConstantKind

AnyKind

data TypedSymbol (knd :: SymbolKind) t where

A typed symbol is a symbol that is associated with a type. Note that the same symbol bodies with different types are considered different symbols and can coexist in a term.

Simple symbols can be created with the OverloadedStrings extension:

>>> "a" :: TypedSymbol 'AnyKind Bool
a :: Bool

Constructors

TypedSymbol

Fields

:: forall t (knd :: SymbolKind). (SupportedPrim t, SymbolKindConstraint knd t, IsSymbolKind knd)

=> { unTypedSymbol :: Symbol

} -> TypedSymbol knd t

typedConstantSymbol :: SupportedNonFuncPrim t => Symbol -> TypedSymbol 'ConstantKind t

Create a typed symbol with constant kinds.

typedAnySymbol :: SupportedPrim t => Symbol -> TypedSymbol 'AnyKind t

Create a typed symbol with any kinds.

type TypedConstantSymbol = TypedSymbol 'ConstantKind

Constant symbol

type TypedAnySymbol = TypedSymbol 'AnyKind

Any symbol

data SomeTypedSymbol (knd :: SymbolKind) where

A non-indexed symbol. Type information are checked at runtime.

Constructors

SomeTypedSymbol :: forall (knd :: SymbolKind) t. TypedSymbol knd t -> SomeTypedSymbol knd

type SomeTypedConstantSymbol = SomeTypedSymbol 'ConstantKind

Non-indexed constant symbol

type SomeTypedAnySymbol = SomeTypedSymbol 'AnyKind

Non-indexed any symbol

class IsSymbolKind (ty :: SymbolKind) where

Decision procedure for symbol kinds.

Associated Types

type SymbolKindConstraint (ty :: SymbolKind) :: Type -> Constraint

Methods

decideSymbolKind :: Either (ty :~~: 'ConstantKind) (ty :~~: 'AnyKind)

showUntyped :: forall (knd :: SymbolKind) t. TypedSymbol knd t -> String

Show a typed symbol without the type information.

withSymbolSupported :: forall (knd :: SymbolKind) t a. TypedSymbol knd t -> ((SupportedPrim t, Typeable t) => a) -> a

Introduce the SupportedPrim constraint from the TypedSymbol.

withConstantSymbolSupported :: TypedSymbol 'ConstantKind t -> ((SupportedNonFuncPrim t, Typeable t) => a) -> a

Introduce the SupportedPrim constraint from the TypedSymbol.

someTypedSymbol :: forall (knd :: SymbolKind) t. TypedSymbol knd t -> SomeTypedSymbol knd

Construct a SomeTypedSymbol from a TypedSymbol.

eqHeteroSymbol :: forall (ta :: SymbolKind) a (tb :: SymbolKind) b. TypedSymbol ta a -> TypedSymbol tb b -> Bool

Compare two TypedSymbols for equality.

castSomeTypedSymbol :: forall (knd' :: SymbolKind) (knd :: SymbolKind). IsSymbolKind knd' => SomeTypedSymbol knd -> Maybe (SomeTypedSymbol knd')

Cast a typed symbol to a different kind. Check if the kind is compatible.

withSymbolKind :: forall (knd :: SymbolKind) t a. TypedSymbol knd t -> (IsSymbolKind knd => a) -> a

Introduce the IsSymbolKind constraint from the TypedSymbol.

Terms

data FPTrait

Traits for IEEE floating point numbers.

Constructors

FPIsNaN

FPIsPositive

FPIsNegative

FPIsPositiveInfinite

FPIsNegativeInfinite

FPIsInfinite

FPIsPositiveZero

FPIsNegativeZero

FPIsZero

FPIsNormal

FPIsSubnormal

FPIsPoint

data FPUnaryOp

Unary floating point operations.

Constructors

FPAbs

FPNeg

data FPBinaryOp

Binary floating point operations.

Constructors

FPRem

FPMinimum

FPMinimumNumber

FPMaximum

FPMaximumNumber

data FPRoundingUnaryOp

Unary floating point operations with rounding modes.

Constructors

FPSqrt

FPRoundToIntegral

data FPRoundingBinaryOp

Binary floating point operations with rounding modes.

Constructors

FPAdd

FPSub

FPMul

FPDiv

data FloatingUnaryOp

Unary floating point operations.

Constructors

FloatingExp

FloatingLog

FloatingSqrt

FloatingSin

FloatingCos

FloatingTan

FloatingAsin

FloatingAcos

FloatingAtan

FloatingSinh

FloatingCosh

FloatingTanh

FloatingAsinh

FloatingAcosh

FloatingAtanh

data Term t where

Internal representation for Grisette symbolic terms.

Constructors

ConTerm :: forall t. SupportedPrim t => WeakThreadId -> !Digest -> Id -> Ident -> !t -> Term t

SymTerm :: forall t. WeakThreadId -> !Digest -> Id -> Ident -> !(TypedSymbol 'AnyKind t) -> Term t

ForallTerm :: forall t1. WeakThreadId -> !Digest -> Id -> Ident -> !(TypedSymbol 'ConstantKind t1) -> !(Term Bool) -> Term Bool

ExistsTerm :: forall t1. WeakThreadId -> !Digest -> Id -> Ident -> !(TypedSymbol 'ConstantKind t1) -> !(Term Bool) -> Term Bool

NotTerm :: WeakThreadId -> !Digest -> Id -> Ident -> !(Term Bool) -> Term Bool

OrTerm :: WeakThreadId -> !Digest -> Id -> Ident -> !(Term Bool) -> !(Term Bool) -> Term Bool

AndTerm :: WeakThreadId -> !Digest -> Id -> Ident -> !(Term Bool) -> !(Term Bool) -> Term Bool

EqTerm :: forall t1. WeakThreadId -> !Digest -> Id -> Ident -> !(Term t1) -> !(Term t1) -> Term Bool

DistinctTerm :: forall t1. WeakThreadId -> !Digest -> Id -> Ident -> !(NonEmpty (Term t1)) -> Term Bool

ITETerm :: forall t. SupportedPrim t => WeakThreadId -> !Digest -> Id -> Ident -> !(Term Bool) -> !(Term t) -> !(Term t) -> Term t

AddNumTerm :: forall t. (SupportedPrim t, PEvalNumTerm t) => WeakThreadId -> !Digest -> Id -> Ident -> !(Term t) -> !(Term t) -> Term t

NegNumTerm :: forall t. (SupportedPrim t, PEvalNumTerm t) => WeakThreadId -> !Digest -> Id -> Ident -> !(Term t) -> Term t

MulNumTerm :: forall t. (SupportedPrim t, PEvalNumTerm t) => WeakThreadId -> !Digest -> Id -> Ident -> !(Term t) -> !(Term t) -> Term t

AbsNumTerm :: forall t. (SupportedPrim t, PEvalNumTerm t) => WeakThreadId -> !Digest -> Id -> Ident -> !(Term t) -> Term t

SignumNumTerm :: forall t. (SupportedPrim t, PEvalNumTerm t) => WeakThreadId -> !Digest -> Id -> Ident -> !(Term t) -> Term t

LtOrdTerm :: forall t1. (SupportedPrim t1, PEvalOrdTerm t1) => WeakThreadId -> !Digest -> Id -> Ident -> !(Term t1) -> !(Term t1) -> Term Bool

LeOrdTerm :: forall t1. (SupportedPrim t1, PEvalOrdTerm t1) => WeakThreadId -> !Digest -> Id -> Ident -> !(Term t1) -> !(Term t1) -> Term Bool

AndBitsTerm :: forall t. (SupportedPrim t, PEvalBitwiseTerm t) => WeakThreadId -> !Digest -> Id -> Ident -> !(Term t) -> !(Term t) -> Term t

OrBitsTerm :: forall t. (SupportedPrim t, PEvalBitwiseTerm t) => WeakThreadId -> !Digest -> Id -> Ident -> !(Term t) -> !(Term t) -> Term t

XorBitsTerm :: forall t. (SupportedPrim t, PEvalBitwiseTerm t) => WeakThreadId -> !Digest -> Id -> Ident -> !(Term t) -> !(Term t) -> Term t

ComplementBitsTerm :: forall t. (SupportedPrim t, PEvalBitwiseTerm t) => WeakThreadId -> !Digest -> Id -> Ident -> !(Term t) -> Term t

ShiftLeftTerm :: forall t. (SupportedPrim t, PEvalShiftTerm t) => WeakThreadId -> !Digest -> Id -> Ident -> !(Term t) -> !(Term t) -> Term t

ShiftRightTerm :: forall t. (SupportedPrim t, PEvalShiftTerm t) => WeakThreadId -> !Digest -> Id -> Ident -> !(Term t) -> !(Term t) -> Term t

RotateLeftTerm :: forall t. (SupportedPrim t, PEvalRotateTerm t) => WeakThreadId -> !Digest -> Id -> Ident -> !(Term t) -> !(Term t) -> Term t

RotateRightTerm :: forall t. (SupportedPrim t, PEvalRotateTerm t) => WeakThreadId -> !Digest -> Id -> Ident -> !(Term t) -> !(Term t) -> Term t

BitCastTerm :: forall t a. (SupportedPrim t, PEvalBitCastTerm a t) => WeakThreadId -> !Digest -> Id -> Ident -> !(Term a) -> Term t

BitCastOrTerm :: forall t a. (SupportedPrim t, PEvalBitCastOrTerm a t) => WeakThreadId -> !Digest -> Id -> Ident -> !(Term t) -> !(Term a) -> Term t

BVConcatTerm :: forall (bv :: Natural -> Type) (l :: Nat) (r :: Nat). (PEvalBVTerm bv, KnownNat l, KnownNat r, KnownNat (l + r), 1 <= l, 1 <= r, 1 <= (l + r), SupportedPrim (bv (l + r))) => WeakThreadId -> !Digest -> Id -> Ident -> !(Term (bv l)) -> !(Term (bv r)) -> Term (bv (l + r))

BVSelectTerm :: forall (bv :: Natural -> Type) (n :: Nat) (ix :: Nat) (w :: Nat). (PEvalBVTerm bv, KnownNat n, KnownNat ix, KnownNat w, 1 <= n, 1 <= w, (ix + w) <= n, SupportedPrim (bv w)) => WeakThreadId -> !Digest -> Id -> Ident -> !(Proxy ix) -> !(Proxy w) -> !(Term (bv n)) -> Term (bv w)

BVExtendTerm :: forall (bv :: Natural -> Type) (l :: Nat) (r :: Nat). (PEvalBVTerm bv, KnownNat l, KnownNat r, 1 <= l, 1 <= r, l <= r, SupportedPrim (bv r)) => WeakThreadId -> !Digest -> Id -> Ident -> !Bool -> !(Proxy r) -> !(Term (bv l)) -> Term (bv r)

ApplyTerm :: forall f a t. (PEvalApplyTerm f a t, SupportedPrim t) => WeakThreadId -> !Digest -> Id -> Ident -> !(Term f) -> !(Term a) -> Term t

DivIntegralTerm :: forall t. (SupportedPrim t, PEvalDivModIntegralTerm t) => WeakThreadId -> !Digest -> Id -> Ident -> !(Term t) -> !(Term t) -> Term t

ModIntegralTerm :: forall t. (SupportedPrim t, PEvalDivModIntegralTerm t) => WeakThreadId -> !Digest -> Id -> Ident -> !(Term t) -> !(Term t) -> Term t

QuotIntegralTerm :: forall t. (SupportedPrim t, PEvalDivModIntegralTerm t) => WeakThreadId -> !Digest -> Id -> Ident -> !(Term t) -> !(Term t) -> Term t

RemIntegralTerm :: forall t. (SupportedPrim t, PEvalDivModIntegralTerm t) => WeakThreadId -> !Digest -> Id -> Ident -> !(Term t) -> !(Term t) -> Term t

FPTraitTerm :: forall (eb :: Nat) (sb :: Nat). (ValidFP eb sb, SupportedPrim (FP eb sb)) => WeakThreadId -> !Digest -> Id -> Ident -> !FPTrait -> !(Term (FP eb sb)) -> Term Bool

FdivTerm :: forall t. (SupportedPrim t, PEvalFractionalTerm t) => WeakThreadId -> !Digest -> Id -> Ident -> !(Term t) -> !(Term t) -> Term t

RecipTerm :: forall t. (SupportedPrim t, PEvalFractionalTerm t) => WeakThreadId -> !Digest -> Id -> Ident -> !(Term t) -> Term t

FloatingUnaryTerm :: forall t. (SupportedPrim t, PEvalFloatingTerm t) => WeakThreadId -> !Digest -> Id -> Ident -> !FloatingUnaryOp -> !(Term t) -> Term t

PowerTerm :: forall t. (SupportedPrim t, PEvalFloatingTerm t) => WeakThreadId -> !Digest -> Id -> Ident -> !(Term t) -> !(Term t) -> Term t

FPUnaryTerm :: forall (eb :: Nat) (sb :: Nat). (ValidFP eb sb, SupportedPrim (FP eb sb)) => WeakThreadId -> !Digest -> Id -> Ident -> !FPUnaryOp -> !(Term (FP eb sb)) -> Term (FP eb sb)

FPBinaryTerm :: forall (eb :: Nat) (sb :: Nat). (ValidFP eb sb, SupportedPrim (FP eb sb)) => WeakThreadId -> !Digest -> Id -> Ident -> !FPBinaryOp -> !(Term (FP eb sb)) -> !(Term (FP eb sb)) -> Term (FP eb sb)

FPRoundingUnaryTerm :: forall (eb :: Nat) (sb :: Nat). (ValidFP eb sb, SupportedPrim (FP eb sb)) => WeakThreadId -> !Digest -> Id -> Ident -> !FPRoundingUnaryOp -> !(Term FPRoundingMode) -> !(Term (FP eb sb)) -> Term (FP eb sb)

FPRoundingBinaryTerm :: forall (eb :: Nat) (sb :: Nat). (ValidFP eb sb, SupportedPrim (FP eb sb)) => WeakThreadId -> !Digest -> Id -> Ident -> !FPRoundingBinaryOp -> !(Term FPRoundingMode) -> !(Term (FP eb sb)) -> !(Term (FP eb sb)) -> Term (FP eb sb)

FPFMATerm :: forall (eb :: Nat) (sb :: Nat). (ValidFP eb sb, SupportedPrim (FP eb sb)) => WeakThreadId -> !Digest -> Id -> Ident -> !(Term FPRoundingMode) -> !(Term (FP eb sb)) -> !(Term (FP eb sb)) -> !(Term (FP eb sb)) -> Term (FP eb sb)

FromIntegralTerm :: forall a t. (PEvalFromIntegralTerm a t, SupportedPrim t) => WeakThreadId -> !Digest -> Id -> Ident -> !(Term a) -> Term t

FromFPOrTerm :: forall t (eb :: Nat) (sb :: Nat). (PEvalIEEEFPConvertibleTerm t, ValidFP eb sb, SupportedPrim t) => WeakThreadId -> !Digest -> Id -> Ident -> !(Term t) -> !(Term FPRoundingMode) -> !(Term (FP eb sb)) -> Term t

ToFPTerm :: forall a (eb :: Nat) (sb :: Nat). (PEvalIEEEFPConvertibleTerm a, ValidFP eb sb, SupportedPrim (FP eb sb)) => WeakThreadId -> !Digest -> Id -> Ident -> !(Term FPRoundingMode) -> !(Term a) -> Proxy eb -> Proxy sb -> Term (FP eb sb)

defaultValueDynamic :: SupportedPrim t => proxy t -> ModelValue

The default value in a dynamic ModelValue.

pattern DynTerm :: SupportedPrim a => Term a -> Term b

Pattern for term with dynamic typing.

toCurThread :: Term t -> IO (Term t)

Convert a term to the current thread.

termId :: Term t -> Id

Return the ID of a term.

termIdent :: Term t -> Ident

The identity of the term.

typeHashId :: Term t -> TypeHashId

Return the ID and the type representation of a term.

introSupportedPrimConstraint :: Term t -> ((SupportedPrim t, Typeable t) => a) -> a

Introduce the SupportedPrim constraint from a term.

pformatTerm :: Term t -> String

Pretty-print a term.

data ModelValue where

A value with its type information.

Constructors

ModelValue :: forall v. SupportedPrim v => v -> ModelValue

toModelValue :: SupportedPrim a => a -> ModelValue

Convert to a model value.

unsafeFromModelValue :: Typeable a => ModelValue -> a

Convert from a model value. Crashes if the types does not match.

Interning

data UTerm t where

Term without identity (before internalizing).

Constructors

UConTerm :: forall t. SupportedPrim t => !t -> UTerm t

USymTerm :: forall t. !(TypedSymbol 'AnyKind t) -> UTerm t

UForallTerm :: forall t1. !(TypedSymbol 'ConstantKind t1) -> !(Term Bool) -> UTerm Bool

UExistsTerm :: forall t1. !(TypedSymbol 'ConstantKind t1) -> !(Term Bool) -> UTerm Bool

UNotTerm :: !(Term Bool) -> UTerm Bool

UOrTerm :: !(Term Bool) -> !(Term Bool) -> UTerm Bool

UAndTerm :: !(Term Bool) -> !(Term Bool) -> UTerm Bool

UEqTerm :: forall t1. !(Term t1) -> !(Term t1) -> UTerm Bool

UDistinctTerm :: forall t1. !(NonEmpty (Term t1)) -> UTerm Bool

UITETerm :: forall t. SupportedPrim t => !(Term Bool) -> !(Term t) -> !(Term t) -> UTerm t

UAddNumTerm :: forall t. (SupportedPrim t, PEvalNumTerm t) => !(Term t) -> !(Term t) -> UTerm t

UNegNumTerm :: forall t. (SupportedPrim t, PEvalNumTerm t) => !(Term t) -> UTerm t

UMulNumTerm :: forall t. (SupportedPrim t, PEvalNumTerm t) => !(Term t) -> !(Term t) -> UTerm t

UAbsNumTerm :: forall t. (SupportedPrim t, PEvalNumTerm t) => !(Term t) -> UTerm t

USignumNumTerm :: forall t. (SupportedPrim t, PEvalNumTerm t) => !(Term t) -> UTerm t

ULtOrdTerm :: forall t1. (SupportedPrim t1, PEvalOrdTerm t1) => !(Term t1) -> !(Term t1) -> UTerm Bool

ULeOrdTerm :: forall t1. (SupportedPrim t1, PEvalOrdTerm t1) => !(Term t1) -> !(Term t1) -> UTerm Bool

UAndBitsTerm :: forall t. (SupportedPrim t, PEvalBitwiseTerm t) => !(Term t) -> !(Term t) -> UTerm t

UOrBitsTerm :: forall t. (SupportedPrim t, PEvalBitwiseTerm t) => !(Term t) -> !(Term t) -> UTerm t

UXorBitsTerm :: forall t. (SupportedPrim t, PEvalBitwiseTerm t) => !(Term t) -> !(Term t) -> UTerm t

UComplementBitsTerm :: forall t. (SupportedPrim t, PEvalBitwiseTerm t) => !(Term t) -> UTerm t

UShiftLeftTerm :: forall t. (SupportedPrim t, PEvalShiftTerm t) => !(Term t) -> !(Term t) -> UTerm t

UShiftRightTerm :: forall t. (SupportedPrim t, PEvalShiftTerm t) => !(Term t) -> !(Term t) -> UTerm t

URotateLeftTerm :: forall t. (SupportedPrim t, PEvalRotateTerm t) => !(Term t) -> !(Term t) -> UTerm t

URotateRightTerm :: forall t. (SupportedPrim t, PEvalRotateTerm t) => !(Term t) -> !(Term t) -> UTerm t

UBitCastTerm :: forall t a. (SupportedPrim t, PEvalBitCastTerm a t) => !(Term a) -> UTerm t

UBitCastOrTerm :: forall t a. (SupportedPrim t, PEvalBitCastOrTerm a t) => !(Term t) -> !(Term a) -> UTerm t

UBVConcatTerm :: forall (bv :: Natural -> Type) (l :: Nat) (r :: Nat). (PEvalBVTerm bv, KnownNat l, KnownNat r, KnownNat (l + r), 1 <= l, 1 <= r, 1 <= (l + r), SupportedPrim (bv (l + r))) => !(Term (bv l)) -> !(Term (bv r)) -> UTerm (bv (l + r))

UBVSelectTerm :: forall (bv :: Natural -> Type) (n :: Nat) (ix :: Nat) (w :: Nat). (PEvalBVTerm bv, KnownNat n, KnownNat ix, KnownNat w, 1 <= n, 1 <= w, (ix + w) <= n, SupportedPrim (bv w)) => !(Proxy ix) -> !(Proxy w) -> !(Term (bv n)) -> UTerm (bv w)

UBVExtendTerm :: forall (bv :: Natural -> Type) (l :: Nat) (r :: Nat). (PEvalBVTerm bv, KnownNat l, KnownNat r, 1 <= l, 1 <= r, l <= r, SupportedPrim (bv r)) => !Bool -> !(Proxy r) -> !(Term (bv l)) -> UTerm (bv r)

UApplyTerm :: forall f a t. (PEvalApplyTerm f a t, SupportedPrim t) => Term f -> Term a -> UTerm t

UDivIntegralTerm :: forall t. (SupportedPrim t, PEvalDivModIntegralTerm t) => !(Term t) -> !(Term t) -> UTerm t

UModIntegralTerm :: forall t. (SupportedPrim t, PEvalDivModIntegralTerm t) => !(Term t) -> !(Term t) -> UTerm t

UQuotIntegralTerm :: forall t. (SupportedPrim t, PEvalDivModIntegralTerm t) => !(Term t) -> !(Term t) -> UTerm t

URemIntegralTerm :: forall t. (SupportedPrim t, PEvalDivModIntegralTerm t) => !(Term t) -> !(Term t) -> UTerm t

UFPTraitTerm :: forall (eb :: Nat) (sb :: Nat). (ValidFP eb sb, SupportedPrim (FP eb sb)) => !FPTrait -> !(Term (FP eb sb)) -> UTerm Bool

UFdivTerm :: forall t. (SupportedPrim t, PEvalFractionalTerm t) => !(Term t) -> !(Term t) -> UTerm t

URecipTerm :: forall t. (SupportedPrim t, PEvalFractionalTerm t) => !(Term t) -> UTerm t

UFloatingUnaryTerm :: forall t. (SupportedPrim t, PEvalFloatingTerm t) => !FloatingUnaryOp -> !(Term t) -> UTerm t

UPowerTerm :: forall t. (SupportedPrim t, PEvalFloatingTerm t) => !(Term t) -> !(Term t) -> UTerm t

UFPUnaryTerm :: forall (eb :: Nat) (sb :: Nat). (ValidFP eb sb, SupportedPrim (FP eb sb)) => !FPUnaryOp -> !(Term (FP eb sb)) -> UTerm (FP eb sb)

UFPBinaryTerm :: forall (eb :: Nat) (sb :: Nat). (ValidFP eb sb, SupportedPrim (FP eb sb)) => !FPBinaryOp -> !(Term (FP eb sb)) -> !(Term (FP eb sb)) -> UTerm (FP eb sb)

UFPRoundingUnaryTerm :: forall (eb :: Nat) (sb :: Nat). (ValidFP eb sb, SupportedPrim (FP eb sb)) => !FPRoundingUnaryOp -> !(Term FPRoundingMode) -> !(Term (FP eb sb)) -> UTerm (FP eb sb)

UFPRoundingBinaryTerm :: forall (eb :: Nat) (sb :: Nat). (ValidFP eb sb, SupportedPrim (FP eb sb)) => !FPRoundingBinaryOp -> !(Term FPRoundingMode) -> !(Term (FP eb sb)) -> !(Term (FP eb sb)) -> UTerm (FP eb sb)

UFPFMATerm :: forall (eb :: Nat) (sb :: Nat). (ValidFP eb sb, SupportedPrim (FP eb sb)) => !(Term FPRoundingMode) -> !(Term (FP eb sb)) -> !(Term (FP eb sb)) -> !(Term (FP eb sb)) -> UTerm (FP eb sb)

UFromIntegralTerm :: forall a t. (PEvalFromIntegralTerm a t, SupportedPrim t) => !(Term a) -> UTerm t

UFromFPOrTerm :: forall t (eb :: Nat) (sb :: Nat). (PEvalIEEEFPConvertibleTerm t, SupportedPrim t, ValidFP eb sb) => Term t -> !(Term FPRoundingMode) -> !(Term (FP eb sb)) -> UTerm t

UToFPTerm :: forall a (eb :: Nat) (sb :: Nat). (PEvalIEEEFPConvertibleTerm a, ValidFP eb sb, SupportedPrim (FP eb sb)) => !(Term FPRoundingMode) -> !(Term a) -> Proxy eb -> Proxy sb -> UTerm (FP eb sb)

prettyPrintTerm :: Term t -> Doc ann

Pretty-print a term, possibly eliding parts of it.

forallTerm :: TypedSymbol 'ConstantKind t -> Term Bool -> Term Bool

Construct and internalizing a ForallTerm.

existsTerm :: TypedSymbol 'ConstantKind t -> Term Bool -> Term Bool

Construct and internalizing a ExistsTerm.

conTerm :: SupportedPrim t => t -> Term t

Construct and internalizing a ConTerm.

symTerm :: forall (knd :: SymbolKind) t. TypedSymbol knd t -> Term t

Construct and internalizing a SymTerm.

ssymTerm :: SupportedPrim t => Identifier -> Term t

Construct and internalizing a SymTerm with an identifier, using simple symbols.

isymTerm :: SupportedPrim t => Identifier -> Int -> Term t

Construct and internalizing a SymTerm with an identifier and an index, using indexed symbols.

notTerm :: Term Bool -> Term Bool

Construct and internalizing a NotTerm.

orTerm :: Term Bool -> Term Bool -> Term Bool

Construct and internalizing a OrTerm.

andTerm :: Term Bool -> Term Bool -> Term Bool

Construct and internalizing a AndTerm.

eqTerm :: Term a -> Term a -> Term Bool

Construct and internalizing a EqTerm.

distinctTerm :: NonEmpty (Term a) -> Term Bool

Construct and internalizing a DistinctTerm.

iteTerm :: Term Bool -> Term a -> Term a -> Term a

Construct and internalizing a ITETerm.

addNumTerm :: PEvalNumTerm a => Term a -> Term a -> Term a

Construct and internalizing a AddNumTerm.

negNumTerm :: PEvalNumTerm a => Term a -> Term a

Construct and internalizing a NegNumTerm.

mulNumTerm :: PEvalNumTerm a => Term a -> Term a -> Term a

Construct and internalizing a MulNumTerm.

absNumTerm :: PEvalNumTerm a => Term a -> Term a

Construct and internalizing a AbsNumTerm.

signumNumTerm :: PEvalNumTerm a => Term a -> Term a

Construct and internalizing a SignumNumTerm.

ltOrdTerm :: PEvalOrdTerm a => Term a -> Term a -> Term Bool

Construct and internalizing a LtOrdTerm.

leOrdTerm :: PEvalOrdTerm a => Term a -> Term a -> Term Bool

Construct and internalizing a LeOrdTerm.

andBitsTerm :: PEvalBitwiseTerm a => Term a -> Term a -> Term a

Construct and internalizing a AndBitsTerm.

orBitsTerm :: PEvalBitwiseTerm a => Term a -> Term a -> Term a

Construct and internalizing a OrBitsTerm.

xorBitsTerm :: PEvalBitwiseTerm a => Term a -> Term a -> Term a

Construct and internalizing a XorBitsTerm.

complementBitsTerm :: PEvalBitwiseTerm a => Term a -> Term a

Construct and internalizing a ComplementBitsTerm.

shiftLeftTerm :: PEvalShiftTerm a => Term a -> Term a -> Term a

Construct and internalizing a ShiftLeftTerm.

shiftRightTerm :: PEvalShiftTerm a => Term a -> Term a -> Term a

Construct and internalizing a ShiftRightTerm.

rotateLeftTerm :: PEvalRotateTerm a => Term a -> Term a -> Term a

Construct and internalizing a RotateLeftTerm.

rotateRightTerm :: PEvalRotateTerm a => Term a -> Term a -> Term a

Construct and internalizing a RotateRightTerm.

bitCastTerm :: (PEvalBitCastTerm a b, SupportedPrim b) => Term a -> Term b

Construct and internalizing a BitCastTerm.

bitCastOrTerm :: PEvalBitCastOrTerm a b => Term b -> Term a -> Term b

Construct and internalizing a BitCastOrTerm.

bvConcatTerm :: forall bv (l :: Nat) (r :: Nat). (PEvalBVTerm bv, KnownNat l, KnownNat r, KnownNat (l + r), 1 <= l, 1 <= r, 1 <= (l + r), SupportedPrim (bv (l + r))) => Term (bv l) -> Term (bv r) -> Term (bv (l + r))

Construct and internalizing a BVConcatTerm.

bvSelectTerm :: forall bv (n :: Nat) (ix :: Nat) (w :: Nat) p q. (PEvalBVTerm bv, KnownNat n, KnownNat ix, KnownNat w, 1 <= n, 1 <= w, (ix + w) <= n, SupportedPrim (bv w)) => p ix -> q w -> Term (bv n) -> Term (bv w)

Construct and internalizing a BVSelectTerm.

bvExtendTerm :: forall bv (l :: Nat) (r :: Nat) proxy. (PEvalBVTerm bv, KnownNat l, KnownNat r, 1 <= l, 1 <= r, l <= r, SupportedPrim (bv r)) => Bool -> proxy r -> Term (bv l) -> Term (bv r)

Construct and internalizing a BVExtendTerm.

bvsignExtendTerm :: forall bv (l :: Nat) (r :: Nat) proxy. (PEvalBVTerm bv, KnownNat l, KnownNat r, 1 <= l, 1 <= r, l <= r, SupportedPrim (bv r)) => proxy r -> Term (bv l) -> Term (bv r)

Construct and internalizing a BVExtendTerm with sign extension.

bvzeroExtendTerm :: forall bv (l :: Nat) (r :: Nat) proxy. (PEvalBVTerm bv, KnownNat l, KnownNat r, 1 <= l, 1 <= r, l <= r, SupportedPrim (bv r)) => proxy r -> Term (bv l) -> Term (bv r)

Construct and internalizing a BVExtendTerm with zero extension.

applyTerm :: (PEvalApplyTerm f a b, SupportedPrim b) => Term f -> Term a -> Term b

Construct and internalizing a ApplyTerm.

divIntegralTerm :: PEvalDivModIntegralTerm a => Term a -> Term a -> Term a

Construct and internalizing a DivIntegralTerm.

modIntegralTerm :: PEvalDivModIntegralTerm a => Term a -> Term a -> Term a

Construct and internalizing a ModIntegralTerm.

quotIntegralTerm :: PEvalDivModIntegralTerm a => Term a -> Term a -> Term a

Construct and internalizing a QuotIntegralTerm.

remIntegralTerm :: PEvalDivModIntegralTerm a => Term a -> Term a -> Term a

Construct and internalizing a RemIntegralTerm.

fpTraitTerm :: forall (eb :: Nat) (sb :: Nat). (ValidFP eb sb, SupportedPrim (FP eb sb)) => FPTrait -> Term (FP eb sb) -> Term Bool

Construct and internalizing a FPTraitTerm.

fdivTerm :: PEvalFractionalTerm a => Term a -> Term a -> Term a

Construct and internalizing a FdivTerm.

recipTerm :: PEvalFractionalTerm a => Term a -> Term a

Construct and internalizing a RecipTerm.

floatingUnaryTerm :: PEvalFloatingTerm a => FloatingUnaryOp -> Term a -> Term a

Construct and internalizing a FloatingUnaryTerm.

powerTerm :: PEvalFloatingTerm a => Term a -> Term a -> Term a

Construct and internalizing a PowerTerm.

fpUnaryTerm :: forall (eb :: Nat) (sb :: Nat). (ValidFP eb sb, SupportedPrim (FP eb sb)) => FPUnaryOp -> Term (FP eb sb) -> Term (FP eb sb)

Construct and internalizing a FPUnaryTerm.

fpBinaryTerm :: forall (eb :: Nat) (sb :: Nat). (ValidFP eb sb, SupportedPrim (FP eb sb)) => FPBinaryOp -> Term (FP eb sb) -> Term (FP eb sb) -> Term (FP eb sb)

Construct and internalizing a FPBinaryTerm.

fpRoundingUnaryTerm :: forall (eb :: Nat) (sb :: Nat). (ValidFP eb sb, SupportedPrim (FP eb sb)) => FPRoundingUnaryOp -> Term FPRoundingMode -> Term (FP eb sb) -> Term (FP eb sb)

Construct and internalizing a FPRoundingUnaryTerm.

fpRoundingBinaryTerm :: forall (eb :: Nat) (sb :: Nat). (ValidFP eb sb, SupportedPrim (FP eb sb)) => FPRoundingBinaryOp -> Term FPRoundingMode -> Term (FP eb sb) -> Term (FP eb sb) -> Term (FP eb sb)

Construct and internalizing a FPRoundingBinaryTerm.

fpFMATerm :: forall (eb :: Nat) (sb :: Nat). (ValidFP eb sb, SupportedPrim (FP eb sb)) => Term FPRoundingMode -> Term (FP eb sb) -> Term (FP eb sb) -> Term (FP eb sb) -> Term (FP eb sb)

Construct and internalizing a FPFMATerm.

fromIntegralTerm :: (PEvalFromIntegralTerm a b, SupportedPrim b) => Term a -> Term b

Construct and internalizing a FromIntegralTerm.

fromFPOrTerm :: forall a (eb :: Nat) (sb :: Nat). (PEvalIEEEFPConvertibleTerm a, ValidFP eb sb) => Term a -> Term FPRoundingMode -> Term (FP eb sb) -> Term a

Construct and internalizing a FromFPOrTerm.

toFPTerm :: forall a (eb :: Nat) (sb :: Nat). (PEvalIEEEFPConvertibleTerm a, ValidFP eb sb, SupportedPrim (FP eb sb)) => Term FPRoundingMode -> Term a -> Term (FP eb sb)

Construct and internalizing a ToFPTerm.

Support for boolean type

trueTerm :: Term Bool

Construct and internalizing True term.

falseTerm :: Term Bool

Construct and internalizing False term.

pattern BoolConTerm :: Bool -> Term a

Pattern matcher for concrete Bool terms.

pattern TrueTerm :: Term a

Pattern matcher for True term.

pattern FalseTerm :: Term a

Pattern matcher for False term.

pattern BoolTerm :: Term Bool -> Term a

Pattern matcher for Bool terms.

pevalNotTerm :: Term Bool -> Term Bool

Partial evaluation for not terms.

pevalOrTerm :: Term Bool -> Term Bool -> Term Bool

Partial evaluation for or terms.

pevalAndTerm :: Term Bool -> Term Bool -> Term Bool

Partial evaluation for and terms.

pevalImplyTerm :: Term Bool -> Term Bool -> Term Bool

Partial evaluation for imply terms.

pevalXorTerm :: Term Bool -> Term Bool -> Term Bool

Partial evaluation for xor terms.

pevalITEBasic :: SupportedPrim a => Term Bool -> Term a -> Term a -> Maybe (Term a)

Basic partial evaluation for ITE terms.

pevalITEBasicTerm :: SupportedPrim a => Term Bool -> Term a -> Term a -> Term a

Basic partial evaluation for ITE terms.

pevalDefaultEqTerm :: SupportedNonFuncPrim a => Term a -> Term a -> Term Bool

Default partial evaluation for equality terms.

type NonFuncPrimConstraint a = (SymVal (NonFuncSBVBaseType a), EqSymbolic (SBVType a), Mergeable (SBVType a), SMTDefinable (SBVType a), Mergeable (SBVType a), SBVType a ~ SBV (NonFuncSBVBaseType a), PrimConstraint a)

Type class for resolving the constraint for a supported non-function primitive type.

class (SupportedPrim a, Ord a) => NonFuncSBVRep a

Type class for resolving the base type for the SBV type for the primitive type.

Associated Types

type NonFuncSBVBaseType a

class NonFuncSBVRep a => SupportedNonFuncPrim a where

Indicates that a type is supported, can be represented as a symbolic term, is not a function type, and can be lowered to an SBV term.

Methods

conNonFuncSBVTerm :: a -> SBV (NonFuncSBVBaseType a)

symNonFuncSBVTerm :: SBVFreshMonad m => String -> m (SBV (NonFuncSBVBaseType a))

withNonFuncPrim :: (NonFuncPrimConstraint a => r) -> r

class SBVRep t

Type class for resolving the SBV type for the primitive type.

Associated Types

type SBVType t

class MonadIO m => SBVFreshMonad (m :: Type -> Type) where

Monads that supports generating sbv fresh variables.

Methods

sbvFresh :: SymVal a => String -> m (SBV a)

translateTypeError :: HasCallStack => Maybe String -> TypeRep a -> b

Error message for unsupported types.

parseSMTModelResultError :: HasCallStack => TypeRep a -> ([([CV], CV)], CV) -> a

Error message for failure to parse the SBV model result.

partitionCVArg :: SupportedNonFuncPrim a => [([CV], CV)] -> [(a, [([CV], CV)])]

Partition the list of CVs for models for functions.

parseScalarSMTModelResult :: forall v r. (SatModel r, Typeable v) => (r -> v) -> ([([CV], CV)], CV) -> v

Parse the scalar model result.