Grisette.Internal.Unified.UnifiedData

Description

Documentation

type family GetData (mode :: EvalModeTag) v = (r :: Type) | r -> mode v

Get a unified data type. Resolves to v in C mode, and Union v in S mode.

wrapData :: UnifiedDataImpl mode v u => v -> u

Wraps a value into the unified data type.

extractData :: (UnifiedDataImpl mode v u, Mergeable v, Monad m, UnifiedBranching mode m) => u -> m v

Extracts a value from the unified data type.

class UnifiedDataImpl mode v (GetData mode v) => UnifiedData (mode :: EvalModeTag) v

This class is needed as constraint in user code prior to GHC 9.2.1. See the notes in IsMode.

class (forall v. UnifiedData bool v, forall v. Mergeable v => UnifiedDataSimpleMergeable v) => AllUnifiedData (bool :: EvalModeTag)

Evaluation mode with unified data types.